### 2024-09-16 #7237 @editorialbot
[PRE REVIEW]: Envismetrics: A Python-based software for electrochemical kinetic analysis
**Submitting author:** @Woffee (<a href="http://orcid.org/0000-0001-8514-0608">Omowunmi Sadik</a>)
**Repository:** https://github.com/Woffee/Envismetrics
**Branch with paper.md** (empty if default branch): main
**Version:** V0.9

Thanks for submitting your paper to JOSS @https://github.com/Woffee. **Currently, there isn't a JOSS editor assigned** to your paper.
