### 2024-09-30 #7237 @ppxasjsm
@Woffee, I am your editor for this submission. If you have any questions please reach out. In the mean time, could you please take a look at the suggested improvements from the EiC? If you could also suggest some reviewers for your submission that would be great. You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. If you give me 3-5 suggestions for reviewers with their github handles without the '@', I can reach out to them for the review.
