### 2024-10-07 #7237 @Grant-Xue
Thank you for your feedback and for assisting us with the initial steps of the submission process. I am one of the authors of the project, and we have addressed the issues mentioned in your message. Specifically, we:

Fixed the reported potential DOI issues by adding ISSN to the books and reformatting the DOI entries in the .bib file.
Created and added a CONTRIBUTING.md file, and linked it in the README section of our GitHub repository.

Additionally, we will provide the reviewers' list shortly.
