### 2024-10-08 #7237 @Grant-Xue
I hope this message finds you well. I wanted to inform you that the corresponding author for our submission is Dr. Omowunmi Sadik. Her contact email is omowunmi.sadik@njit.edu.

Please note that Dr. Sadik does not currently have a GitHub account. I kindly request that any important communications or updates related to the review process be sent to her via email, in addition to the GitHub review platform. She plans to create an account in the future.

Thank you for your assistance, and please let me know if any additional information is required.
