### 2024-10-09 #7237 @ppxasjsm
Thank you @Grant-Xue. It is not necessary for the corresponding author to be involved in the submission per se. The review should still be visible for an account. Will you be the main contact for the submission to address any issues with the code etc?

Any reviewer suggestions would be massively appreciated to speed up the process of finding people to review this submission.
