### 2024-10-09 #7237 @Grant-Xue
Dear JOSS Editor Toni Mey

Thank you for your message. I will be the main contact for the submission
and will address any issues related to the code or the review process.

Regarding potential reviewers, I believe the following individuals have
extensive knowledge in electrochemical analysis and would provide valuable
feedback. However, they may not be registered as JOSS reviewers on GitHub:

   1.

   *Professor Omotayo A. Arotiba*
   Director, Center for Nanomaterials and Science Research
   Department of Chemistry, University of Johannesburg
   Email: ***@***.***
   2.

   *Professor Ashok Mulchandani*
   W. Ruel Johnson Chair in Environmental Engineering
   Distinguished Professor, Chemical & Environmental Engineering
   University of California, Riverside
   Email: ***@***.***
   3.

   *Professor Emmanuel Iwuoha*
   Senior Professor, SARChI (Tier 1)
   University of the Western Cape
   Email: ***@***.***
   4.

   *Professor Sudipta Seal*
   Department of Material Science and Engineering
   University of Central Florida
   Email: ***@***.***
   5.

   *Professor Silvana Andreescu*
   Egon Matijevic Chair of Chemistry / Co-Director of CAMP
   Department of Chemistry & Molecular Science
   Clarkson University
   Email: ***@***.***

Please let me know if any additional information is needed or if I can
assist further with the review process.

Best regards,
