### 2024-10-09 #7237 @ppxasjsm
It may be easier if you could suggest someone from the existing reviewer list, as people who have not used github before, may not be familiar with our review process. Can you check this list please: [the list of people](https://reviewers.joss.theoj.org/lookup)
