### 2024-10-14 #7237 @Grant-Xue
Dear Editor @ppxasjsm

Thank you for your message. I have reviewed the suggested improvements from the EiC and will work on addressing them promptly.

In the meantime, after searching for related works. I would like to propose the following reviewers for my submission:

mdmurbach
EricaEgg
alberto-battistel
yaomz16

These reviewers have expertise in areas relevant to my submission, including electrochemistry, data analysis, and Python-based software.

Please feel free to let me know if there's anything else I can assist with.
