### 2024-10-22 #7237 @ppxasjsm
Hi @EricaEgg, is there any chance you are available to review this submission for JOSS?
