### 2025-03-31 #7237 @Grant-Xue
Dear @ppxasjsm and @Kevin-Mattheus-Moerman,

Thank you again for your continued efforts in coordinating the review for our JOSS submission. As the process has been ongoing for over six months, we would like to assist further by proposing a few more potential reviewers.

After carefully reviewing the most similar JOSS papers suggested by @editorialbot, we have identified the following individuals whose expertise and past review history seem particularly well-aligned with our software:

- TomTranter – Known for a practical and user-focused review style, suitable for applied electrochemical data tools like ours.
- yangbai90 – Strong engineering and electrochemistry background, emphasizes reproducibility and usability in reviews.
- MaximevdHeijden – Involved in battery and data-analysis projects, with a constructive and reasonable review approach.

We believe these reviewers would provide meaningful, constructive insights while helping move the review process forward.

Please let us know if any of them could be approached, or if we can offer additional suggestions. We truly appreciate your support and look forward to the next steps.
