### 2025-03-31 #7237 @Grant-Xue
@ppxasjsm Just a quick note: outside of our current submission, I’d be happy to help as a reviewer for other chemistry or electrochemistry-related submissions in JOSS. I recently completed my PhD in Materials Science and Engineering (MTSE) with a focus on electrochemical analysis and data-driven modeling, and I’d love to contribute if needed.

Please feel free to reach out if there’s an opportunity. Thanks again!
