### 2025-06-20 #7237 @Kevin-Mattheus-Moerman
@ericfell, @rena-96, @lancekavalsky, @lscalfi, @mbabar09, @nmerovingian, @oliverrdz, @MoonshinetheP  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled _Envismetrics: A Python-based software for electrochemical kinetic analysis_ .

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/Woffee/Envismetrics) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07237/joss.07237/10.21105.joss.07237.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of a recent review if you want to see what the process is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
