### 2025-06-20 #7237 @rena-96
Sure!

Renata

Am Fr., 20. Juni 2025 um 13:17 Uhr schrieb Kevin Mattheus Moerman <
***@***.***>:

> *Kevin-Mattheus-Moerman* left a comment (openjournals/joss-reviews#7237)
> <https://github.com/openjournals/joss-reviews/issues/7237#issuecomment-2991061062>
>
> @ericfell <https://github.com/ericfell>, @rena-96
> <https://github.com/rena-96>, @lancekavalsky
> <https://github.com/lancekavalsky>, @lscalfi <https://github.com/lscalfi>,
> @mbabar09 <https://github.com/mbabar09>, @nmerovingian
> <https://github.com/nmerovingian>, @oliverrdz
> <https://github.com/oliverrdz>, @MoonshinetheP
> <https://github.com/MoonshinetheP> would you be interested in reviewing
> this submission for JOSS (the Journal of Open Source Software
> <https://joss.theoj.org>? The work is titled *Envismetrics: A
> Python-based software for electrochemical kinetic analysis* .
>
> JOSS reviews take place on GitHub and focus on the software 💾
> <https://github.com/Woffee/Envismetrics> as well as a short paper 📜
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.07237/joss.07237/10.21105.joss.07237.pdf>.
> We typically aim for a 2-3 week review but we can be more flexible if
> needed. You can find more about our review process in our review
> guidelines
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. Here is
> an example of a recent review if you want to see what the process is like:
> #7564 <https://github.com/openjournals/joss-reviews/issues/7564>.
>
> If you are interested you can let me know by responding here.
>
> Thanks!
>
