### 2025-06-20 #8436 @editorialbot
[REVIEW]: Envismetrics: A Python-based software for electrochemical kinetic analysis
**Submitting author:** @https://github.com/Woffee (<a href="http://orcid.org/0000-0001-8514-0608">Omowunmi Sadik</a>)
**Repository:** https://github.com/Woffee/Envismetrics
**Branch with paper.md** (empty if default branch): main
**Version:** V0.9

@MoonshinetheP & @oliverrdz & @mbabar09 & @rena-96, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
