### 2025-06-25 #8436 @Grant-Xue
Thank you @rena-96 for pointing that out!

We’ve now relaunched the demo server — you can access it here: http://34.162.1.1:8080/

I’ve done some preliminary testing and everything seems to be working well. If you encounter any issues or have any questions while using the tool, please don’t hesitate to let me know. I’m happy to assist throughout the review process!
