### 2025-06-26 Woffee/Envismetrics#1 @MoonshinetheP
MIT License - Personal details possibly need updating
Hello! I'm posting this issue as part of [JOSS review #8436](github.com/openjournals/joss-reviews/issues/8436).

The MIT license features a person not on the Author's list, as well as a year before the repository was created. I imagine the license was copied from somewhere else, but please make sure to edit the personal information so that the copyright applies to you.

If you want to extend copyright ownership to more than one person or to specify a range of years in which the software was developed, it is also possible ([this article explains it quite well](https://gustavopinto.medium.com/explaining-the-mit-license-ddc8bc44a14a)).

I'll post more issues as I am going throught the code/paper, but I thought this one was important to do so that everyone can tick off the 'License' box on their checklist.

Best Regards,
MoonshinetheP
