### 2025-06-26 Woffee/Envismetrics#2 @MoonshinetheP
Possible irregularities in authorship
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

I'm tagging @Kevin-Mattheus-Moerman for advice on this (I hope you don't mind; I assume you are still our editor).

One of the reviewer checklist items is:

> Contribution and authorship: Has the submitting author (@Woffee) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I'm having a little trouble judging the second question in this item for the following reasons:

1. The manuscript has five authors, but there are only two contributors in the repository. From what was written in the [pre-review](https://github.com/openjournals/joss-reviews/issues/7237) thread, it would seem likely that the last (corresponding) author is a group supervisor without access to GitHub. I think there is a relevant section in the JOSS guidelines:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are.

However, there is another line in the guidelines:

> The authors themselves assume responsibility for deciding who should be credited with co-authorship

It's possible the last author contributed to the writing of the software without a GitHub account, but even with just a pure supervisory role, I think it is up to the authors to decide how the co-authorship works. The current format is in line with the rules for typical research journals, if not strictly following the JOSS guidelines. However, should this division of work not be declared in some way in the manuscript? Likewise, the contribution of the remaining two authors is not currently clear from the manuscript nor the repository. Again, it's possible these authors were responsible for testing or some other aspect in development, but should this not be stated?

2. There are asterisks indicating equal contribution for the first two authors. However, from the commit history, it would appear the second author has contributed significantly more (in both volume and content) to the project. Therefore, I think the asterisks should be removed. This also raises the question of first-authorship; for mainstream scientific journals, the first author is usually the person that contributed the mosts (i.e. did the majority of the research or wrote the manuscript). Here I don't know what the format should be, because the software development is clearly the more involved activity. I also understand there could be other factors that have led to the order of the two being the way it is, but again would it not make sense to specify the roles of each person (or at least give the reviewers an explanation for this arrangement)?

Perhaps such instances of confusion could be avoided by using some brief statement such as is seen in the [CRediT system](https://www.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement) which is becoming popular in mainstream scientific journals?

Apologies if none of the above is an issue, just looks a bit odd to me coming from more of a typical scientific reviewing background.

Best Regards,
MoonshinetheP
