### 2025-06-27 #8436 @rena-96
Ok, now the download from the website works.

 I also have some additional (minor) comments regarding your paper, which I believe would enhance the clarity of the project for the users.

This paper illustrates a software for the online and local analysis of electrochemical data. Additionally, educational content related to the analysis is available from the online tool. The paper is quite clear, and the tool shows the mentioned functionalities.

However, I believe one should address those minor comments:
-	Summary: The word “vital” is a bit dramatic for a scientific statement. I would suggest using ‘fundamental’
-	Summary: references are missing when listing the analysis kinds
-	Statement of need: I think it would help to place the statement in a clearer context and avoid language that sounds overly self-promotional. Instead of saying it’s the best, maybe focus on explaining why it’s useful and how it differs from other software. It might also strengthen the argument to elaborate a bit more on motivation, what makes it important? In general, adding more precise and scientific reasoning behind the statements would make it more informative and compelling. For example, the sentence ‘Whether users are professional researchers seeking to save time or individuals lacking basic knowledge of the relevant equations, Envismetrics encourages reproducible, easy-to-use, and transparent analysis’. It is a listing of desirable characteristics, but it is not very precise, in my opinion.
-	In general, sometimes you use the hyphen, sometimes you do not. I suggest fixing those typos.
-	Figure 2: should revise the axis, those are labeled in different ways (1/2 as superscript). Sometimes the fitting lines also cover the datasets as they are too thick (red lines) in comparison to the data points.
-	Sometimes you define the variable names. Sometimes you do not. I suggest you clearly state that you are following the notation of the given reference. What is n? What is n’? Are you taking the same assumptions as the given reference?
-	You should check how you are citing your references. I think they are just typos, but you have a lot of repetitions, for instance “The method is suggested 116 by Nicholson (1965) and Lavagnini et al. (2004) Lavagnini et al. (2004).”
Installation: if someone uses Anaconda with Windows or a similar setting, the source command is not working. I suggest explaining how to do it for the users for local installation.

I hope those comments would help.
