### 2025-06-28 Woffee/Envismetrics#3 @MoonshinetheP
Improving installation guide
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

One of the reviewer checklist items is:

> Installation: Does installation proceed as outlined in the documentation?

Please make the following changes so that I can sign off this item:

1.  I would suggest swapping the order so that users are instructed to first clone the repository (what is referred to as Installation) and then create the virtual environment (what is referred to as Prerequisites). While it is not necessary to keep the virtual environment in the local repository and some people like to keep their virtual environments in one centralised folder, I would imagine it is more beginner friendly to do everything withing the local repository and let more advanced users decide where to keep their virtual environment. You could add a line to say you recommend keeping the virtual environment in such and such a way.
2. Consider expanding on instructions for specific OS's. You have the instructions on how to set up a virtual environment but seemingly only for Linux/Mac OS. You stated a few times that the software is cross-platform, so best to make sure all instructions are for each OS.
3. You may also want to include some brief description on how to make sure the user has an up-to-date Git/Python installation. It is not necessary in most repositories, but you have to consider that the people who would get the most out of your software are electrochemists who have little experience in programming, and accordingly guide them through the installation process. I know one of the other reviewers mentioned something about Anaconda and its a good point - a lot of electrochemists are more likely to go through this easier route of package installation and will not know
4. On a similar topic, it is maybe worth clarifying: do you intend for all users to clone and run locally, or will this be an option aside from you also running a server with the software? If the latter, it needs to be clarified better.
5. gunicorn does not work with Windows, so you need to suggest an alternative in the instructions for Windows users

Best Regards,
MoonshinetheP
