### 2025-06-29 Woffee/Envismetrics#1 @Woffee
Dear MoonshinetheP,

Thank you very much for pointing this out as part of the JOSS review.

You’re right. The original license was copied and had not yet been updated. I have now replaced it with a revised MIT license reflecting the correct name (Huize Xue) and year (2024). The updated license is committed to the repository.

Please let me know if anything further needs adjustment. I appreciate your careful review.

Best regards,
Wenbo Wang
