### 2025-06-29 Woffee/Envismetrics#1 @MoonshinetheP
Hi Wenbo,

A quick follow up and then I think I can sign off on the license:

1. There is a link to the license in the top left of the readme.md (in the md this is: [![MIT License][license-shield]][license-url]). However, the link redirects to a different license (unlicense), again in the repository of othneildrew (I understand now how the MIT license came to have this name/year). The badge is a nice touch, but perhaps not so useful if it points to the wrong place (I don't know how to fix it, perhaps [this](https://shields.io/badges) can help).

2. I am not sure on the exact rules of who is the copyright holder in a repository with multiple contributors. I always understood that it should be the repository holder (i.e. you), but that other contributors can also be included. I'm not sure of the situation in your laboratory (whether the first author is PI but not corresponding, whether you are a temporary student etc.), some situations may warrant changes in authorship/ownership (I briefly touched on this in another issue), but I am not sure if the same situations warrant a change/omission in copyright holders. For example, I am always making the repository myself, so that there is no issue with ownership if a student leaves, but always including them alongside myself in the license (because after all, it is their work in the repository). This could be fixed using the multiple people option in the license, so please discuss with your authors on the approach which works best for you and who to include.
