### 2025-06-29 Woffee/Envismetrics#4 @MoonshinetheP
Cleaning repository
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

A minor issue here, but perhaps consider cleaning up the repository a little. There is currently one level containing:

- Source code
- Documentation, tests and configuration files
- JOSS manuscript (PDF and MD), bibliography and images for manuscript

There are also example files and test files, which could be easily combined (especially as it seems the test files are not explicitly referred to in the tests - more on this later). This makes the repository a little hard to navigate (and will only get worse as you add more features).

I think a better practise would be to have on the first level:

- Individual files - readme.md, pytest.ini, LICENSE.txt, CONTRIBUTING.md, .gitignore, config.py
- Folder for source code containing - BaseModule.py, CA.py, CV.py, HDV.py, app.py, static/, templates/
- Folder containing tests
- Folder containing test data
- Folder containing JOSS manuscript items - paper.md (and pdf), bibliography.bib, Image_Set/

I might have missed something, and bear in mind suggestions may break dependencies, but it would certainly be a lot cleaner to look at. Anyway, just a suggestion.

Best Regards,
MoonshinetheP
