### 2025-06-29 Woffee/Envismetrics#5 @MoonshinetheP
Manuscript comments -
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

One of the reviewer checklist items is:

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

As far as I understand, the only thing this could refer to in your case is the subplots in Figure 2. Please somewhere indicate which files were used to generate these subplots and which settings were entered into the application to attain the appearance of these subplots in the manuscript (something like an Experimental Details in a typical scientific research article). This not only would help us reviewers determine the reproducibility of the code, but could also serve as a tutorial for users (you already upload some example files, but guide them on how to use these files to get the results you did).

Best Regards,
MoonshinetheP
