### 2025-06-29 Woffee/Envismetrics#5 @MoonshinetheP
This would also help us tick off another item:

> Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

This is difficult to judge currently. I assume the data in the manuscript is in Test_Set/, but some guide to the data would be nice (I'm not sure I am up to manually comparing each dataset in this folder with the results in manuscript).
