### 2025-06-30 #8436 @Grant-Xue
We sincerely thank you for your careful reading of our manuscript and for the constructive comments that helped us significantly improve the clarity and quality of the paper. We have addressed each of your suggestions in detail below.

1. Use of the word “vital” in the Summary
Comment: The word “vital” is a bit dramatic for a scientific statement.
Response: We have replaced “vital” with “fundamental” to reflect a more precise and neutral tone in the scientific context.
2. Missing references when listing the types of analysis
Comment: Summary: references are missing when listing the analysis kinds.
Response: Added proper references for all listed analytical methods, including:
- Levich and Koutecký–Levich analysis [@doi:10.1021/ar50110a004; @doi:10.1002/1521-4109...]
- Randles–Ševčík and transfer coefficient methods [@doi:10.1002/adts.202500346; @LEFTHERIOTIS2007259]
- Cottrell equation-based chronoamperometry [@HERATH20084324; @RODRIGUEZLUCAS2025145648]
3. Statement of Need sounds overly promotional
Comment: Reframe to focus on usefulness and scientific context rather than sounding like a product pitch.
Response: We have revised the Statement of Need section to provide a more objective comparison between Envismetrics and other tools (e.g., proprietary and custom-built solutions). We emphasize its cross-platform functionality, generalizability, and transparency in a scientific tone.
4. Inconsistent use of hyphenation
Comment: Sometimes you use the hyphen, sometimes you do not.
Response: We have standardized the use of hyphenated compound modifiers (e.g., “cross-platform,” “instrument-specific”) throughout the manuscript.
5. Axis label formatting and line thickness in Figure 2
Comment: Axis labels should be consistent (e.g., using superscript for 1/2), and fitting lines should not obscure data points.
Response: We revised all axis labels for consistent scientific formatting (e.g., [𝜈(V/s)]1/2) and adjusted plot line thickness and transparency to ensure data visibility.
6. Inconsistent variable definition
Comment: Sometimes you define the variable names. Sometimes you do not.
Response: We added Table 2, which systematically defines all key electrochemical symbols, their meanings, units, and context of usage. A note is also included to clarify dual usage of symbols like v.
7. Redundant citations
Comment: You have repetitions like “Lavagnini et al. (2004) Lavagnini et al. (2004).”
Response: We reviewed and corrected all in-text references to avoid duplication. Now, each citation appears only once and in correct syntax (e.g., [@lavagnini2004extended]).
8. Clarify installation instructions for Anaconda users
Comment: Source command may not work on Anaconda for Windows.
Response: We clarified in the manuscript that the online version requires no installation. For users wishing to run locally, detailed Anaconda-compatible setup instructions are available in the GitHub repository's readme part, which is now explicitly referenced.

Again, thank you for your valuable feedback. We believe these changes have significantly improved the quality and clarity of our submission. We appreciate your time and consideration.
