### 2025-07-01 Woffee/Envismetrics#3 @Woffee
Dear MoonshinetheP,

Thank you so much for your thoughtful and constructive feedback on the Envismetrics README. I really appreciate the time you took to consider the perspective of users who may not have a strong programming background.

I have carefully reviewed all of your suggestions and made the updates accordingly.

Please feel free to revisit the README to review the updated content and let me know if anything further stands out.

Regards,
Wenbo Wang
