### 2025-07-01 Woffee/Envismetrics#1 @Woffee
Dear MoonshinetheP,
Thank you again for your valuable feedback. Here is our response:

1. We’ve temporarily removed the license badge until we can fix it to correctly point to the appropriate license file.

2. Huize Xue is the first author of both this project and our submitted paper. He wrote most of the core logic for the project, and the project would not exist without his foundational work. Since he is not experienced in web development, he provided the non-GUI codebase, and I built the web interface on top of it. I have also been primarily responsible for the web development and server maintenance.

Given his central contribution to the core functionality, we believe he is the most important contributor to this project. After our discussion, I’ve updated the license file to list both him and me as copyright holders.

Please let me know if there’s anything else that needs to be revised.

Regards,
Wenbo
