### 2025-07-01 Woffee/Envismetrics#1 @MoonshinetheP
Hi Wenbo,

Thanks for fixing the issue, I've ticked the item off on our checklist. Just to clarify, my point here was not to question the exact nature of any author's contribution, but to say that you (at least) should also hold copyright. I don't have any problem with Huize Xui holding copyright, and I think the current fix is appropriate.

Your comments are maybe more relevant to another issue (#2) in which I asked whether Huize Xue's and your own contribution were equal, when the commit history suggests otherwise. If as you say, he built the Python code (possibly elsewhere or offline) and you built the web app using this code, then I think equal contribution is appropriate, as is first authorship for Huize Xue. Ultimately however, it is up to you and the other authors to decide; I only wanted to point out the irregularity.

If you like, we can continue the discussion there. I will close this Issue as you have made the appropriate fix.
