### 2025-07-01 Woffee/Envismetrics#3 @MoonshinetheP
Hi Wenbo,

The updated installation guide is much better, so I will close this Issue. I will say, I was not able to get app.py to run on port 8080 (for some reason, this was being used by something else), but another port was free and after changing I was able to run app.py and use the web app. Maybe it would be good to briefly (one sentence or so) mention how to change port if it doesn't work and how to check if the port is busy:
`netstat -ano | findstr :8080`
