### 2025-07-01 Woffee/Envismetrics#6 @MoonshinetheP
Manuscript - General comments
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

Overall, the manuscript is well-written, concise and nicely structured. I think after some corrections (mentioned in subsequent Issues) it will be appropriate for publication in JOSS.

I do however have one generic criticsm - I think you are focussing too much on the '_kinetic analysis_' component of your software and not at all on the more justifiable '_automated analysis_' aspect to your work. I also think that using the term '_electrochemical kinetic analysis_' in the title (and throughout the manuscript) is a misstep. See why:

- You are correct in using the term '_kinetic parameters_' to describe diffusion coefficent, but most would conflate this term with '_electron transfer kinetics_' and (maybe incorrectly) associate it more specifically to describe the standard rate constant and transfer coefficient. It would be worth clarifying, especially for novices, that the rate of the electrochemical reaction is dependant on both electron transfer and mass transport (which is often the rate-determining step) and that the kinetic parameters you want to derive can be split into those in the electron transfer domain and those in the mass transport domain.

- Following on from this, it is worth noting that a large portion of your software's capabilities focus on the determination of one specific kinetic parameter, the diffusion coefficient, using various equations (Levich, Randles-Sevcik, Cottrell). However, it is often the case that the diffusion coefficient is a known quantity (or can at least be approximated - there is not a lot of range in its value for a solute in a liquid solvent) and that these equations are instead used to determine other non-kinetic parameters (_e.g._ concentration, electrode radius, number of electrons). Actually, I would even say that of all the equations used, only Cottrell is suitable for determination of diffusion coefficient, Levich and Randles-Sevcik have better uses (I will make another Issue for this).

- Consequently, I think that specifically labelling the software as designed for 'kinetic analysis' can: (1) limit the scope of the software, both now and with future updates (_e.g._ calculating formal potential, identifying iR drop); (2) make it more complicated for you to clarify and describe which features (future and present) don't fit into this label; and (3) neglect (what I think is) the true advantage of your software, the ability to automate the routine analysis of electrochemical data.

- Likewise, keeping the application of the software more broad could improve clarity and help attract more novice users.

- Finally, the terms '_kinetic analysis_' and '_electrochemical kinetic analysis_' do not seem correct. You are analysing electrochemical data to get kinetic information, not directly analysing the kinetics of the electrochemical reaction. If anything, this term should be changed to something like '_analysis of kinetic parameters_' or '_determination of kinetic parameters through analysis of electrochemical data'_.

I would therefore suggest shifting the focus towards general electrochemical analysis (but automated) and perhaps changing the title of the paper to '_Envisimetrics: A Python-based software for automated analysis of electrochemical data_'. I don't think it requires too much effort, just some subtle changes throughout the manuscript. This could increase the appeal of your software and help keep the scope broad when considering future updates.

Best Regards,
MoonshinetheP
