### 2025-07-01 Woffee/Envismetrics#7 @MoonshinetheP
Manuscript - Abstract
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

### General comments

- In general, I don't know if JOSS allows an abstract - it doesn't feature in the [guidelines ](https://joss.readthedocs.io/en/latest/paper.html) and I cannot find one in any of my downloaded JOSS articles. It seems the summary should serve as the abstract (but more from a perspective of what problem your software solves and what exactly does it do). Still, I will give corrections on this abstract in case elements end up incorporated into the Summary.

### Specific corrections

- **Line 10 -** The adjective '_innovative_' is a bit self-aggrandising for what is essentially software enabling people to do routine electrochemical analysis. I would instead focus on the ease-of-use of your software or its accessibility, which are more objective appraisals

- **Line 11 -** Change '_to streamline the electrochemical kinetic analysis process_' to '_to streamline the analysis of electrochemical data_'.

- **Line 11 -** I don't know if the term '_software suite_' applies here. It's just one software.

- **Line 11-13 -** Personally I would not describe the data processing or analysis steps as '_efficient_', because if I was to manually perform the same processes it would likely be in the same fashion (there are only so many ways to plot Randles-Sevcik for example, so not a lot of room for inefficiency). A better descriptor could lie in the automated approach (i.e. you take care of everything for the user in this context) applied to numerous datasets. A better line might be '_This software offers a comprehensive toolbox that enables researchers to automate aspects of the processing and analysis of electrochemical data_'. I removed the part about different potentiostats, because as I will get to, it is a little redundant.

- **Line 13-14 -** I don't see much wrong in the sentence starting '_Key features_', except that all these features (with exception of importing and processing of data) are only for CV. Either make it general (plotting, filtering, analysis, output) or give some example for each actual use of the software (HDV and CA).

- **Line 16 -** I think '_simplifying_' is redundant when followed by '_automated_'. How does your software simplify beyond automating the processing/analysis steps?

- **Line 17 -** '_Ease of use_' of what? It is not clear in the sentence

Best Regards,
MoonshinetheP
