### 2025-07-01 Woffee/Envismetrics#8 @MoonshinetheP
Manuscript - Summary
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

### General comments

- I will make this point here, although it applies throughout the manuscript and also to the software itself. There is a slight problem with terminology when referring to the techniques your software is compatible with. Hydrodynamic voltammetry is a pretty broad term covering voltammetry in stirred solutions, voltammetry at RDEs and RRDEs, the use of microfluidics, wall jet electrodes _etc._ You can do Levich and Koutecky-Levich analysis in certain conditions on some of these systems, but your software is only accounting for RDEs. Personally, I think you should replace this terminology with something like 'Linear sweep voltammetry at a rotating disk electrode (RDE): Levich and Koutecky-Levich analysis in laminar flow conditions.' to more accurately reflect what the technique/analysis is. Cyclic voltammetry can stay as it is if you want it to, but it is worth noting that you could also do the same analysis on a linear sweep voltammogram (which is after all only one half of the cyclic voltammogram). The difference then is not the technique, but the electrode used (i.e. Linear/cyclic voltammetry on a stationary electrode vs. linear voltammetry on an RDE). Also maybe worth specifying that this is for planar diffusion only (so most likely macroelectrodes).

- I think towards the end of the Summary you should introduce Envisimetrics as a platform to perform the indicated analyses and it general benefits using the text that is currently in the Abstract.

### Specific corrections

- Line 20 - Change '_Accurate determination of kinetic electrochemical parameters and thermodynamic constants_' to '_Accurate determination of kinetic parameters and thermodynamic properties from electrochemical data_'? I changed '_thermodynamic constants_' to '_thermodynamic properties_' to avoid confusion, as the former more often refers to static values (F, R, kB _etc._).

- Line 21 - Change 'predicting and optimizing the performance of redox reactions in various applications' to 'understanding redox reactions' used in various applications? It doesn't make 100% sense that you are predicting the performance of something that you already had to do in order to get the parameters needed for prediction. Likewise, even if you know these parameters, it is not 100% logical that you can optimise the reaction, I would say this mostly comes from understanding the mechanism (i.e. avoiding unwanted chemical steps) than knowing exact kinetic parameters.

- Line 22 - Change 'These kinetic parameters' to 'These quantities' or 'These values'? Formal potential is not a kinetic parameter and is necessary (kind of) to perform simulations.

- Line 22 - 23 - I would rather say that you use such parameters in simulations to validate the mechanism. It's not a huge distinction, but you can often determine mechanism from data without simulations and in the case that you can't, this is mostly limited by the users ability to interpret the data (see for example Gundry et al. 2022). If you know what the mechanism is and what the parameters you need are, simulation is more likely going to help validate your electrochemical data and maybe predict behaviour of the reaction in conditions that are more difficult to enact experimentally.

- Line 23 - 25 - The whole sentence starting 'However, many ...' could go. It doesn't add much to the context of the problem, isn't 100% true (unless studying something completely novel, there are often plenty of tables and literature examples containing formal potential and diffusion coefficient for commonly used redox compounds), and raises more questions than it answers (which experimental conditions affect such parameters?). If you want something like this, I would rather say something like '_Whilst such parameters can be found for selected redox compounds in the literature, evaluation of new redox compounds/reactions needs to be performed experimentally.'_.

- Line 26 - 28 - See the general comment about use of CV and HDV. I don't see why you isolated CA ('and step methods such as...') in comparison to the first two techniques (i.e. why not 'voltammetric techniques such as ... and step methods such as ...')

- Line 34 - 'essential' is a bad choice of word. If you mean that the kinetic parameters are basic in nature, I would use either 'basic' or 'fundamental'. If you mean the kinetic parameters are necessary for something, then you should state what that is for (_i.e._ kinetic parameters essential in ....)

- Line 34 - 'formal potential' is a thermodynamic quality, not a kinetic parameter

- Line 35 - 'standard rate constant'? Rate constant is a more vague term, so best to avoid confusion

- Line 35 - 'widely' is redundant here

- Line 35 - Last sentence is optional in my view, you already talked about simulations. Maybe combine with that sentence?

Best Regards,
MoonshinetheP
