### 2025-07-01 Woffee/Envismetrics#9 @MoonshinetheP
Manuscript - Statement of Need
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

### General comments

- A big problem in this section is the comparison of your software with instrument-specific proprietary software. This ignores the fact that unless programmatically controlling the instrument, you still need this software to obtain the data. What you should be comparing your software to is the standard methods of processing and analysing the data (as stated in the manuscript, Excel - but also SigmaPlot and Origin to give some common examples). In the context of Data Formats, as you say Autolab-specific formats, whereas your software is capable of using .xlxs, .txt and .csv - but you have to get the data in those formats from the Autolab software. So yes, whilst you can save .nox files, I don't know of anyone that is not extracting the data in some plaintext format and processing this.

### Specific corrections

- Line 39 - See previous comments about 'electrochemical kinetic analysis'

- Line 40 - I will say here, but it applies everywhere: a potentiostat is an electronic circuit, not an instrument. Since it is only an electronic circuit, a potentiostat cannot run any software or provide data formats. Usually the instruments themselves are called electrochemical workstations (although potentiostat has unfortunately drifted into common usage)

- Line 41 - 43 - Sentence structure is a bit off here.
-
-  Line 54 - I'm not sure I would call the Figures provided by the software 'publication-ready'. If you want this, you might need to provide some editing capablilites and a method of saving the figures.

- Line 59 - Your software does not facilitate data collection, just analysis

- Line 66 - 67 - It is perhaps a bit bold to call your software 'a leading tool in electrochemical analysis'. Maybe tone it down to say it's a 'powerful tool for electrochemical analysis'?

- Table 1 - A little hard to read; could do with spacing between rows and no hyphenation if avoidable.

- Table 1 - Some of the appraisals of NOVA and Envismetrics are highly subjective in nature. I will not comment on which, because as I say I think you should compare vs. plotiing/analysis tools instead. Perhaps if you like, we can get into more details in the comments.

- Table 1 - One point here in your favour. When you compare against other plotting/analysis software and not instrument control software, you will see that Origin is not available for MacOS (you say here NOVA is not either, but then you need a Windows PC to get your data).

- Table 1 - There is no label on the caption (i.e. Table 1: ...)

Best Regards,
MoonshinetheP
