### 2025-07-01 Woffee/Envismetrics#10 @MoonshinetheP
Manuscript - Current Functions
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

### General comments

- There are not too many things to point out here that have not been brought up in previous Issues. It would be nice however if you placed the relevant Figures exactly where you discuss each module (TL & TR of Fig 2 in RDE section, BL & BR in CV section) and add a figure showing CA module.

- Please put all the equations used in the manuscript (Cottrell, Levich, Koutecky-Levich are missing) and definitions for variables and constants where it is not clear what they are (or have been already defined).

### Specific corrections

- Line 71 - 77 - Change according to previous comments on Statement of Need.

- Line 81 - 82 - It is hard to know that you need to enter a sigma value of 0 to avoid Gaussian filtering, so maybe say here (as you say here that it is optional, but that is not clear when using). Also, maybe as this is the first time you discuss smoothing the data, maybe it is worth explaining the process briefly and describing what exactly sigma is doing (other than giving smoother results with larger values). Could also give some warning not to smooth too much at risk of losing features.

- Line 84 - Maybe 'Levich and Koutecky-Levich analyses are essential for analysing electrochemical reactions under laminar flow convection conditions'? Or if you want to put more specific reasons for using each equation, 'Levich analysis can be used to calculate diffusion coefficient (or other non-kinetic properties)' and 'Koutecky-Levich analysis is used to calculate the standard rate constant'. They are not both used solely for determination of diffusion coefficient (KL is not used for this at all).

- Line 85 - 87 - This is incorrect. Levich equation is a mass transport relationship, you will see that there is no term for the potential in the equation, because you are assuming that you are at a high enough potential that electron transfer is not in any way limiting the rate of the electrochemical reaction. Only the mass transport is limiting the rate in this case. If you have no current plateau, either because of slow electron transfer kinetics or turbulent convective flow, the Levich equation cannot apply (no steady state). It also doesn't make sense to plot the Levich equation at several potentials as any potentials within the plateau will give the same plot, whilst any outside the plateau are not suitable (no mass transport limited control). Furthermore, you cannot have a plot of diffusion coefficient vs. rotation rate - diffusion coefficient is insensitive to rotation rate, and the changing limiting current under different rotation rates is a function of the Nernst layer shortening (D can be influenced by various solvent properties, but this is not relevant here).

- Line 85 - 87 - Koutecky-Levich is normally used to calculate the standard rate constant, not the diffusion coeffient. You could use KL to find diffusion coefficient, but only because the term inside the Koutecky-Levich equation containing D (i.e. the gradient) is just the Levich equation. It seems a bit off to have two options for doing essentially the same thing. KL analysis on the current plateau reduces to Levich equation and can be extrapolated back to x = 0, y =0.

- Line 88 - 89 - The Koutecky-Levich plot should be inverse current vs angular rotation rate/frequency.

- Line 89 - 93 - I think you probably get this by now, but the diffusion coefficient does not change with potential either.

- Line 96 - I've not seen the term 'rate constant value' be used before, it is usually 'scan rate'.

- Line 113  -  Empty brackets?

- Line 114 - Maybe the symbol for standard rate constant is not formatting properly?

- Line 113 - 116 - This entire section could be expanded on a little bit. Reading through Nicholson, I'm not sure how you managed to extract this dimensionless kinetic parameter - is it through peak seperation? It seems like it should only work in certain cases (not too high k0, not too low), so I am not sure how you will ensure it is used responsibly.

- Line 118 - 124 - I think it would also be good to explain here how you perform Tafel analysis from peak identification in Function 2. How do you identify the linear region?

- Line 132 - 137 - I will probably come back to this when other corrections are made. In general this paragraph is fine, but it repeats itself a bit and the symbols are not formatted.

- Line 143 - 151 - Nicely written paragraph and I like that it serves as a testimonial for the software.

Best Regards,
MoonshinetheP
