### 2025-07-01 Woffee/Envismetrics#11 @MoonshinetheP
Manuscript - References
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

### General comments

- Some references are not cited in the main body of the manuscript: Garg _et al._ (2021), Sanecki _et al._ (2003), and Wang _et al._ (2020).

- The referencing style in the main body of text is kind of all over the place. Sometimes there is no hyperlink to the references, sometimes it is only on the year, sometimes the link covers both the year and the author. Sometimes only the year is in parentheses, sometimes it is both year and author. Sometimes the references are bunched up at the end of the sentence, whereas sometimes they are in the middle of the sentence. Please make it uniform.

### Specific corrections

- Line 116 - Lavagnini _et al._ is repeated twice.

- Line 121 - 122 - Here you name Danlei Li _et al._, but the reference also gives this information. I would replace the name with the reference (if you can place the reference elsewhere other than the end of the sentence)

- Line 123 - References Bacil _et al._ by name, but gives a different year to the previous Bacil reference (I assume you mean to keep the same reference) and also references Zanello instead.

- Line 149 - 'Xue_2023'?

- Line 167 - You're not using journal abbreviations anywhere else, why here?

- Line 179 - 180 - The journal is just called 'Electroanalysis'

- Line 198 - Reference has your supervisor as 'Sadik, O. "Wunmi"', whereas they are 'Sadik, O.' elsewhere

- Line 200 - Link is broken

- Line 212 - 213 - RSC has somehow ended up in the author's field, leading to the reference being Zanello and Chemistry.

Best Regards,
MoonshinetheP
