### 2025-07-01 Woffee/Envismetrics#12 @MoonshinetheP
Conditions for uploading data need to be clearer
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

I've tried to upload some of my own data (CA .txt from Autolab loaded into .csv) into the web application, but it seemingly gets stuck on the loading page. Some comments related to this data uploading (for all modules):

- Please make it clear the upload has failed rather than giving an eternally loading progress bar.

- In general, I think that requiring RPM or scan rate to be in the filename itself is a little limiting. There should be an option to manually enter these values.

- There are a few comments in the manuscript about reading files from different instruments. To be honest this is a tricky process that probably needs you to understand how each instrument saves its data (even if it is in the same plaintext format). Compared to options like Origin, this is a huge disadvantage of your software, so maybe it is worth loading the data into a PD and showing this to the user so they can decide what to import (and what is time, current, potential etc.).

- Would be nice to be able to remove a file

- Some clearer instructions on how you want the data to be formatted is necessary
