### 2025-07-01 Woffee/Envismetrics#13 @MoonshinetheP
Insufficient warnings for improper use
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

I'm finding that it's quite easy to do incorrect analysis using this software (not ideal for intended novice audience). Some suggestions:

- For CV module, implement some check for the planar diffusion regime. Some of the data in the Test_Set folder has scan rates as low as 10 mV/s - it's highly unlikely you are still in a purely planar diffusion regime when scanning this slowly, so Randles-Sevcik is no longer appropriate (it may still work, but the equation specifically relates to planar diffusion). In the planar diffusion regime, all voltammograms should 'overlap' when current is normalised by square root of scan rate. Any voltammograms that do not should be discarded from analysis (giving user option to discard data which does not fit is in general a great idea).

- To go even one step further, warn the user that Randles-Sevcik is not the best method of determining diffusion coefficient; this would be chronoamperommetry. You are comparing analysis of one data point per scan rate, peak height (which itself is susceptible to surface roughness, potential step, iR drop), to a technique in which 100 - 1000s of data points can be fitted to Cottrell. There is no comparison, and it is of little surprise that Cottrell often gives more accurate diffusion coefficient values.

- On the same note however, it is worth warning users that Cottrell deviates at short times (due to charging currents) and long times (>40 s, due to natural convection).

- Randles-Sevcik and Cottrell are also only applicable to diffusing species under mass transport control (no poor electron transfer for Cottrell, as you point out RS has an irreversible version). No surface reactions, no adsorbed species, no non-Faradaic processes.

- Levich analysis at an RDE is only viable on the limiting current plateau. If there is no plateau (poor ET kinetics or turbulent flow), it is not applicable. If there is a plateau and then a 'drop' (this is not likely in steady state unless you exhaust the redox compound), then it probably wasn't a plateau.

- I think the standard rate constant calculation in the CV module is only valid in certain conditions, but you need to make users know what these are.
