### 2025-07-01 Woffee/Envismetrics#14 @MoonshinetheP
Add comments and descriptions to source code
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

In some places it is difficult to follow what certain functions actually do and why certain approaches have been taken. This is necessary for (1) us to assess whether the code is doing what it should be and (2) others to collaborate or develop their own code based on this software. The source code really needs better comments (#) next to lines which may be hard to follow and descriptions (''' ''') under functions that explain what the purpose is. This is a general requirement across all .py files.
