### 2025-07-01 Woffee/Envismetrics#15 @MoonshinetheP
Insufficient testing
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

The test files are not really serving any purpose in their current condition. I find it hard myself to 'automate' the tests in electrochemical software, so I will ignore this requirment, but testing is still an important step in showing your code works as expected under known and random conditions. You should be thinking of making tests that validate each aspect of the source code, or each function.

test_CA.py only seems to check that a number inside a pre-specified filename can be extracted during file upload:

- It is not super clear why this matters, but if it does then surely it would be better to test random filenames to make sure the function works as intended.

- It would be more pertinent to make a test checking the data has been extracted correctly after uploading, that correct columns are assigned, that data length has not been changed at all, that there are no NaN

- Why not check the Cottrell equation on simulated data?

test_CV.py similarly checks the filename and scan rate specified within the filename

- Why not check peak searching function on more complicated simulated data or something you know might fail?

- Also could check Gaussian filtering is working as intended.

- Here is where you could test function checking for planar diffusion in 1D and 2D simulations (if you have capabilities)

- Worth checkin RS, standard rate constant determination and transfer coefficient calculation work as intended on known or simulated data

test_HDV.py checks some basic maths which does not need testing

- In addition to similar tests as in other modules, maybe some test to identify or discern plateaus in RDE data?
