### 2025-07-01 Woffee/Envismetrics#16 @MoonshinetheP
More flexibility in determining alternative parameters
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

This is an optional thing to consider, but perhaps its worth adapting the application so that other parameters (concentration, electrode area, number of electrons) can be determined instead if desired. It would certainly broaden the appeal of your software.
