### 2025-07-01 Woffee/Envismetrics#17 @MoonshinetheP
Confusing data selection
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

Some aspects to the selection of data in the web application are kind of confusing:

- A potential range is entered manually, whereas it would be much more intuitive to select a range on a slider or something similar (could even update the plotted data). It requires reading the instructions just to understand what to do, which makes the software less intuitive.

- Discarding scan rates is very confusing - wouldn't it be better to have checked boxes or something?

- Same for cycle range. Also, it is hard to understand from the plotted data alone which scans I want to omit (and even how many scans there are)

- Sometimes there are already three entries for potential range, or diffusion coefficient. Should possibly scale with number of identified peaks.
