### 2025-07-02 Woffee/Envismetrics#2 @Grant-Xue
@MoonshinetheP

Thank you very much for your careful reading and constructive feedback. We truly appreciate your attention to detail, especially regarding authorship and transparency. We have addressed all the points raised as follows:

1. Authorship Transparency
We have added a detailed breakdown of author contributions using the CRediT taxonomy, now included in both the manuscript and the repository README.md. This clearly explains each author’s role and responsibility in the project.
[Done] See: paper.md → ## Author Contributions (CRediT Taxonomy)

2. Equal Contribution Clarification
We understand the concern regarding the equal-contrib marker for the first two authors. Based on your suggestion and a more precise review of our actual contributions, we have removed the “equal-contrib” tag and clarified the first authorship order in the contribution section.

To explain:
Huize Xue (first author) led the scientific and algorithmic development, implemented the core analysis modules in Python, performed all experimental validation, and wrote the manuscript.
Wenbo Wang (second author) focused on the online version, server deployment, and GitHub-side integration (commits are mostly his due to his role handling GitHub and the web platform).
The authorship order now reflects both the scientific contribution and technical support more clearly.

[Done] “equal-contrib” removed and clarified.

3. Affiliation Duplication
You were absolutely right—authors 4 and 5 shared the same department but were listed as separate affiliations. We have merged them into a single affiliation entry in the metadata.

[Done] Fixed in paper.md → now only four affiliations.

4. Acknowledgment Missing Author
Thank you for catching that. The omission of Xinxin Zhou in the README.md acknowledgments was unintentional and has been corrected.

[Done] Added to README.md acknowledgments section.

**Additional Clarification on Author Roles:**
Here is a brief summary of the actual division of work:

Huize Xue: Main developer of all Python-based analysis modules; led electrochemical design and theory; conducted experiments; wrote the manuscript.

Wenbo Wang: Built the web interface and hosted the platform; handled most GitHub commits as he managed the deployment.

Xinxin Zhou: Helped test user modules and wrote documentation.

Fuqin Zhou: Assisted with exploratory data formatting and figure export.

Prof. Omowunmi Sadik: Principal investigator who guided the research vision, reviewed the manuscript, and provided strategic input throughout.

We hope this clarifies the authorship arrangement. Please let us know if any further clarification is needed. Again, we appreciate your careful attention to these important matters.

Best regards,
Huize Xue (on behalf of all co-authors)
