### 2025-07-03 Woffee/Envismetrics#2 @MoonshinetheP
Hi @Grant-Xue,

Thanks for the changes.

Regarding the second point:
- I don't want you to come away with the impression that I don't think you and Wenbo contributed equally. My point was more that from solely looking at the repository commit history, it was hard to justify this claim of equal contribution. From the breakdown of work you provided here (and Wenbo provided in [#1](https://github.com/Woffee/Envismetrics/issues/1)), I would happily accept the equal contribution signifiers if you so wish to put them there. I think it would be a nice touch given the work Wenbo has put into the software (basically giving him rights to later claim that he is a kind of co-first author).

Ultimately, as always I will leave it up to you - I've already accepted the changes and will close this issue after you've seen this.
