### 2025-07-03 #8436 @mbabar09
Hello,
I was trying to run it by cloning the github and making a new environment using Anaconda (as outlined in the readme), the final run with gunicorn is throwing errors on my mac, --

gunicorn -w 4 -b 0.0.0.0:8080 app:app
[2025-07-03 13:34:00 -0400] [70552] [INFO] Starting gunicorn 23.0.0
[2025-07-03 13:34:00 -0400] [70552] [INFO] Listening at: http://0.0.0.0:8080 (70552)
[2025-07-03 13:34:00 -0400] [70552] [INFO] Using worker: sync
[2025-07-03 13:34:00 -0400] [70554] [INFO] Booting worker with pid: 70554
[2025-07-03 13:34:00 -0400] [70555] [INFO] Booting worker with pid: 70555
[2025-07-03 13:34:00 -0400] [70556] [INFO] Booting worker with pid: 70556
[2025-07-03 13:34:00 -0400] [70557] [INFO] Booting worker with pid: 70557
[2025-07-03 13:34:52 -0400] [70556] [ERROR] Exception in worker process
Traceback (most recent call last):
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/arbiter.py", line 608, in spawn_worker
    worker.init_process()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 135, in init_process
    self.load_wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 147, in load_wsgi
    self.wsgi = self.app.wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/base.py", line 66, in wsgi
    self.callable = self.load()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 57, in load
    return self.load_wsgiapp()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 47, in load_wsgiapp
    return util.import_app(self.app_uri)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/util.py", line 370, in import_app
    mod = importlib.import_module(module)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/importlib/__init__.py", line 127, in import_module
  File "<frozen importlib._bootstrap>", line 1030, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1007, in _find_and_load
  File "<frozen importlib._bootstrap>", line 986, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 680, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 850, in exec_module
  File "<frozen importlib._bootstrap>", line 228, in _call_with_frames_removed
  File "/Users/mbabar/Desktop/Papers/Review/Envismetrics/app.py", line 39, in <module>
    logging.basicConfig(level=logging.INFO,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 2003, in basicConfig
    h = FileHandler(filename, mode,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1146, in __init__
    StreamHandler.__init__(self, self._open())
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1175, in _open
    return open(self.baseFilename, self.mode, encoding=self.encoding,
FileNotFoundError: [Errno 2] No such file or directory: '/Users/mbabar/Desktop/Papers/Review/Envismetrics/logs/tnote-2025-07-03.log'
[2025-07-03 13:34:52 -0400] [70555] [ERROR] Exception in worker process
Traceback (most recent call last):
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/arbiter.py", line 608, in spawn_worker
    worker.init_process()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 135, in init_process
    self.load_wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 147, in load_wsgi
    self.wsgi = self.app.wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/base.py", line 66, in wsgi
    self.callable = self.load()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 57, in load
    return self.load_wsgiapp()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 47, in load_wsgiapp
    return util.import_app(self.app_uri)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/util.py", line 370, in import_app
    mod = importlib.import_module(module)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/importlib/__init__.py", line 127, in import_module
  File "<frozen importlib._bootstrap>", line 1030, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1007, in _find_and_load
  File "<frozen importlib._bootstrap>", line 986, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 680, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 850, in exec_module
  File "<frozen importlib._bootstrap>", line 228, in _call_with_frames_removed
  File "/Users/mbabar/Desktop/Papers/Review/Envismetrics/app.py", line 39, in <module>
    logging.basicConfig(level=logging.INFO,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 2003, in basicConfig
    h = FileHandler(filename, mode,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1146, in __init__
    StreamHandler.__init__(self, self._open())
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1175, in _open
    return open(self.baseFilename, self.mode, encoding=self.encoding,
FileNotFoundError: [Errno 2] No such file or directory: '/Users/mbabar/Desktop/Papers/Review/Envismetrics/logs/tnote-2025-07-03.log'
[2025-07-03 13:34:52 -0400] [70554] [ERROR] Exception in worker process
Traceback (most recent call last):
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/arbiter.py", line 608, in spawn_worker
    worker.init_process()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 135, in init_process
    self.load_wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 147, in load_wsgi
    self.wsgi = self.app.wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/base.py", line 66, in wsgi
    self.callable = self.load()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 57, in load
    return self.load_wsgiapp()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 47, in load_wsgiapp
    return util.import_app(self.app_uri)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/util.py", line 370, in import_app
    mod = importlib.import_module(module)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/importlib/__init__.py", line 127, in import_module
  File "<frozen importlib._bootstrap>", line 1030, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1007, in _find_and_load
  File "<frozen importlib._bootstrap>", line 986, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 680, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 850, in exec_module
  File "<frozen importlib._bootstrap>", line 228, in _call_with_frames_removed
  File "/Users/mbabar/Desktop/Papers/Review/Envismetrics/app.py", line 39, in <module>
    logging.basicConfig(level=logging.INFO,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 2003, in basicConfig
    h = FileHandler(filename, mode,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1146, in __init__
    StreamHandler.__init__(self, self._open())
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1175, in _open
    return open(self.baseFilename, self.mode, encoding=self.encoding,
FileNotFoundError: [Errno 2] No such file or directory: '/Users/mbabar/Desktop/Papers/Review/Envismetrics/logs/tnote-2025-07-03.log'
[2025-07-03 13:34:52 -0400] [70557] [ERROR] Exception in worker process
Traceback (most recent call last):
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/arbiter.py", line 608, in spawn_worker
    worker.init_process()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 135, in init_process
    self.load_wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/workers/base.py", line 147, in load_wsgi
    self.wsgi = self.app.wsgi()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/base.py", line 66, in wsgi
    self.callable = self.load()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 57, in load
    return self.load_wsgiapp()
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/app/wsgiapp.py", line 47, in load_wsgiapp
    return util.import_app(self.app_uri)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/site-packages/gunicorn/util.py", line 370, in import_app
    mod = importlib.import_module(module)
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/importlib/__init__.py", line 127, in import_module
  File "<frozen importlib._bootstrap>", line 1030, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1007, in _find_and_load
  File "<frozen importlib._bootstrap>", line 986, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 680, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 850, in exec_module
  File "<frozen importlib._bootstrap>", line 228, in _call_with_frames_removed
  File "/Users/mbabar/Desktop/Papers/Review/Envismetrics/app.py", line 39, in <module>
    logging.basicConfig(level=logging.INFO,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 2003, in basicConfig
    h = FileHandler(filename, mode,
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1146, in __init__
    StreamHandler.__init__(self, self._open())
  File "/Users/mbabar/miniconda3/envs/envismetrics/lib/python3.9/logging/__init__.py", line 1175, in _open
    return open(self.baseFilename, self.mode, encoding=self.encoding,
FileNotFoundError: [Errno 2] No such file or directory: '/Users/mbabar/Desktop/Papers/Review/Envismetrics/logs/tnote-2025-07-03.log'
[2025-07-03 13:34:52 -0400] [70556] [INFO] Worker exiting (pid: 70556)
[2025-07-03 13:34:52 -0400] [70555] [INFO] Worker exiting (pid: 70555)
[2025-07-03 13:34:52 -0400] [70554] [INFO] Worker exiting (pid: 70554)
[2025-07-03 13:34:52 -0400] [70557] [INFO] Worker exiting (pid: 70557)
[2025-07-03 13:34:52 -0400] [70552] [ERROR] Worker (pid:70556) exited with code 3
[2025-07-03 13:35:22 -0400] [70552] [ERROR] Shutting down: Master
[2025-07-03 13:35:22 -0400] [70552] [ERROR] Reason: Worker failed to boot.
