### 2025-07-03 Woffee/Envismetrics#2 @Grant-Xue
Hi @MoonshinetheP,

Thank you for the clarification and for your understanding.

We appreciate your thoughtful perspective on authorship transparency. After internal discussion and based on your feedback, we have decided to keep the equal-contrib markers for both Wenbo and myself. As you mentioned, the project reflects a joint effort between electrochemical method development and web implementation, and we believe equal contribution best represents that collaboration.

Thanks again for your thorough and constructive review throughout this process!

Best regards,
Huize
