### 2025-07-03 Woffee/Envismetrics#4 @Grant-Xue
Hi @MoonshinetheP,

Thank you very much for your thoughtful suggestions regarding the repository structure — they were very helpful.

We’ve completed a thorough cleanup and reorganization of the project files based on your feedback. Here's a summary of what we’ve done:

✅ Repository structure revised
All high-level individual configuration and metadata files (e.g., README.md, LICENSE.txt, pytest.ini, config.py, etc.) remain at the top level.

The source code has been grouped under a consistent folder (scr/) structure with BaseModule.py, CV.py, HDV.py, CA.py, app.py, as well as static/ and templates/ directories for the web app.

Tests are placed in a dedicated folder, with corresponding test data in data/test_data/.

Example input files (i.e., templates for CV, HDV, CA formats) are placed in data/example_files/, to help users avoid misformatted uploads.

All JOSS submission materials (paper.md, bibliography.bib, images, etc.) are now in a clean joss/ folder.

The outdated /Test_Set/ directory has been fully removed, and its contents reorganized under the new data folders.

📁 New documentation
We’ve also updated the main README.md to include a tree view of the directory layout and explanations for each folder’s purpose, so new users can navigate easily.

Thanks again for your time and dedication to improving the clarity and usability of this project. Please let us know if there’s anything else you’d like us to tweak!
