### 2025-07-04 #8436 @Grant-Xue
@mbabar09 Thanks for catching this! The error occurred because the logs/ folder didn’t exist yet on your machine — app.py attempts to create a log file but fails silently when the directory isn’t found.
Let me add a code to automatically create the logs/ directory if it’s missing, so you won’t run into this again. Thanks for reporting!
