### 2025-07-04 #8436 @mbabar09
I added 7 xlsx files with different scan rates in this folder, its still getting stuck at Function 3 in the CV module.
