### 2025-07-05 #8436 @Grant-Xue
@mbabar09 Thank you for testing it with multiple scan rates!

You're right — sometimes Function 3 may get stuck temporarily due to a known issue with cached state not resetting properly after file uploads. We’re actively working on a fix for this.

🔄 Temporary solution:
Please try refreshing the page (or clearing the upload and reloading the files). This usually resolves the issue and allows Function 3 to proceed normally.

Thanks again for your patience and for helping us improve the tool!
