### 2025-07-05 Woffee/Envismetrics#5 @Grant-Xue
@MoonshinetheP To reproduce Figure 2 – HDV subplots 1 and 2, follow these steps using the online tool:

✅ Example Dataset
Folder: data/test_data/02272024_HDV_D41

File type: .xlsx files for linear sweep voltammetry under different rotation speeds (rpm)

Each file represents one experimental curve with a unique rotation speed

✅ Online Module
Module: Hydrodynamic Voltammetry (HDV)

After uploading all files in the folder and submit, the interface automatically moves to HDV-2

⚙️ Parameter Configuration (HDV-2 Page)
All values shown below are the default settings of the online tool, and they match the example data used in Figure 2. Users do not need to adjust them unless using custom data.

| Parameter                           | Default Value                     | Explanation                                                                                           |
|-------------------------------------|-----------------------------------|-------------------------------------------------------------------------------------------------------|
| **Concentration of solute (C)**     | `0.000000894454 mol/cm³`          | Bulk concentration of analyte (e.g., DMAB or Fe²⁺) in the electrolyte solution                        |
| **Electrode surface area (A)**      | `0.15 cm²`                         | Geometric area of the working electrode, used to calculate diffusion-limited current                 |
| **Kinematic viscosity (ν)**         | `0.01 cm²/s`                       | Assumed value for aqueous solution at ~25°C                                                           |
| **Number of electron transfer (n)** | `1`                                | Number of electrons involved in the redox process (e.g., Fe²⁺ → Fe³⁺ → n = 1)                         |
| **Method**                          | `Levich plot and Levich analysis` | Chooses the type of plot and calculation (Levich equation-based analysis of limiting current)        |
| **Applied potential range (V)**     | `(-1, 1)`                          | Voltage window from which currents are sampled, typically chosen to include the limiting region      |
| **Number of potentials to display** | `9`                                | Number of equally spaced potentials to extract from each curve. This determines how many lines appear in Figure 2 Subplot 1 |
| **Potential step interval**         | `100` (mV)                         | Distance between sampled potentials. This controls the **dot density** seen in Figure 2 Subplot 2    |

✅ Continue to HDV-3.1
proceed to the Levich Plot and Levich Analysis page
This is the page where Figure 2 HDV subplots 1 and 2 are generated

📈 Output Description
When using the above configuration with the folder 02272024_HDV_D41, the tool generates:

Subplot 1 (Figure 2A): Levich plot → Linear fit of limiting current vs. ω¹ᐟ²

Subplot 2 (Figure 2B): Levich analysis output → Diffusion coefficient and plot residuals

Note: The dataset contains at least 5 rotation speeds, making it sufficient for meaningful Levich analysis. All curves are aligned with default parameter assumptions.

When using your own dataset, if you're unsure how to set a parameter, just click the **"?"** icon next to it — each one explains what the field means and how to choose values.
