### 2025-07-05 Woffee/Envismetrics#5 @Grant-Xue
To reproduce the results in Subplots 3 & 4:

Go to the CV module in the online tool.
Upload all .xlsx files from the 06102024_CV_D_AF folder.
Keep all parameter fields at their default values.

I'm very happy to introduce more details if you have any further questions or concerns!
