### 2025-07-06 Woffee/Envismetrics#5 @MoonshinetheP
@Grant-Xue

When I wrote this issue, I was in the process of taking a brief look through the checklist before starting a more thorough read of your manuscript. As such, the comments I made regarding the Levich/Koutecky-Levich analyses when reviewing the manuscript should take precendence.

In general, my point was more that I want you to write somewhere (either in the manuscript itself or in the markdown with reference in the manuscript) what exact files were used to generate this data, what settings were used, and ideally, what the data actually represents (electrochemical cell, reaction, peak identification etc.). This makes it more useful for developers wanting to test the capabilities of your software.

Regardless, some comments on your suggestions:

- The data you've suggested for linear sweep voltammetry on an RDE is not the same as seen in Figure 2 a+b of the manuscript. The regression performed on the 'Limiting current' vs. Rotation Rate data is different and the current vs. potential plot is also not the same (you can recreate it using your 'diffusion coefficient' plot in Fig. 2b, since D is proportional to i in Levich).

- There is also something very strange going on with the 'Potential step interval' field in the HDV-2 Envismetrics stage. You said to select 100 mV, but the plots are giving more like 215 mV steps. It is also clear that Fig. 2b is not using this 100 mV, there are a lot more data points.

- Disregarding the fact that the data you suggested is (1) seemingly of a noble metal (maybe gold) in hydroxide (and possibly something else, I see DMAB in the filename) and therefore not entirely allowed in RS analysis (no diffusion, although maybe DMAB is okay, not familiar with it) and (2) performed at too low scan rates for RS to apply on most electrodes used in the laboratory, I don't get the same plot in Fig. 2c (not the same appearance exactly, not the same current). Also it is seemingly failing to identify a pair of peaks around -0.15 V. For the RS plot, I obviously don't get any of the simulated data you seem to have plotted in Fig. 2d, although I do get RS analysis for reduction peaks, which is not in the manuscript.
