### 2025-07-06 Woffee/Envismetrics#18 @MoonshinetheP
Fix CA module so that it plots multiple data sets from t = 0
This issue is related to [JOSS review #8436](https://github.com/Woffee/Envismetrics/issues/github.com/openjournals/joss-reviews/issues/8436).

@Woffee @Grant-Xue

When multiple current transients are uploaded, they appear concatenated on the 'Potential vs. Time' and 'Current vs. Time' plots of CA-2. This gives the impression of multi-step choronoamperommetry, not multiple single steps. I imagine some users may want to use the software to plot two or more transients in the same time (although in this case, I am not sure potential vs. time makes sense unless you give each line a different color - although even then, unless Envismetrics can deal with multistep without changing regression time range, there is not much interesting in potential vs. time of CA anyway).

In the next step (Cottrell and regression), each individual transient has its own plot, which is also a lot if processing many transients, but is better - although this should also allow multiple parameters to be entered for different transients (say one transient has concentration of A, other has concentration of B, you want to compare).

Also, minor correction but since in both cases the y-axis is the function of the x-axis (i.e. potential/current is a function of time), so the labels should be as 'Potential vs. Time' and 'Current vs. Time', not 'Time vs. Potential' and 'Time vs. Current'.

By the way, I saw that one reviewer mentioned no data in the CA output - I also noticed it before, but I must have forgot to mention. Anyway, the point is that the fit of the '1st interval' is being overwritten by first column or just skipped. It is not present when analysing multiple transients.
