### 2025-07-06 Woffee/Envismetrics#18 @MoonshinetheP
Also, was thinking of making another issue for the weighting of Cottrell fit, but I think it is more a ease of use issue. If you take some CA data from test folder (I took all files from 05132024_CA_D) and plot, you see that the R-squared is really low on almost all the plots (the Cottrell fits are bad). In general this is because there is more data at later t when you plot 1/sqrt(t), and since pure diffusion is spoilt by convection at later times, this produces a bad fit with earlier data. Would be good to weigh data more for analysis at short times (avoiding charging of course), but I suppose that can also be done with the regression time range. I think a selection bar would be more intuitive than a float between 0 and 1, but in any case something on the plot to indicate where the fit is made would be good.
