### 2025-07-06 Woffee/Envismetrics#18 @MoonshinetheP
Final comments on CA for now:

- Seems like all data is being plotted as 120 s regardless if the data is shorter or not (could also be that longer transients are being repeated and shorter ones ignored - seemed fine in CA-2)

- Concantenated transients are because it it plotting time, not corrected time

- Only number of electrons and concentration seems to change diffusion coefficient determination. Area has no effect, and regression time range also appears to make no difference.

- What is strange is that Cottrell plot changes when concentration or number of electrons changes. It seems you plot current vs. the part of the Cottrell equation without sqrt(D). Strange way of doing it, instead of plotting current vs. sqrt(t) and accounting for constants in gradient calculation.
