### 2025-07-08 Woffee/Envismetrics#5 @Grant-Xue
Hi @MoonshinetheP,

Thanks again for your detailed feedback!

We've now added a complete walkthrough to reproduce all subplots in Figure 2 of the manuscript. You can find it here:

📄 [Reproducing Figure 2](https://github.com/Woffee/Envismetrics/blob/main/data/reproducibility_guide.md)

This guide includes:

1. Exact filenames used for each subplot
2. Detailed parameter settings used in each module (HDV and CV)
3. Matching outputs and how to interpret them
4. Notes on default values, dataset requirements, and edge cases

The reproduction file is located directly in the [data/test_data](https://github.com/Woffee/Envismetrics/tree/main/data/test_data) folder for easy access, and we’ve updated the main README.md to link to this as well.

Thanks again for helping us make this better!
