### 2025-07-08 Woffee/Envismetrics#7 @Grant-Xue
Thank you for the thoughtful feedback.

We've revised the summary accordingly to align with your suggestions, particularly:

- Removed subjective terms like “innovative” and “efficient”
- Replaced “software suite” with “application”
- Clarified that the software automates routine analysis of electrochemical data rather than only performing kinetic analysis
- Highlighted that modules cover CV, CA, and HDV techniques
- Removed unnecessary phrasing (“ease of use”, “simplifying”) in favor of clearer descriptions of the graphical web interface and its accessibility

We’re happy to incorporate these improvements and appreciate your help in making the manuscript clearer and more precise.

If the inclusion of a summary is not permitted by JOSS, we understand and are prepared to remove or adjust it accordingly.

Best regards,
