### 2025-07-08 Woffee/Envismetrics#8 @Grant-Xue
Thank you very much for your thoughtful and detailed suggestions regarding the Summary section. We've carefully addressed each of your points with the following revisions:

- Replaced the term "hydrodynamic voltammetry" with "linear sweep voltammetry using a rotating disk electrode (LSV at RDE)" to avoid overgeneralization.
- Clarified that LSV at RDE is conducted under laminar flow and planar diffusion conditions.
- Changed "thermodynamic constants" to "thermodynamic properties" to avoid confusion with universal constants like F or R.
- Rephrased "predicting and optimizing redox reactions" to "understanding redox reactions", making the objective more logically sound.
- Replaced "These kinetic parameters" with "These values" and explicitly included formal potential, diffusion coefficient, and other relevant parameters.
- Adjusted phrasing about simulation use: instead of suggesting it determines mechanisms, we now state that it helps validate reaction pathways.
- Removed the sentence implying many parameters are unavailable in literature; replaced with: "Although literature values exist for some compounds, experimental evaluation is necessary for new systems."
- Combined the technique discussion for clarity: "Techniques such as CV, LSV at RDE, and CA..." rather than separating CA.
- Replaced "essential kinetic parameters" with either "fundamental" or direct mention of the parameters (e.g., diffusion coefficient, formal potential).
- Standardized the terminology by using "standard rate constant" consistently throughout.
- Removed the final redundant sentence about simulations to streamline the paragraph.

Please let us know if any further adjustments are needed—we greatly appreciate your insights and feedback!
