### 2025-07-08 Woffee/Envismetrics#9 @Grant-Xue
@MoonshinetheP

Summary of Changes in Statement of Need section:

- Refocused comparison scope: We shifted the comparison from instrument-specific software (e.g., NOVA) to widely used data analysis tools such as Excel, Origin, and SigmaPlot, in line with your suggestion.
- Clarified data acquisition role: We now explicitly state that Envismetrics does not perform data collection, and users must first export data from proprietary software (e.g., Autolab) into plaintext formats such as .csv, .xlsx, or .txt.
- Revised terminology: "Electrochemical kinetic analysis" has been changed to "automated analysis of electrochemical data" to better reflect the scope of our software and avoid overstatement.
- Adjusted figure description: Rather than calling them "publication-ready," we describe Envismetrics plots as clean and exportable, while acknowledging that further formatting may be required for final publication.
- Neutralized promotional language: The term “leading tool” has been replaced with a more objective and descriptive phrase: “modular, web-based platform.”
- Table 1 revisions:
1.   Reformatted for clarity and consistency, with better alignment across rows and tools.
2.   Removed subjective or ambiguous phrasing; focused on objective capability assessment.
3.   Added explicit rows for data collection and publication output, addressing your comments on functional boundaries.
4.   Terminology correction: We replaced incorrect references to "potentiostat software" with "electrochemical workstation software", acknowledging that potentiostats are circuits and not standalone software platforms.

Thank you again for your constructive feedback—we’ve incorporated it thoroughly, and we’re happy to revise further if needed!
