### 2025-07-10 Woffee/Envismetrics#12 @Woffee
Dear MoonshinetheP,

Thank you for your comments. I have now completed the following feature:
- [x] Would be nice to be able to remove a file

I will continue to make updates.
