### 2025-07-11 #8436 @Grant-Xue
@rena-96 @MoonshinetheP Thank you for your reminder. We are working on the new ability to remove a file, causing a chain effect. We are working on solving this issue.
