### 2025-07-11 #8436 @oliverrdz
Hello.

So far it seems a useful tool. Congrats! I have reviewed the manuscript and created issue [20](https://github.com/Woffee/Envismetrics/issues/20). Please address the comments regarding electrochemical concepts, as they are well known by the community plus they would serve for educational purposes, something the software claims. As a proof of this, @MoonshinetheP has raise similar concerns in issue [10](https://github.com/Woffee/Envismetrics/issues/10).

I also opened an issue asking to update the README file, the structure section is different than the one implemented.

I will test the software locally tomorrow. If it is possible to have the server running so I can try it, that would be nice, otherwise I will only check it locally. By the way, I advice to run the flask app on an actual server service such as apache or nginx with WSGI when publishing the paper. Or maybe having a docker file for easy installation. Having an exposed port (the 8080 in http://34.162.1.1:8080/) is a huge security risk, not to mention that this is not signed (https). If this is too difficult or it would take too long to implement, I suggest you don't mention that it can be used live and instead only mention that you are working on a live demo.
