### 2025-07-14 #8436 @Grant-Xue
@oliverrdz I just checked — the server is running at the moment. Thank you again for your deployment feedback.

You're absolutely right: the current live instance (`http://34.162.1.1:8080/`) is a **temporary testing endpoint**, and we do **not** intend to maintain it as a production server. We fully agree that exposing a raw port is a security risk.

We are now working on:
-  Removing the live link from the manuscript and marking the demo as “under development”;
-  Providing a Dockerfile to allow safe and local deployment;
-  Planning for future deployment using WSGI with Nginx or Apache for proper server hardening.

Thanks again for pointing this out!
