### 2025-07-14 Woffee/Envismetrics#12 @Grant-Xue
Hi @MoonshinetheP,
Thank you for your thorough feedback — here are our point-by-point responses:

1. Upload Failure Clarity
We agree that an endless loading bar can be confusing. We are adding a new file validation step after upload:
if the uploaded file lacks necessary structure or key columns (e.g., Time (s), Current (A), Potential (V)), a warning message will be displayed to the user. This will help clearly distinguish between successful and failed uploads.

2. RPM or Scan Rate Required in Filename
We understand your concern. In our typical workflow, scan rate or RPM is not included inside the raw data files exported from electrochemical software (e.g., NOVA, EC-Lab, or Gamry).
Therefore, users often include these values in the filename as a manual record to avoid ambiguity during analysis. While this is not ideal, it remains the most consistent automatic identifier at present.
We are exploring options to allow manual entry of scan rate/RPM in future updates.

3. Compatibility with Data from Different Instruments
This is an excellent suggestion. Building a universal interactive file parser would indeed improve flexibility. However, such a feature requires significant development effort and interface design, which are currently beyond our team's available resources.
For now, we ensure compatibility by manually supporting and documenting the output formats of major software platforms.
We agree this is a limitation compared to tools like Origin, and we’ve added this feature to our long-term development roadmap.

4. File Removal Feature
Agreed — this has been implemented. Users can now remove uploaded files within the session without needing to refresh the page.

5. Clearer Data Format Instructions
We have added a dedicated [README_data_structure.md](https://github.com/Woffee/Envismetrics/blob/main/data/README_data_structure.md) file to the repository.
This document includes:
- Expected file formats and required column headers for each module;
- Example input tables;
- Notes on common formatting issues and how to resolve them.

Let us know if there’s anything else we can clarify or improve further!
