### 2025-07-14 Woffee/Envismetrics#16 @Grant-Xue
We appreciate this thoughtful suggestion. Indeed, equations such as Randles–Ševčík and Cottrell can be algebraically rearranged to solve for various unknowns—such as analyte concentration, electrode area, or number of electrons transferred—provided that the remaining parameters are known and reliable.

While our current implementation is primarily geared toward estimating the diffusion coefficient
𝐷
D, which reflects the most common use case in our research, we agree that enabling users to solve for other variables would significantly enhance the software’s educational and practical utility.

We have added this feature to our development roadmap as part of a planned “Advanced Mode”, which will allow users to specify the parameter they wish to compute. This enhancement aligns well with the educational goals of our platform and will help broaden its applicability to a wider user base.
