### 2025-07-14 Woffee/Envismetrics#14 @Grant-Xue
Thank you for your feedback.

We have now added detailed inline comments (#) and docstrings ('''...''') throughout the codebase to improve clarity and accessibility. This includes both line-level explanations for non-obvious logic and function-level documentation describing the purpose, inputs, and outputs of each major method.

The following core modules have been fully updated with these enhancements:

[CA_commented.py](https://github.com/Woffee/Envismetrics/blob/main/src/CA_commented.py)

[CV_commented.py](https://github.com/Woffee/Envismetrics/blob/main/src/CV_commented.py)

[HDV_commented.py](https://github.com/Woffee/Envismetrics/blob/main/src/HDV_commented.py)

We hope this improves the readability and reusability of the code for reviewers and future users.

Please let us know if further clarification or documentation is needed.
