### 2025-07-14 Woffee/Envismetrics#16 @MoonshinetheP
@Grant-Xue

Adding to the development roadmap is certainly a good idea for now, although I don't see whether you have a separate development roadmap somewhere or you just have not updated the one in the readme.md.

I will close the comment for now, but I want to point out one thing:

- Although this issue is just a suggestion, bear in mind that it might be harder to adapt your code to accommodate such an approach   (or any similar update) later on if your code is written in more 'specific' or 'concrete' fashion. So, I don't expect you to change anything for the review, but just as some advice - consider whether the code is flexible for such updates and if possible make it more adaptable for the future (I will try and go through code at some point, but you know your own work best and can make a better judgement)
