### 2025-07-15 Woffee/Envismetrics#18 @Grant-Xue
@MoonshinetheP

- Part 1. Clarification regarding the "concatenated" appearance in CA-1/CA-2 plots

Thank you for raising this observation.

We’d like to clarify that the apparent concatenation of current transients in the 'Potential vs. Time' and 'Current vs. Time' plots under CA-2 is due to the structure of the input dataset, rather than the software explicitly stitching data together.

For instance, in the test folder 05132024_CA_D, there are 49 .xlsx files. These originate from 7 separate experiments, each consisting of 7 transients labeled with the same leading number (e.g., 1_..., 2_..., ..., 7_...). These files are organized sequentially in terms of time — for example:

2_DMAB_2s_CA.xlsx: first timestamp ~120.77 s

3_DMAB_2s_CA.xlsx: first timestamp ~123.77 s

This creates the illusion of a continuous time series, but no concatenation is being performed by the software. The files are simply being plotted in the order they appear, with raw time values intact.

To avoid this stacking behavior (and misinterpretation as multi-step chronoamperometry), we recommend that users upload only one transient per index (1–7) per analysis session. Duplicate indices (e.g., multiple files starting with 2_...) may result in overlapping curves in CA-1/CA-2.

We’ll consider adding a warning or data validator to flag duplicated prefixes in future updates.

- part 2. Feature Limitation: Global vs. Per-Transient Parameters (CA-2)

Thanks for this very helpful suggestion.

You're absolutely right — in the current implementation, key parameters such as the number of electrons (n) and concentration (c) are treated as global values, applied uniformly across all uploaded transients. This works for datasets with identical experimental conditions, but limits the utility when comparing multiple transients with varying concentrations or reaction stoichiometry.

Planned Solution

We are now extending support for per-transient parameter input via an "Advanced Mode." In this mode, users will be able to input a list of values (e.g., c = [1e-3, 5e-4, 2e-3, ...]), where the list length matches the number of uploaded files. Each entry in the list will be mapped to the corresponding transient.

For now, this will be implemented via JSON-style list input in the UI, but in future updates, we plan to improve usability further by introducing a user-friendly parameter input table (e.g., spreadsheet-style or uploadable CSV).

The goal is to prioritize functionality first (i.e., enabling correct and flexible computation), followed by enhancements to the user interaction layer. What do you think?

- Part 3. Y-axis Label Format (CA-1 & CA-2 Plots)

Thanks for pointing this out — we’ve updated all figure labels and titles to follow the correct Y vs. X convention, e.g., "Potential vs. Time" and "Current vs. Time".

This ensures consistency with standard scientific plotting practices.

Status: Working.

- Part 4. Missing Data in CA-2 Output (Interval 1 Skipped)

Thanks for raising this — the skipping of the first interval is intentional by design.

In typical chronoamperometry experiments, the first potential step (often ~120 s in our data) is used to precondition the electrode, for example:
1. to remove residual adsorbates or surface films,
2. to stabilize the diffusion environment near the electrode,
3. and to ensure consistent initial concentration profiles before analysis.
This initialization period is not included in the regression-based diffusion analysis because its current transient may not reflect pure diffusion behavior, and including it could lead to misleading Cottrell fits.

Therefore, we intentionally exclude the first interval and analyze from the second step onward to ensure the calculated diffusion coefficients are meaningful.

We agree that this behavior should be better documented or even made optional. We'll update the code comments and consider adding a user toggle/flag to include all intervals if desired.

Status: Will improve documentation

- Part 5. Low R² in Cottrell Fits and Weighting Suggestion

Thank you for the thoughtful observation.

We agree that in many chronoamperometric measurements, especially in unstirred or semi-open systems, late-time data can be heavily influenced by convection, which affects the linearity of the Cottrell fit. This is a well-known issue in electrochemical analysis, and we appreciate your pointing it out.

Currently, Envismetrics allows users to define a custom fitting range via a float-based [start, end] input, which is effective in excluding problematic time regions from the fit (e.g., very early charging or late convection). However, we acknowledge that this input method lacks visual feedback and is not as intuitive as a GUI slider.

Planned improvements:
In the next update, we will visually highlight the fitted region directly on the regression plots (e.g., colored points or shading) so users can see exactly which portion of the data is used in the Cottrell fit.
We may also consider adding a toggle or preset range selector in the interface for commonly used fit intervals (e.g., "early-time", "mid-time", etc.).

On weighting early-time data:
We recognize the merit of weighted regression (e.g., giving more weight to early-time points) and its potential to improve fit accuracy. However, this approach introduces new assumptions about weighting functions and introduces variability across datasets, especially when the time resolution is inconsistent.

To avoid overcomplicating the user experience or introducing non-transparent heuristics, we currently keep the regression unweighted but allow explicit user control over the fitting region.

That said, we are open to adding a weighted fit option in an "advanced mode" in the future, where users can choose their own weighting scheme if they are aware of the tradeoffs.

- Part 6.  Uniform 120s Plot Length – Clarification

Thanks again for the detailed observation.

Indeed, many current transients shown in CA-1 plots appear to have a duration of exactly 120 s, and this may give the impression that data is either being truncated or artificially extended.

 Clarification:
This is not an error — it's by experimental design.
In our chronoamperometry datasets, the first file for each experiment (usually those with filenames starting with "1") is deliberately set to 120 seconds to serve as a pre-conditioning step. The goal is to:

1. Initialize the electrode/electrolyte interface,
2. Remove residual adsorbates or contaminants,
3. Establish a baseline diffusion layer before the main measurements.

This design helps ensure consistent diffusion conditions before the key potential pulses are applied.

📁 For example, in 1_DMAB_2s_CA.xlsx, the time column ranges from 0 to 120 s.
Subsequent steps (e.g., 2_DMAB_2s_CA.xlsx, etc.) contain much shorter potential pulses (e.g., 2 s, 10 s, etc.).

There is no artificial truncation or repetition — the data durations reflect the real experimental time of each transient.

- Part 7: X-axis Choice in Cottrell Plot (Bt vs. I Instead of 1/√t vs. I)

Thanks for raising this important point.

You're absolutely right — in traditional Cottrell analysis, the current is plotted against 1/√t, and the diffusion coefficient D is derived from the slope. In Envismetrics, we define: B(t) and plot I vs. B(t) directly.

This allows the slope of the linear fit to directly equal √D, so that:

This was intended to simplify the user's workflow — no additional back-calculation is needed. This makes the output more immediately actionable, especially for users with limited background in electrochemistry.

What we will do:
We agree this non-standard axis might be unintuitive for some users. Therefore:
We will add a note in the documentation to clearly explain that our Bt axis is equivalent to 1/√t but scaled by known constants.
Optionally, we may add a toggle to allow plotting I vs. 1/√t in a future update for better educational value and physical intuition.

On Plot Duplication:
In many practical cases, users analyze chronoamperometry involving multiple potential steps (e.g., 7 platforms). Since Envismetrics already generates 7 individual regression plots (one for each step), duplicating these plots in a second format (e.g., I vs 1/√t) would significantly increase visual clutter without adding real value.
Our goal is to keep the interface clear, focused, and actionable, especially when processing a large number of transients. Instead of duplicating, we opt to explain the axis meaning clearly and offer the possibility of an optional toggle in the future.
