### 2025-07-15 #8436 @MoonshinetheP
@Grant-Xue

I was a bit hasty to send the link about Triage access, as I did not read beforehand that this is a role specific to organisations (or at least that is what it seems like). The only way for us to be able to assign labels is for us to be added as collaborators, which I am not certain is allowed if we are reviewing (maybe someone else has some more informed opinion on this?). Perhaps it is best for now that we leave it up to you to decide the urgency of each issue.

I want to bring something else up.

Something about your message has made me a little suspicious. I hope I am wrong, but it seems quite clear to me the above message is at least partially AI generated (using a large language model - LLM). For example:

> It absolutely makes sense to help streamline issue labeling and prioritization, especially for distinguishing between suggestions and bugs.

- LLMs tend to add this kind of sentence - grammatically correct, but adding very little context or content; oddly complex in some places and certainly not how a human would write such a sentence (I don't think any human would even write such a sentence, its completely unneccesary).

> Please let us know your GitHub handles if needed

- You wrote my GitHub 'handle' already at the beginning of the message...
- Again, LLMs are writing quite vague instructions (i.e. adding 'if needed') where humans would write much more concise requests

>  (if not already contributors)

- I mean .... come on ....

Before we get carried away, I want to make some things clear. Firstly, I don't mean this to be an accusation, but rather as a way to impart some advice on any potential use of LLMs. Secondly, I can see multiple valid reasons for using an LLM in the reviewing process:

- **To overcome a language barrier** - I'm a native English speaker myself, but many of the people I work with are not and rely on LLMs to correct their text. I see no problem with it, and find it equivalent to autocorrect on Word for all intents and purposes.

- **As an advanced search engine** - I myself and many of my students use LLMs to avoid reading lengthy documentation in situations where that is an unneccesary step to understanding a problem. I think this is applicable when faced with comments from reviewers that may be outside your expertise.

- **To help structure replies** - On the fence of LLM use here, but I do think that if you have some long reply to make, some template made by LLM is valid for helping you structure your thoughts

Furthermore, as far as I understand, [JOSS ](https://joss.readthedocs.io/en/latest/review_criteria.html#using-ai-to-generate-code) has no position on the use of LLMs to generate code, and I assume this extends towards other aspects of manuscript submission. All to say - if you are using LLMs in the reviewing process, don't hide it and make sure you use it responsibly. You've already done the hard part (making the software), the review is meant to be the easy bit and you shouldn't feel the need to take shortcuts.

On that last note of responsible use, some restrictions:

- Anything that you use an LLM to summarise during the review process, you need to 100% understand what it outputs and you probably need to show some ability to edit the output to fit your needs. This applies to comments in the review process and code written in the repository.

- This should apply doubly so when it comes to changes in the code. There is a new trend of 'vibe coding', but if you use this, you still need to validate the approach the LLM is taking makes sense. I've seen some odd functions in the source code which are doing strange things that can be done in much easier ways - maybe human error, but maybe LLM giving you strange way of doing something (I sometimes use LLMs to suggest approaches to some particular code with mixed results - sometimes it takes the non-standard route).

- On a personal level, I have at least tried to put some effort into reviewing your manuscript and making comments that I hope will help improve your software. Therefore, I'm not sure I would appreciate my work being replied to by an AI model (it is actually quite a disrespectful practice - would not be fair the other way round either, if the reviewer didn't give time/effort). I would rather read a reply from YOU, even it is brief and to the point, and regardless if their are any language barriers or any other aspects to such replies you are trying to avoid.

If none of this applies, then ignore it. I am however, reasonably certain that several aspects to the submitted work and the review process are reliant on LLM usage. I don't 100% trust such tools, but an 'AI detection' application flagged several comments on the issues I've raised (in general, the tone of your comments throughout the reviewing process has also made me a little suspicious) and _even some of the manuscript_ as potentially generated by an LLM. I'm also a bit suspicious of the commented variations of the .py files, partly because you uploaded them as separate files and also because of how clean they look; but as long as they are okay and you understand what is written, I am fine with it. Again, it's not 'illegal' to use an LLM to aid you in coding/writing/reviewing, but it is ethically a little murky and requires careful attention to make sure you understand what is being written.
