### 2025-07-16 Woffee/Envismetrics#13 @Grant-Xue
@MoonshinetheP We have made the following change
1. Applicability of Randles–Ševčík Equation and Comparison with Cottrell Analysis
Thank you for the suggestion regarding the applicability of the Randles–Ševčík equation and the limitations of using a single-point approximation.

To improve clarity for novice users, we have added a contextual warning on the CV-3.1 (Randles–Ševčík) interface:

⚠️ New to Randles–Ševčík? [Click to learn more]

This links to our documentation page, which explains the applicability assumptions of the Randles–Ševčík equation—particularly under low scan rates (<10 mV/s) where planar diffusion may no longer hold:
📄 http://34.162.1.1:8080/doc#section4

In addition, we have included a note near the output highlighting that this method is based on a single data point (peak current) and may be sensitive to baseline drift, surface conditions, or iR drop. For more robust diffusion analysis, we recommend using the Cottrell-based approach in the CA module, which performs regression over thousands of data points.

While we have not yet implemented automatic i/√v overlap checking, we agree that such diagnostics would enhance usability and are considering this for future updates.

2. Deviations in Chronoamperometry (CA): Early-Time Charging & Late-Time Convection
To alert users of potential deviations in CA experiments, we’ve added a contextual warning to CA-1:

⚠️ New to Chronoamperometry? [Click to learn more]
📄 http://34.162.1.1:8080/doc#section6_1

The documentation outlines how charging currents dominate at early times and convection effects can emerge at longer durations, both of which can impact the accuracy of Cottrell-based diffusion coefficient extraction. Users are encouraged to choose regression intervals accordingly.

3. Levich Analysis Validity in HDV Module
In the HDV-1 module, we have added a similar warning:

⚠️ New to Rotating Disk Electrode? [Click to learn more]
📄 http://34.162.1.1:8080/doc#section1

This reminds users that Levich analysis is only valid in regions of stable current plateaus. If the current is still rising or decaying (i.e., not steady-state), the resulting diffusion coefficient will be unreliable.

4. Standard Rate Constant (k⁰) Calculation in CV Module
We have also added a warning to CV-4.1 (k⁰ calculation module):

⚠️ New to standard rate constant calculation? [Click to learn more]
📄 http://34.162.1.1:8080/doc#section10

The documentation clarifies that k⁰ estimation assumes certain conditions (e.g., reversible or quasi-reversible systems, well-defined peak separation). Users are advised to validate these assumptions before applying the method.
