### 2025-07-16 Woffee/Envismetrics#12 @Grant-Xue
@MoonshinetheP
We’ve made the following changes based on your suggestions:

Upload Failure Clarity – If the uploaded file does not match the expected structure or column headers, a warning message now alerts the user.

File Removal Feature – Users can now remove uploaded files directly during a session without refreshing the page.

Clearer Data Format Instructions – A new guide has been added:
📄 [README_data_structure.md](https://github.com/Woffee/Envismetrics/blob/main/data/README_data_structure.md)

Please let us know if anything remains unclear or if further improvements are needed!
