### 2025-07-16 Woffee/Envismetrics#10 @Grant-Xue
@MoonshinetheP

**About the Levich equation usage**

Thank you for raising important theoretical points regarding the application of the Levich and Koutecky–Levich (KL) equations. We fully agree that the diffusion coefficient D is a physical constant, determined by the nature of the analyte, solvent, and temperature, not by the applied potential or rotation rate.

However, in our analysis workflow, we do not intend to suggest that D physically changes. Rather, we calculate D repeatedly across different potentials using the definitional form of the Levich equation. The equation defines how current should behave under mass-transport-limited conditions if D is known, or conversely, allows one to extract D from current data if all other parameters are fixed. Because experimental current values can fluctuate (due to convection, noise, or subtle surface effects), calculating D at multiple potentials within the plateau region serves several diagnostic purposes:

- It helps users identify regions where mass-transport control holds (i.e., D is consistent across adjacent potentials);
- It allows averaging across these points to reduce statistical noise;
- It helps flag potential regions where the Levich assumptions break down (e.g., near the foot or tail of the wave).

Thus, the D vs. potential plot in Envismetrics is not an assertion of D varying with E — it is a diagnostic tool derived from repeatedly applying the definition equation to observed current responses. We will explicitly clarify this point in both the manuscript and user documentation to avoid misinterpretation.

As for the KL analysis, although it is often used to extract kinetic parameters like k0 , it also contains the Levich slope in its linearized form. Therefore, under mass-transport-limited conditions, KL regression can still be used to estimate D — not as a redundant method, but as a cross-check or when experimental conditions require separating kinetic and transport contributions.

We thank you for emphasizing theoretical rigor, and we have updated the manuscript to reflect these clarifications. In particular, we now explicitly state that:

Diffusion coefficient is extracted via equation definitions;
D vs. E plots are not implying physical dependence, but are visual aids to guide valid data selection.

Let us know if you recommend an alternative phrasing or structure to further improve clarity.
