### 2025-07-16 Woffee/Envismetrics#10 @Grant-Xue
Hi @MoonshinetheP @Woffee @Grant-Xue,

Thanks again for your detailed and constructive comments. We've addressed all points as follows:

- General Comments

1. Figures Placement: Figures are now inserted inline with the relevant module sections — HDV (TL & TR), CV (BL & BR), and CA (added new figure as requested).
2. Equations Added: Cottrell, Levich, and Koutecky–Levich equations have been included, with definitions for all variables and constants. Initially omitted for brevity, they are now restored for clarity.

- Specific Corrections (the line number is not the original line since the new added content)

1. Line 71–77: Revised the Statement of Need section to avoid redundancy and better reflect the unique purpose of Envismetrics.
2. Line 81–82: Expanded the Gaussian filtering description. It now:
- Clearly states that sigma = 0 disables filtering;
- Explains how sigma affects smoothness;
- Warns against over-smoothing which may obscure features.
3. Lines 84–87:
The entire Levich and Koutecky–Levich section has been rewritten:
- Levich is framed as a mass-transport-limited method to estimate $D$.
- KL is correctly described as used primarily for estimating the standard rate constant $k_0$, although it shares the same slope term with Levich.
- Explained that multiple $D$ estimates across potential are for diagnostic purposes only, not to imply physical dependence.
- A warning is added regarding steady-state applicability.
4. Line 88–89: Corrected the KL plot definition to be inverse current vs. angular frequency.
5. Line 89–93: Clarified that the diffusion coefficient is a constant and does not change with potential — we only visualize its behavior to help identify valid fitting regions.
6. Line 96: All uses of “rate constant value” have been standardized to “standard rate constant.”

- Kinetics & Tafel Analysis

1. Line 113–116: Expanded the section on Nicholson–Lavagnini methodology:
- Explained how $\Psi$ is extracted from $\Delta E_p$.
- Included assumptions and limitations for proper application.
2. Line 114: Symbol formatting is now correct.
3. Line 118–124: Clarified that peak detection comes from Function 2.
- Linear region fitting is not yet implemented due to variability in slope behavior, but we plan to support flexible region selection in future releases.

Looking forward to your thoughts on this section after all other changes. And thank you for the positive feedback!
