### 2025-07-16 Woffee/Envismetrics#11 @Grant-Xue
Hi @MoonshinetheP,

Thank you for your careful review and detailed comments. We have addressed all the issues you mentioned as follows:

---

###  General Comments

- **Missing citations**
  The following references were accidentally omitted during earlier revisions and have now been restored:
  - **Garg et al. (2021)** — cited at **line 62**
  - **Sanecki et al. (2003)** — cited at **line 48**
  - **Wang et al. (2020)** — cited at **line 48**

- **Inconsistent hyperlinking/citation style**
  We use the JOSS markdown citation template of `[@ref]` or `[@ref1; @ref2]`. The actual hyperlink appearance may depend on the rendering engine (e.g., GitHub preview vs. JOSS processing via Pandoc).
That said, we have carefully reviewed and updated the manuscript to ensure the citation style is now applied consistently throughout. The BibTeX citation keys (e.g., @Xue_2023) are auto-generated by reference management tools, and we’ve retained them in their default form for reproducibility. However, we’re happy to make further adjustments if JOSS requires a specific format.

---

###  Specific Corrections

- **Line 116 – Lavagnini et al. is repeated twice**
  This has been resolved by consolidating the reference. The corrected sentence now appears at **line 209**:
  > "...was extended by Lavagnini et al. to cover a broader range of peak separations ($\Delta E_p$) [@nicholson1965theory; @lavagnini2004extended]."

- **Line 121–122 – Danlei Li et al. named redundantly**
  The author name has been removed and replaced by a standalone citation:
  > "Additionally, a mass-transport corrected version has been proposed and implemented in this module [@LI2018117]." (now **line 248**)

- **Line 123 – Bacil et al. / Zanello citation mix-up**
  Corrected to cite the correct work by Bacil et al.:
  > "...including the study of dopamine oxidation at gold electrodes conducted by Bacil and co-workers [@C9CP05527D]."

- **Line 149 – 'Xue_2023'?**
  This is the correct BibTeX key for our ECS Meeting Abstract [@Xue_2023], now used at **line 316**.

- **Line 167 – Journal abbreviations inconsistency**
  We checked and found no remaining journal abbreviations in the main text. The format is now consistent throughout.

- **Line 179–180 – Journal name "Electroanalysis"**
  Confirmed and corrected in the bibliography entry (see bib line 322):
  `journal = {Electroanalysis},`

- **Line 198 – Author format for Sadik**
  This entry has been revised. It was previously using an internal working title for an unpublished manuscript.

- **Line 212–213 – RSC in author field of Zanello citation**
  The incorrect author field has been corrected to:
  ```bibtex
  author = {Zanello, P.}
  ```

---

Let us know if any other issues arise. We appreciate your thorough review and feedback to improve the clarity and consistency of this submission.
