### 2025-07-18 Woffee/Envismetrics#17 @Woffee
Hi MoonshinetheP, thanks for your comments. I have now made the following updates:

- Replaced manual text input for Peak Range (Top) and Peak Range (Bottom) with interactive range sliders.
  - Each peak now uses a dual-handle slider to define a value range.
  - Sliders support float values from -1.5 to 1.5, with a step size of 0.01.
- Grouped each peak’s settings (Top Range, Bottom Range) into a single visual block for better usability.
- Added a “Number of Peaks” input field. Changing this value dynamically adds or removes peak configuration blocks.
- Added “Scan Rate Range” sliders for each peak.
- Added a “Cycle Range” slider.
