### 2025-07-23 #8436 @Grant-Xue
@MoonshinetheP @oliverrdz @woffee
Dear @MoonshinetheP @oliverrdz @Woffee,

Thank you all for your detailed and thoughtful feedback — we’ve carefully reviewed and addressed each of your comments across the Issues and annotated PDF. We’ve made substantial updates to the manuscript, figures, and code documentation based on your suggestions.

We believe the manuscript is now significantly improved and better aligned with community expectations.

We would appreciate it if you could take a look at the current version and let us know if anything further is needed. Your feedback has been extremely helpful and we’re happy to continue refining things if necessary.

Many thanks again!
