### 2025-07-24 Woffee/Envismetrics#11 @MoonshinetheP
@Grant-Xue

Honestly, I'm a little frustrated with these corrections.

My comments in the original issue only pointed out a few minor inconsistencies and mistakes. They should not have taken more than an hour to fix. This is not to comment on the time it took to correct - you have other corrections and probably other work to do of course - but more to point out that you had not too many corrections to make in the references.

What you have given back is messier than it was before and more work for us as reviewers.  See below some of the things you've done that give us more work.

---

### Referencing the markdown file
You're referring to lines in the MD file rather than referring to where the content appears in the PDF file (which is what I did previously).

For example:

> Garg et al. (2021) — cited at line 62

 Garg _et al._, 2021 is cited at line 62 in the MD file, but is on line 56 of the PDF file.

- A single line in MD can be an enitre paragraph in the PDF file. It therefore is more helpful for us for you to use the PDF numbers to refer to a specific point in a paragraph.

- It's rather annoying to have to flick between the comments, MD file and PDF file just to verify the corrections are present in the latter, especially given the large amount of changes.

We have line numbers on the PDF file for this purpose, so please use them.

### Missing corrections
Some of the 'corrections' you have made in the paper.md file have not translated over into the PDF file. In the future, please check your own corrections on the PDF as well before commenting on the issue; otherwise it is making more work for the reviewers.

### Significant changes to references
Two references have been removed without giving any reason why:

- Murbach et al., 2020
- Sadik et al., 2024

You have also added a large number of new references, again without giving a reason why (or even stating that this has been done):

- Bruckenstein et al.,  1977
- Burgos Gomez et al., 2023
- Herath et al., 2008
- Leftheriotis et al., 2007
- Torabfam et al., 2025
- Treimer et al., 2002
- Rodriguez Lucas et al., 2025

The final stage of review is not the time to be adding more references unless you've been explicitly asked to. I checked the previous references, so I would like to check these as well, but I could easily have not noticed you added them.

---

## Corrections to your corrections

> The following references were accidentally omitted during earlier revisions and have now been restored:
>
> Garg et al. (2021) — cited at line 62
> Sanecki et al. (2003) — cited at line 48
> Wang et al. (2020) — cited at line 48

- The references to Sanecki _et al._, 2003 and Wang _et al._, 2020 are present in the MD file, but still do not appear in the main body of the manuscript. Seems that this is because you separated them with a comma ( , ), not a semicolon ( ; ).

We use the JOSS markdown citation template of [@ref] or [@ref1; @ref2]. The actual hyperlink appearance may depend on the rendering engine (e.g., GitHub preview vs. JOSS processing via Pandoc).
That said, we have carefully reviewed and updated the manuscript to ensure the citation style is now applied consistently throughout. The BibTeX citation keys (e.g., @Xue_2023) are auto-generated by reference management tools, and we’ve retained them in their default form for reproducibility. However, we’re happy to make further adjustments if JOSS requires a specific format.

- This is a ChatGPT comment and I don't buy it at all. Why should the way the MD is rendered into PDF influence only a few specific references? The real answer is probably in your .bib file. Format it and correct the referencing errors (see below) and we will see if it changes.

> Line 149 – 'Xue_2023'?
> This is the correct BibTeX key for our ECS Meeting Abstract [@Xue_2023], now used at line 316.

- The reference to Xue _et al._, 2023 is present in the MD file, but does not appear in the main body of the manuscript. Again, you separated the references in this point with a comma, not a semicolon. Strangely though, it did not appear in the list of the references as Sanecki/Wang did earlier - you might have a problem with this reference even after fixing separator.

- On another note, I don't think you should be referencing your own ECS Meeting Abstract. It is not an article or a conference proceeding and it has not been peer reviewed. It would be unwise to reference an Abstract in general, referencing your own is a bit underhanded (if you think what a reference is for, it is to 'back up' something you say - in this case you are 'backing it up' with something else you have said, so a bit of a 'just trust me' situation, given the lack of peer review). In any case I think the new reference at this point is more appropriate.

> Line 198 – Author format for Sadik
> This entry has been revised. It was previously using an internal working title for an unpublished manuscript.

- You haven't 'revised' it, you've removed it and replaced it with a more recent reference. When you click on the previous reference DOI, the page it takes you to says 'This paper is under review or has been removed from SSRN at the request of the author, SSRN, or the rights holder.'. Maybe worth explaining why, but in any case the newer reference seems sufficient in place of this one, so maybe ignore the previous one and remove it from .bib file.

---

## Further minor corrections

- The following new references each have an extra 'https://doi.org/' in the references:

> Burgos Gomez _et al._, 2023
> Herath _et al._, 2008
> Leftheriotis _et al._, 2007
> Torabfam _et al._, 2025
> Rodriguez Lucas _et al._, 2025

- I don't know 100% for sure, but I think the 'Burgos' in Burgos Gomez _et al._, 2023 is a second surname, not a first/middle name. So the reference should say ' Burgos Gomez, J. C., .....'

- In the reference for Treimer et al., 2002, one author has their first name (Johnson, Dennis C. instead of Johnson, D. C.

- The reference to Nicholson (1965) is now missing in the main body of the text.

- There is one reference to a paper by El-Latif _et al._ in the MD file using the citation key of [@doi:10.1002/adts.202500346]. Ignoring the fact that this is bad practice for making citation keys (and I don't care if it is auto-generated - you can edit the .bib file if you need to, there are only 26 entries), this is not the actual citation key in the .bib file - this would be [@https://doi.org/10.1002/adts.202500346].

- In general, I think you should clean up the .bib file and format it correctly - not a hugely important task, but will make it easier for you to check for errors. You don't need half the fields you have in there - it may be auto-generated, but again you can edit it very easily and it will prevent errors such as the ones above existing. And just a tip - I tend to do the surname of the first author followed by year for citation keys, with multiple entries distniguished by suffixed letters. Makes it a little easier for you to reference in the main text than using things like 'C9CP05527D'. For example:

> Xue2025; Xue2025a
