### 2025-07-24 Woffee/Envismetrics#14 @MoonshinetheP
@Grant-Xue The docstrings are much better. The inline comments are still sparse and are more often explaining **what** you did rather than **why you did it that way** (which is usually the point of such comments). Still keep the ones you have explaining different variables etc. - these are good.

I will get to the content of the code shortly, although it is a little confusing - the .diff shows that there are some functional changes to the '_commented.py' variants, not only just addition of comments. When I asked for comments on the code, I meant for you to add them to the code being used, not to create a separate .py file with comments that is never used. Also strange to add the new functionalities to this code.

In general, there should only be one file for each of the three 'modules' - there is another CA file (CA_1.py) which also has some notable .diff changes. I don't see any reason for why it was uploaded, as the commit comment just says 'uploaded files'.

Also, I would like you to make the same comments on BaseModule.py. It's not a complex file, but it is important for people to understand why you have done things in the way you have - after all, each Module uses the BaseModule() parent class.

@Woffee I assume it is you writing app.py - I'm not too familiar with Flask, although I can just about follow. You may want to make similar comments detailing explicitly what page corresponds to each app route (if I understood correctly that this is what these are doing), because it is not entirely clear how the workflow is structured and how it corresponds to the web application (at least not to me). I will leave it up to you though, I am not sure of best practice in this regard (I think @oliverrdz has a lot of experience here)
