### 2025-07-24 Woffee/Envismetrics#17 @MoonshinetheP
@Woffee

I really like the changes you've made. A few suggestions:

- The potential range slider seems fixed between -1.5 to 1.5 V. Would be more intuitive to have the min and max in the actual potential range of the voltammogram - would also make the app more foolproof, as if you currently select the potential ranges that don't exist in the data, it of course returns an error. I think you should be able to extract max and min from data upload stage, then set slider resolution as potential step size (or literally as potential[slider_index], since even if step is 1 mV, likely it is not always exactly that - still display text as 1 mV increments of course).

- Maybe a bug, but if you select 1 or 2 or 4+ peaks, then go back a step, the 'number of peaks' combo box is still at the number of peaks you previously chose, but there are always three peak 'blocks'. The ranges/sliders also reset, but I don't know if this is something you can fix on flask. In any case, I would set the default slider ranges as their full ranges (whole potential range, all scan rates, all number of scans)

- More complicated, but if you could update the matplotlib figure as you move the slider, so that a box highlighting the peak region is shown - well, it would be really cool, but more difficult so I understand if you avoid this for now

- Scan rate range is a nice touch, but it might also be neccesary to omit a scan rate in the middle of the range (maybe it is an outlier). I don't know how you could do it, maybe just as checkboxes if there are not too many scan rates, maybe as text entry 'Avoid : 100' otherwise.

- Similar comment with the scan range, although I don't think it is as important. I do notice it is between 0-100 at the moment though, which seems excessive. There should be a way of extracting the number of scans

- I don't really know what the 'Scan rate to display (mV/s)' and 'Cycle number to display' fields are for now. If I selected the Cycle numbers in previous slider, why would I then select anything different to display? I will maybe focus more on this when addressing the modules, but it seems if you select one specific CV (50 mV/s, cycle 1) it plots all peaks it detected in each scan rate on this one (I think, although something is off). I also see some other functionality problems in the peak detection
