### 2025-07-24 Woffee/Envismetrics#17 @MoonshinetheP
Also, do you think you could implement something similar with 'Regression Time Range' in the CA module (using actual times, not 0 and 1) and 'Applied potential range (V)' in the HDV module?

Doesn't seem like it would be too difficult given what you accomplished on the CV module.

And finally, make sure the subsequent functions that use peak detection take the number of peaks into account (Function 4: Rate constant calculation assumes 3 peaks, others assume 1 peak).
