### 2025-07-24 Woffee/Envismetrics#18 @MoonshinetheP
@Grant-Xue

> We’d like to clarify that the apparent concatenation of current transients in the 'Potential vs. Time' and 'Current vs. Time' plots under CA-2 is due to the structure of the input dataset, rather than the software explicitly stitching data together.
>
> For instance, in the test folder 05132024_CA_D, there are 49 .xlsx files. These originate from 7 separate experiments, each consisting of 7 transients labeled with the same leading number (e.g., 1_..., 2_..., ..., 7_...). These files are organized sequentially in terms of time — for example:
>
> 2_DMAB_2s_CA.xlsx: first timestamp ~120.77 s
>
> 3_DMAB_2s_CA.xlsx: first timestamp ~123.77 s
>
> This creates the illusion of a continuous time series, but no concatenation is being performed by the software. The files are simply being plotted in the order they appear, with raw time values intact.

- Yes, I understand the workstation is usually adding any pretreatment, conditioning and equilibration times. That is why they also give you the 'Corrected time / s' column, which you could use instead to make sure each transient starts at t = 0. Even if they didn't you could subtract the first time from the the rest of the 'Time / s' column. For Cottrell, you need time starting from 0 (you even take the step in CA.py to normalise to 0, which would be unneccesary if you used corrected time).

- I don't really understand what is going on with the data in your test folders - it would make sense if you had several transients (as you say with duration indicated in the filename) one after another, and the time reflecting that, but in this dataset: '1_...' all seem to be 120 s regardless of duration in filename; '2_...' all seem to start at 121.76 s; '3_...' seem to start at 123.77 s + all previous durations.

> To avoid this stacking behavior (and misinterpretation as multi-step chronoamperometry), we recommend that users upload only one transient per index (1–7) per analysis session. Duplicate indices (e.g., multiple files starting with 2_...) may result in overlapping curves in CA-1/CA-2.
>
> We’ll consider adding a warning or data validator to flag duplicated prefixes in future updates.

- The problem here (and it is a reasonably consistent problem throughout the whole software) is that you are considering only how your data is structured, despite the fact that you are publishing this software in JOSS so that other people can use the work. You assume people are going to want to use the software for the exact same purpose you have used the code - in reality, you need to account for people using it in the way they want (they may not care about 'duplicate prefixes', and they may indeed want to see two consecutive transients). This is a pretty fundamental element to developing software for people to use in the laboratory - make it flexible.

> You're absolutely right — in the current implementation, key parameters such as the number of electrons (n) and concentration (c) are treated as global values, applied uniformly across all uploaded transients. This works for datasets with identical experimental conditions, but limits the utility when comparing multiple transients with varying concentrations or reaction stoichiometry.
>
> We are now extending support for per-transient parameter input via an "Advanced Mode." In this mode, users will be able to input a list of values (e.g., c = [1e-3, 5e-4, 2e-3, ...]), where the list length matches the number of uploaded files. Each entry in the list will be mapped to the corresponding transient.
>
> For now, this will be implemented via JSON-style list input in the UI, but in future updates, we plan to improve usability further by introducing a user-friendly parameter input table (e.g., spreadsheet-style or uploadable CSV).

- A long way to say we will address this in the future. Make sure it is adequately explained in the development roadmap, then.

> Thanks for raising this — the skipping of the first interval is intentional by design.
>
> In typical chronoamperometry experiments, the first potential step (often ~120 s in our data) is used to precondition the electrode, for example:
> to remove residual adsorbates or surface films,
> to stabilize the diffusion environment near the electrode,
> and to ensure consistent initial concentration profiles before analysis.

- No, this is something you are doing in this specific example. Again, if you want to publish the code you use as a software for others to use, it needs to be adaptable to other people's needs. Furthermore, this approach also assumes that people are saving their pretreatment steps as separate files, then uploading them so they can not be shown (and also shows that your approach cannot deal with double-step chronoamperommetry, which is more typical for D estimation than making several transients with different durations one after the other). In general, uploading something means I want to see it in full, you don't save data just to not look at it.

> This initialization period is not included in the regression-based diffusion analysis because its current transient may not reflect pure diffusion behavior, and including it could lead to misleading Cottrell fits.

- I may not want to use your software just for Cottrell fits - after all, isn't the point that you can use it to visualise/process data in an automated way?

> Therefore, we intentionally exclude the first interval and analyze from the second step onward to ensure the calculated diffusion coefficients are meaningful.
>
> We agree that this behavior should be better documented or even made optional. We'll update the code comments and consider adding a user toggle/flag to include all intervals if desired.

- It doesn't need to be better documented, it needs to do what the user wants it to do - plot a transient. The user can decide themselves if the derived data is meaningful. You cannot make that call for them ahead of time unless you become very specific with data upload and labelling, which is a nightmare. If you want an easy and useable software, then just make it plot transients with an optional second step of 'do you want to fit to Cottrell?'.

> You're absolutely right — in traditional Cottrell analysis, the current is plotted against 1/√t, and the diffusion coefficient D is derived from the slope. In Envismetrics, we define: B(t) and plot I vs. B(t) directly.
>
> This allows the slope of the linear fit to directly equal √D, so that:

> This was intended to simplify the user's workflow — no additional back-calculation is needed. This makes the output more immediately actionable, especially for users with limited background in electrochemistry.

- The user doesn't need to do anything, it's all in the calculations within your CA.py files. You have all the relevant constants already when you calculate `Bt = ((n * F * A * C0) / (math.pi ** 0.5)) * t_inverse_05`, you just need to use them to calculate D from the gradient of i vs 1/SQRT(t). A more 'immediately actionable' approach would be to show D, not gradient that equals SQRT(D).

> What we will do:
> We agree this non-standard axis might be unintuitive for some users. Therefore:
> We will add a note in the documentation to clearly explain that our Bt axis is equivalent to 1/√t but scaled by known constants.
> Optionally, we may add a toggle to allow plotting I vs. 1/√t in a future update for better educational value and physical intuition.

- It is arguably more confusing to plot this scaled gradient than it is to just give the gradient and let people figure out how to extract D. Furthermore, if you ever do intend to plot different C, n, etc., this approach is going to cause a mess. If you want Cottrell, just plot Cottrell.

> On Plot Duplication:
> In many practical cases, users analyze chronoamperometry involving multiple potential steps (e.g., 7 platforms). Since Envismetrics already generates 7 individual regression plots (one for each step), duplicating these plots in a second format (e.g., I vs 1/√t) would significantly increase visual clutter without adding real value.
> Our goal is to keep the interface clear, focused, and actionable, especially when processing a large number of transients. Instead of duplicating, we opt to explain the axis meaning clearly and offer the possibility of an optional toggle in the future.

- I'm not sure I particularly agree that 49 plots is keeping the interface clear. Perhaps choosing which ones to group would help? At the end of the day, if I upload more than one transient, it may be because I want to compare the Cottrell fits. It may also just be for automated plotting - it would be better if you had some options.
