### 2025-07-24 Woffee/Envismetrics#13 @MoonshinetheP
In general, good. Will close, but address below comments:

> This links to our documentation page, which explains the applicability assumptions of the Randles–Ševčík equation—particularly under low scan rates (<10 mV/s) where planar diffusion may no longer hold:

- ChatGPT error here. I said you were using scan rates as low as 10 mV/s, not that this is the limit for RS analysis. If you're working with electrodes 1 - 3 mm (most commonly people are not using electrodes larger than this - it is electrode size dependent, so be careful here...), you'll likely transition away from 'purely planar' diffusion at scan rates around 100-200 mV/s.

> In addition, we have included a note near the output highlighting that this method is based on a single data point (peak current) and may be sensitive to baseline drift, surface conditions, or iR drop. For more robust diffusion analysis, we recommend using the Cottrell-based approach in the CA module, which performs regression over thousands of data points.

- Okay good, although the things you've mentioned ('baseline drift', 'surface conditions' and iR drop) are also going to influence data obtained in current transients as well. My point was more that CA is superior to CV because of the larger number of points - less susceptible to variations due to surface roughness at high scan rates and other stuff like that.
