### 2025-07-24 Woffee/Envismetrics#12 @MoonshinetheP
@Woffee @Grant-Xue

Good on ability to remove file and error on upload fail. Comments below:

> We understand your concern. In our typical workflow, scan rate or RPM is not included inside the raw data files exported from electrochemical software (e.g., NOVA, EC-Lab, or Gamry).
Therefore, users often include these values in the filename as a manual record to avoid ambiguity during analysis. While this is not ideal, it remains the most consistent automatic identifier at present.
We are exploring options to allow manual entry of scan rate/RPM in future updates.

- I'm not sure any software exports such information in the data file (maybe in some method file?) - you could work out scan rate from potential step / time step, but this is besides the point - rotation rate I would assume is usually controlled external from electrochemical software/hardware for most people (at least every RDE I've used has been like this).

- My point is more that your file format is very limiting - would be better to have an entry box next to filename allowing you to enter the scan rate/rpm so that I don't need to rename all my files to comply with your format. It may not even be necessary, you could adapt `    pattern = r'(?:^|_)(\d+rpm)\.'    match = re.search(pattern, filename)` to extract the values you need as long as they are anywhere in the filename. The conditions have got even more restrictive, now CA need to start with a number - and everything has to have the module name at the end of the filename (not sure why this is necessary - it seems to work regardless and if I am uploading to CA module, I should know it needs to be CA data).

> This is an excellent suggestion. Building a universal interactive file parser would indeed improve flexibility. However, such a feature requires significant development effort and interface design, which are currently beyond our team's available resources. For now, we ensure compatibility by manually supporting and documenting the output formats of major software platforms.
We agree this is a limitation compared to tools like Origin, and we’ve added this feature to our long-term development roadmap.

- Then you need to make sure all references to your software being useable for different workstations are removed from manuscript, app and MD files. It seems you can only use it on Autolab data.

> We have added a dedicated [README_data_structure.md](https://github.com/Woffee/Envismetrics/blob/main/data/README_data_structure.md) file to the repository.
> This document includes:
>
> Expected file formats and required column headers for each module;
> Example input tables;
> Notes on common formatting issues and how to resolve them.
>

- Good, but you don't mention anything about .txt files in this MD file and they should be equally valid in CA.
