### 2025-07-24 Woffee/Envismetrics#5 @MoonshinetheP
- I get a slightly different Figure 2b near the negative vertex potential, but Fig 2a is fine.

- I cannot get to the same Fig. 2c by following your instructions. The instructions are also outdated.

- You also didn't address the comments here:

> For the RS plot, I obviously don't get any of the simulated data you seem to have plotted in Fig. 2d, although I do get RS analysis for reduction peaks, which is not in the manuscript.

- In this MD file you have added it says:

> Subplot 4 (Figure 2d) includes a new feature in the core code — simulated peak currents based on a given diffusion coefficient (shown as triangles ▲ in the plot).
⚠️ Note: This simulation functionality is not yet available in the online tool. To reproduce this feature, use the core Python script in the source code instead. The online version currently only supports diffusion coefficient fitting from experimental data.

What exactly do you mean by this? From what I can tell, you've:

1. Performed a linear fit to the i vs SQRT(v) data
2. Extracted the slope and intercept from this data
3. Calculated the y data of this fit using the slope and intercept with artificial scan rate data
4. Used this y data to calculate D, thus getting back to where you were in step 2

That isn't simulation. Unless I missed something, you just plotted the same data derived from the experiments.
