### 2025-07-24 Woffee/Envismetrics#7 @MoonshinetheP
I think I said in another issue about using the term 'potentiostat'. In general, if I made a comment in one issue, it applies everywhere (and I think this is an issue we are having in this review), but okay. Abstract is fine now, up to you whether it stays as Abstract or goes in summary like other JOSS papers.
