### 2025-07-24 Woffee/Envismetrics#8 @MoonshinetheP
Perfect except small correction:

> (LSV at RDE, under laminar flow and planar diffusion conditions)

- I would just saw laminar flow - convection dominates
