### 2025-07-24 Woffee/Envismetrics#9 @MoonshinetheP
> Refocused comparison scope: We shifted the comparison from instrument-specific software (e.g., NOVA) to widely used data analysis tools such as Excel, Origin, and SigmaPlot, in line with your suggestion.

- You mentioned the real comparative softwares (Excel, Origin, SigmaPlot) in one paragraph and still kept the workstation software as the main comparison. Not really a change, and still focussing on the wrong thing.

- Also, it is worth pointing out again here that you are only able to process Autolab data (it says something about plaintext from any system - this just isn't true). So again, your comparing that your software is better than Autolab, despite no-one using NOVA for plotting and analysis, and you STILL needing it to get your data.

> Clarified data acquisition role: We now explicitly state that Envismetrics does not perform data collection, and users must first export data from proprietary software (e.g., Autolab) into plaintext formats such as .csv, .xlsx, or .txt.

- Yes, but you also have a line

> Envismetrics works independently of instrument brands - allowing researchers to export plaintext data from proprietary systems and continue their analysis seamlessly.

Which implies Envismetrics is doing something to aid operation of the proprietary systems

> Reformatted for clarity and consistency, with better alignment across rows and tools.

- Better, although it would be nice in some places to have more space between rows. Just a suggestion.

- I'm still not very keen on some of the points you raise in the table. Why bring up Hardware in a software comparison? Why talk about data collection when you are comparing to your software that doesn't collect data? Why specify both installation and platform support, as if you can install on a platform that isn't supported? To me, the entire table is too long, focussing on lots of different things except what makes your software more useable/useful.
