### 2025-07-25 Woffee/Envismetrics#10 @MoonshinetheP
- In line 83 you still talk about supporting data from many potentiostat (again, terminology from other issue - please make sure corrections are made everywhere, not just where the point is made) software. This is not true (also Metrohm is a company, not software). I have not said it so far, but its okay if your tool is currently only viable for Autolab data - they are probably the most popular workstations - but then you need to be more specific.

- Part about Levich and KL is much better, although still problems with Levich (see below).

- You say the CV module supports both irreversible an reversible RS, but where is this in web app and source code?

- I am not really sure if Function 4 is really worth adding in the manuscript/code. Its certainly not a standard method that is used often by the novice audience and seemingly doesn't work so well (I got something like k0 = 5E-04 for K3Fe(CN)6 data in '05172023CV'). I also will need to read through Nicholson again, but I don't understand how Phi is dimensionless if it includes step size (through X), so it is possible there is something missing here.

- This new Klinger-Kochi approach needs a reference

### In response to comments
> However, in our analysis workflow, we do not intend to suggest that D physically changes. Rather, we calculate D repeatedly across different potentials using the definitional form of the Levich equation.

- The point is not that you are suggesting D physically changes with potential, because as you pointed out, this is nonsense. The point is that you are able to apply an equation which should be specific to the plateau to elsewhere within the voltammogram. This is fundamentally incorrect. The plot in Fig 2b is fundamentally incorrect. Levich is not just a current = diffusion coefficient calculator, it is a steady state relationship. To have SS in RDE, you need a plateau. If you are at a plateau, the current should be stable regardless of potential.

> Because experimental current values can fluctuate (due to convection, noise, or subtle surface effects), calculating D at multiple potentials within the plateau region serves several diagnostic purposes:

- You should hope the current 'fluctuates' due to convection in RDE, although if you are talking about natural convection, the forced convection might be a little bit stronger in most cases. I'm not even sure what is meant by subtle surface effects.

- In general, 'fluctuating' current in a plateau can be addressed by numerous approaches. If you took the maximum current within the potential range and calculated D from that, it would still be more reasonable approach. In the past when I needed to calculate limiting current, I averaged the current data over a given potential window (which of course the user could control and select to be elsewhere than the plateau if they didn't know what they were doing).

> It helps users identify regions where mass-transport control holds (i.e., D is consistent across adjacent potentials);
> It allows averaging across these points to reduce statistical noise;
> It helps flag potential regions where the Levich assumptions break down (e.g., near the foot or tail of the wave).

- I don't see how the D vs E plot could reveal a plateau any more effectively than the i vs E plot if you are just converting i to D with scalar constants.

- Its not so much averaging as you've just skipped every 'however many mV' - even worse then, the 'noise' is not averaged out, but you now have less data to select the plateau from

> Thus, the D vs. potential plot in Envismetrics is not an assertion of D varying with E — it is a diagnostic tool derived from repeatedly applying the definition equation to observed current responses. We will explicitly clarify this point in both the manuscript and user documentation to avoid misinterpretation.

- I appreciate the efforts to clarify any misinterpretation, but the approach is incorrect and misleading to the intended audience. There would be no misinterpretation if you used Levich properly. Rather than going out of the way to justify the non-standard plot, please just get rid of it and only plot **limiting** current vs. rotation rate (i.e. Levich)!

> As for the KL analysis, although it is often used to extract kinetic parameters like k0 , it also contains the Levich slope in its linearized form. Therefore, under mass-transport-limited conditions, KL regression can still be used to estimate D — not as a redundant method, but as a cross-check or when experimental conditions require separating kinetic and transport contributions.

- You've misunderstood the point. KL at steady state potentials is just Levich (no kinetic control at high enough potentials). It is not cross-checking anything, its doing exactly the same thing as you did in Levich (hence the redundancy). If you don't believe me, I leave the following quote from a Pine Research AN:

> For the set of currents sampled on the limiting current plateau (see Figure 2, Red circles), an extrapolation back to the vertical axis (i.e., to infinite rotation rate) yields a zero intercept.  This is the identical result obtained for a facile half-reaction (see Figure 1, Right) because these currents are sampled at a high enough overpotential that there are no kinetic limitations.  Only mass transport limits the current, and the usual Levich behavior applies.

- No-one is plotting KL for the primary objective of determining D, maybe they extract D from KL at steady state potential, but the whole point of KL is to work out kinetic parameters k0 and a. The fact your software does not do this needs to be addressed.

Figures Placement: Figures are now inserted inline with the relevant module sections — HDV (TL & TR), CV (BL & BR), and CA (added new figure as requested).

- I see them in the MD file, but they didn't get put into the PDF file. Please check and make sure the Figures are visible.
- In general my comments about the Figures stand. Figs. 2a and 2b are incorrect. Other figures are fine except for previous comments in other issues.

> Line 81–82: Expanded the Gaussian filtering description. It now:
> Clearly states that sigma = 0 disables filtering;
> Explains how sigma affects smoothness;
> Warns against over-smoothing which may obscure features.

- I missed it before, but you have a Filtering section for CA module, which doesn't use filtering. I agree with other reviewer that this repetition should be place in its own part rather than repeated twice.

> Line 118–124: Clarified that peak detection comes from Function 2.
> Linear region fitting is not yet implemented due to variability in slope behavior, but we plan to support flexible region selection in future releases.

- What is the Elbow/Knee method for then if not determining onset of peaks?
- How are you doing Tafel analysis then if you are only identifying the peak and not the slope?
