### 2025-07-25 #8436 @MoonshinetheP
Had a fussy newborn last night, so found the time to finish the review a little earlier. I could make a lot more detailed comments/issues, but given the number of corrections outstanding, it doesn't make much sense right now. I will be pretty fully occupied from now up until the beginning of September, but I can always come back for a quick response or check.

I also thought it would be helpful for the authors if I listed the checklist items I haven't ticked off and why.

### General checks - Reproducibility

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

This is a minor point, but of the three figures showing data within the manuscript, I can only reproduce one (HDV). Authors have responded well to writing instructions for reproducing data, but those instructions for CV are still giving different results to those shown in the manuscript, and the instructions for new CA figures are not present

### Functionality - Functionality

> Functionality: Have the functional claims of the software been confirmed?

I will leave this open until the web app / source code is performing operations in the correct manner (no errors or mistakes) and with satisfactory abilities according to what the documentation/manuscript currently claims. This ties in with the quality of the source code - currently it is rather messy and executing bloated functions in a non-efficient manner, which ultimately is the limiting factor to functionality of the web app.

### Documentation - Statement of need

A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Again, a minor point, but the documentation is outdated given the changes that have been made to the manuscript. In general, this is a problem throughout the manuscript/repository - updates according to a comment or suggestion are only made where this is brought up, but the old content is still present elsewhere within the project.

### Documentation - Functionality documentation

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Current functionality is documented pretty well outside of the manuscript (in Roadmap section - could do with rename), but there has been talk of a Future development roadmap which still is not documented

### Software paper - Statement of need

> A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

Left open because, as I've commented before, this part of the manuscript still focusses on comparison with proprietary electrochemistry software when the real comparison is with plotting software.

### Software paper - References

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

A few minor errors that still need correction.

### Some additional comments

- When it comes to corrections according to comments/issues, some are omitted or only partially completed - please address everything that has been brought up and double check that any corrections are working/present.

- To some degree, the responses to the comments/issues show a lack of understanding of what the original comment/issue reffered to,and, in some cases, core electrochemical concepts. When combined with the use of LLMs has led to some rather long winded responses that kind of miss the point - please double check the responses before sending and edit them down as much as possible (I don't mind even dropping the formalities when it comes to responses)

- Source code is really messy at the moment. Lots of poor practices; functions in parent class overridden by those in child; unused variables, functions and modules; commented out code snippets; global functions used within classes rather than local definitiions; bad naming of variables; very long functions that can be done with already exisiting built in or numpy functions; A LOT of repitition. Currently the source code is far below the standards needed for community contribution and not very Zen Python.

- In general, please just do a manual read through of the documentation and manuscript. There are many small errors and typos which are difficult for us to point out. For example, your email is listed as email@example.com at the end of the MD - easily caught in a basic proofread.
