+++ data/reproducibility_guide.md
@@ -1,4 +1,4 @@
-# Reproducing Manuscript Figure 2
+# Reproducing Manuscript Figures
 
 This guide explains how to reproduce all four subplots in **Figure 2** of the *Envismetrics* manuscript using the online tool(http://34.162.1.1:8080/).
 
@@ -6,12 +6,12 @@ All required input files are included in this GitHub folder:
 🔗 [`/data/test_data/`](https://github.com/Woffee/Envismetrics/tree/main/data/test_data)
 
 ---
-## 🧪 Subplots 1 & 2: Hydrodynamic Voltammetry (HDV Module)
+## Figure 2a and 2b: Hydrodynamic Voltammetry (HDV Module)
 
-**Folder used:**  
+**Dataset location:**  
 [`data/test_data/05202024_HDV_D40_A1`](https://github.com/Woffee/Envismetrics/tree/main/data/test_data/05202024_HDV_D40_A1)
 
-**Files:**  
+**Files used:**
 - `HDV_G_DMAB0.05gL_10mVs_200rpm.xlsx`  
 - `HDV_G_DMAB0.05gL_10mVs_400rpm.xlsx`  
 - `HDV_G_DMAB0.05gL_10mVs_600rpm.xlsx`  
@@ -23,43 +23,45 @@ All required input files are included in this GitHub folder:
 - `HDV_G_DMAB0.05gL_10mVs_1800rpm.xlsx`  
 - `HDV_G_DMAB0.05gL_10mVs_2000rpm.xlsx`
 
-Each file contains a linear sweep voltammetry (LSV) curve measured under a specific rotation speed. These are ideal for **Levich plot** and **diffusion coefficient analysis**.
+Each file contains a linear sweep voltammetry (LSV) curve recorded at a specific electrode rotation speed. These data are used to generate the **Levich plot** and compute the **diffusion coefficient** for DMAB.
 
 ---
 
-### 🔧 Steps to Reproduce
+### Reproduction Steps
 
-1. Go to the [Envismetrics HDV module](http://34.162.1.1:8080/hyd_elec) to open the **HDV-1** page  
-2. Upload **all 10 files** listed above  
-3. Set **Smoothed / Smoothing Level (sigma):** `10`  
-4. Click **Submit** to proceed to the **HDV-2** page  
-5. Use the following parameter settings:
+1. Visit the [Envismetrics HDV module](http://34.162.1.1:8080/hyd_elec)  
+2. Upload **all 10 HDV files** listed above  
+3. Set the **Smoothing Level (σ):** `10`  
+4. Click **Submit** to proceed to **HDV-2**  
+5. Enter the following parameters:
 
-| Parameter                           | Default Value                     | Explanation                                                                              |
-|-------------------------------------|-----------------------------------|------------------------------------------------------------------------------------------|
-| **Concentration of solute (C)**     | `0.000000894454 mol/cm³`          | Bulk concentration of DMAB in the electrolyte                                            |
-| **Electrode surface area (A)**      | `0.15 cm²`                         | Area of the working electrode                                                            |
-| **Kinematic viscosity (ν)**         | `0.01 cm²/s`                       | Default value for aqueous media at ~25°C                                                 |
-| **Number of electron transfer (n)** | `1`                                | Typical for DMAB oxidation                                                               |
-| **Method**                          | `Levich plot and Levich analysis` | Performs Levich linear fit and diffusion coefficient extraction                          |
-| **Applied potential range (V)**     | `(-1, 1)`                          | Voltage range to extract limiting current                                                |
-| **Number of potentials to display** | `9`                                | Determines the number of lines in Subplot 1                                              |
-| **Potential step interval (mV)**    | `50`                               | Determines dot density in Subplot 2                                                      |
+| Parameter                           | Value                             | Description                                                                 |
[…]
