### 2025-08-07 a533ecf
Update paper.md
