+++ data/reproducibility_guide.md
@@ -64,31 +64,34 @@ Each file contains a linear sweep voltammetry (LSV) curve recorded at a specific
 
 
 ---
-## 🧪 Subplots 3 and 4: Cyclic Voltammetry (CV Module)
+## 2. Figure 3a and 3b: Cyclic Voltammetry (CV Module)
 
 **Folder used:**  
-[data/test_data/06102024_CV_D_AF](https://github.com/Woffee/Envismetrics/tree/main/data/test_data/06102024_CV_D_AF)
+[`data/test_data/01112023_CV_DMAB`](https://github.com/Woffee/Envismetrics/tree/main/data/test_data/01112023_CV_DMAB)
 
-**Files:**  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_10mVs_CV.xlsx  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_20mVs_CV.xlsx  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_30mVs_CV.xlsx  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_40mVs_CV.xlsx  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_50mVs_CV.xlsx  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_60mVs_CV.xlsx  
-- GP_0.05gLDMAB_40gLKOH_-1.1to0.7_70mVs_CV.xlsx
+**Files:**
+- `G_P_KOH40.0032gL_DMAB0.0475gL_100mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_10mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_20mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_30mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_40mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_50mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_60mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_70mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_80mVs_CV.xlsx`
+- `G_P_KOH40.0032gL_DMAB0.0475gL_90mVs_CV.xlsx`
 
 Each file contains a cyclic voltammetry (CV) curve of DMAB oxidation recorded at a different scan rate. These are required for **Randles–Ševčík analysis** and **standard rate constant estimation**.
 
 ---
 
-### 🔧 Steps to Reproduce
+### Steps to Reproduce
 
 1. Go to the [CV module](http://34.162.1.1:8080/cv) to open the **CV-1** page  
-2. Upload **all 7 files** listed above  
+2. Upload **all 10 files** listed above  
 3. Keep default settings:  
    - **Gaussian filter sigma**: `10`  
-   - **Cycle of representative**: `6` (can be any number base on experiment setting in here is 1-12)  
+   - **Cycle of representative**: `6` (for this dataset, any cycle between 1–12 is acceptable)  
 4. Click **Submit** to proceed to the **CV-2.1** page  
 5. In **Function 2: Peak searching**, use the following settings:
 
@@ -102,9 +105,10 @@ Each file contains a cyclic voltammetry (CV) curve of DMAB oxidation recorded at
 | **Scan rate to display**    | `20` mV/s                                           | Highlights the curve at 20 mV/s in the display                               |
[…]
