### 2025-08-07 Woffee/Envismetrics#5 @Grant-Xue
Thank you for your careful reading and helpful feedback.

Figures were recently updated based on another reviewer’s comment, but we did not update the corresponding reproduction instructions in time. We are currently in the process of syncing all instruction files with the latest manuscript version.

We have now updated the reproducibility guide:
🔗 [Reproducibility Guide – Current Version](https://github.com/Woffee/Envismetrics/blob/main/data/reproducibility_guide.md)
Following this guide, we can reproduce all manuscript figures accurately, except for some label writing mismatches, which will be corrected shortly.

---

### Regarding the "simulated data" in Figure 2d

> The Randles–Ševčík equation is inherently flexible —
> if you know **D**, you can simulate **Ip**; if you know **Ip**, you can calculate **D**.
> In Figure 2d, we used the **D value calculated from RS analysis** simply as an input to **simulate expected peak currents**.
> However, this D value could have come from **any source** — literature, other experiments, or estimation.
> The purpose of Figure 2d is not to claim a new D value, but to illustrate a general use-case:
> using RS theory to validate or visualize electrochemical mechanisms under different assumptions,
> as commonly done in literature (e.g., [10.1039/C9CP05527D](https://doi.org/10.1039/C9CP05527D)) figure 2.
> In our case, Figure 2d is not intended as a fitted result but a **demonstration of this same conceptual idea** — that simulation lines can be plotted from theoretical models for mechanism testing, especially when some parameters (like n, n′) are uncertain.

At this stage:
- The **online tool only supports RS fitting** to experimental data
- The **simulation mode** is part of the local Python version under development
- Figure 2d serves as a **proof-of-concept** for this extended functionality

We revised the manuscript and documentation to clarify this purpose and avoid confusion.

---

### Final Note

If preferred, we are happy to replace **Figure 2d** with output strictly from the current online version of Envismetrics. However, we believe the conceptual overlay in Figure 2d provides scientific value by showing the extensibility of the tool — not just as a fitting engine, but also as a possible mechanism evaluation framework.

We appreciate your thoughtful review and are actively refining both the code and documentation to better reflect these distinctions.
