### 2025-08-08 59e09c6
Delete src/CA_1.py
