+++ src/CA_1.py
@@ -1,267 +1,0 @@
-# ================================================================
-# ✨ Updates in step2() – 2025-07 by Huize
-#
-# 1. ✅ Added visual distinction for regression vs. non-regression points:
-#    - Points used in linear regression (Bt vs I) are shown in blue.
-#    - Points excluded from the fit are shown in gray.
-#    - Red line shows the linear regression result.
-#
-# 2. ✅ Enabled Advanced Mode support for per-transient parameter input:
-#    - Now supports 'n', 'a', 'c' as either:
-#        • single float (applied to all transients), or
-#        • list of values (e.g., c = [1e-3, 5e-4, 2e-3]) matching the number of uploaded files.
-#    - Also supports JSON-style string input: e.g., "[0.001, 0.0005, 0.002]"
-#    - Automatically validates length and applies corresponding values per transient.
-#
-# 3. 🧩 All updates are backward-compatible with previous interface and parameters.
-# ================================================================
-
-
-import os as os
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.ndimage import gaussian_filter
-from sklearn.linear_model import LinearRegression
-import math
-import re
-import json
-from config import *
-from BaseModule import BaseModule
-
-def get_num(filename):
-    # 定义要匹配的字符串
-    # filename = "(3)PFOA400ppm_75075_50s_CA.xlsx"
-    # filename = "3PFOA400ppm_75075_50s_CA.xlsx"
-
-    # 使用正则表达式匹配数字部分
-    match = re.search(r'(\d+)', filename)
-
-    # 如果匹配成功，则打印括号中的内容
-    if match:
-        result = match.group(1)
-        return int(result)
-    else:
-        return None
-
-
-class CA(BaseModule):
-    def __init__(self, version):
-        super().__init__(version)
-
-    def step1(self):
-        data = self.read_data()
-        for d in data:
-            filename = d['filename']
-            df = d['df']
-
-            t = df['Time (s)']
-            I = df['WE(1).Current (A)']
-            U = df['WE(1).Potential (V)']
-            plt.plot(t, U, linestyle='-', linewidth=1, color='#1f77b4')
-
-        plt.xlabel('time/s')
-        plt.ylabel('Applied potential/V')
-        plt.title('A', loc='left', bbox=dict(facecolor='white', edgecolor='black'))
-        # plt.ylim(-2e-5,2e-5)
-        # plt.grid()
-        # plt.show()
-        to_file1 = os.path.join(self.datapath, "CA_form1_p1.png")
-        plt.savefig(to_file1)
-        plt.close()
-
-
-        for d in data:
-            filename = d['filename']
-            df = d['df']
-
-            t = df['Time (s)']
-            I = df['WE(1).Current (A)']
-            U = df['WE(1).Potential (V)']
-
-            plt.scatter(t, I, s=1, c='#1f77b4')
-
-        plt.xlabel('time/s')
-        plt.ylabel('Current/A')
-        plt.title('B', loc='right', bbox=dict(facecolor='white', edgecolor='black'))
-        # plt.ylim(-2e-5,2e-5)
-        # plt.grid()
-        plt.ticklabel_format(style='sci', axis='y', scilimits=(0, 0))
-
-        to_file2 = os.path.join(self.datapath, "CA_form1_p2.png")
-        plt.savefig(to_file2)
-        plt.close()
-
-        data_file = os.path.join(self.datapath, 'data.json')
-        if os.path.exists(data_file):
-            data = json.loads(open(data_file, 'r').read())
-        else:
-            data = {'version': self.version}
-
-        if 'CA' not in data.keys():
-            data['CA'] = {}
-
-        data['CA']['form1'] = {
-            'status': 'done',
[…]
