### 2025-08-08 07e7b4e
Delete src/CA_commented.py
