+++ src/CA_commented.py
@@ -1,290 +1,0 @@
-"""
-CA.py - Chronoamperometry (CA) Analysis Module
-----------------------------------------------
-
-This module is part of the Envismetrics software suite and provides both visualization
-and diffusion coefficient analysis for Chronoamperometry (CA) data. It reads time-dependent
-potential and current measurements from electrochemical experiments and offers a standardized
-approach to visualize raw signals and compute key kinetic parameters using the Cottrell equation.
-
-Core Functions:
----------------
-1. `step1()`: Raw Data Visualization
-   - Generates two plots per dataset:
-     - A: Applied Potential vs Time
-     - B: Current vs Time
-   - Useful for preliminary data inspection and quality control.
-
-2. `step2(inter, n, a, c, x_range)`: Diffusion Coefficient Calculation
-   - Applies the Cottrell equation to estimate diffusion coefficients (D) from current-time curves.
-   - Performs linear regression on transformed data (Bt vs I), where Bt ~ t^(-1/2).
-   - Produces:
-     - Plot of raw current vs time.
-     - Plot of regression used to extract D.
-     - Summary table (CSV) with slope, D, and R² for each dataset.
-
-Features:
----------
-- Supports batch processing of multiple CA files.
-- Configurable analysis range (`x_range`) for linear regression.
-- Automatically handles file saving, figure generation, and output management.
-- Saves intermediate results and metadata to `data.json` and CSV for reproducibility.
-
-Dependencies:
--------------
-- numpy, pandas, matplotlib, scipy, sklearn
-- config.py
-- BaseModule.py
-
-Date: 2025  
-"""
-
-
-import os
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.ndimage import gaussian_filter
-from sklearn.linear_model import LinearRegression
-import math
-import re
-import json
-from config import *
-from BaseModule import BaseModule
-
-def get_num(filename):
-    """
-    Extracts the first number found in the filename using regex.
-
-    Args:
-        filename (str): File name string (e.g., '3PFOA400ppm_75075_CA.xlsx')
-
-    Returns:
-        int or None: Extracted number, or None if no match found.
-    """
-    match = re.search(r'(\d+)', filename)
-    return int(match.group(1)) if match else None
-
-
-class CA(BaseModule):
-    def __init__(self, version):
-        """
-        Initialize the CA module with a version identifier.
-
-        Args:
-            version (str): Version name or ID for the run.
-        """
-        super().__init__(version)
-
-    def step1(self):
-        """
-        Step 1: Load CA data files, plot potential and current over time, and save outputs.
-        
-        Outputs:
-            - CA_form1_p1.png: Potential vs Time for all files
-            - CA_form1_p2.png: Current vs Time for all files
-            - data.json: Metadata and output file paths
-
-        Returns:
-            dict: Status and result metadata
-        """
-        data = self.read_data()
-        
-        # ----------- Plot Potential vs Time -----------
-        for d in data:
-            filename = d['filename']
-            df = d['df']
-            t = df['Time (s)']
-            U = df['WE(1).Potential (V)']
-            plt.plot(t, U, linestyle='-', linewidth=1, color='#1f77b4')
-
-        plt.xlabel('time/s')
-        plt.ylabel('Applied potential/V')
-        plt.title('A', loc='left', bbox=dict(facecolor='white', edgecolor='black'))
-        to_file1 = os.path.join(self.datapath, "CA_form1_p1.png")
-        plt.savefig(to_file1)
-        plt.close()
-
-        # ----------- Plot Current vs Time -----------
-        for d in data:
-            filename = d['filename']
[…]
