+++ src/CV_commented.py
@@ -1,1559 +1,0 @@
-"""
-CV.py - Cyclic Voltammetry Analysis Module
-------------------------------------------
-
-This module is part of the Envismetrics software suite and provides a complete pipeline
-for analyzing cyclic voltammetry (CV) data obtained from electrochemical experiments.
-
-Core Functions:
----------------
-1. Data Reading and Preprocessing:
-   - Reads .csv, .xlsx, or .txt files with scan rate encoded in the filename.
-   - Extracts potential and current data for individual cycles.
-   - Supports Gaussian smoothing for noise reduction.
-
-2. Step 1: Raw Data Visualization
-   - Plots raw and smoothed CV curves.
-   - Allows selection of a specific scan cycle for plotting.
-
-3. Step 2: Peak Identification
-   - Detects anodic and cathodic peaks in user-defined potential ranges.
-   - Stores peak potentials, currents, and calculates Ef and ΔE0 for each scan.
-
-4. Step 3: Randles–Ševčík Analysis
-   - Uses linear regression to plot Ip vs. sqrt(scan rate).
-   - Calculates diffusion coefficients D for both anodic and cathodic processes.
-
-5. Step 4: Kinetic Parameter Estimation
-   - Estimates heterogeneous rate constants (k0) using the Nicholson method.
-   - Fits ψ–ΔEp and ψ–v⁻¹/² plots to extract slopes for k₀ calculation.
-
-6. Step 5: Tafel Analysis
-   - Two methods implemented to estimate the charge transfer coefficient α.
-   - Method 1: based on d(logJ)/dE (Tafel slope).
-   - Method 2: based on d(ln(I² / (Ip - I)))/dE.
-
-Features:
----------
-- Modular design built on the `BaseModule` foundation.
-- Output is saved in versioned folders under `/outputs`.
-- Fully automatic plotting and result export as .png images and .json/.pkl files.
-- Easily extendable for new electrode systems or devices (Autolab, EC-Lab, etc.)
-
-Usage:
-------
-To use this module independently, run the script directly:
-    python CV.py
-
-You can also use the class methods `start1`, `start2`, `start3`, etc., within your own workflow.
-
-Dependencies:
--------------
-- numpy, pandas, matplotlib, scipy, sklearn
-- config.py (user-defined settings)
-- BaseModule.py (provides save/load utilities)
-"""
-
-
-import os
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.ndimage import gaussian_filter
-from sklearn.linear_model import LinearRegression
-import math
-import re
-import json
-from config import *
-from BaseModule import BaseModule
-import ast
-from datetime import datetime
-
-# Define a color palette for plotting (10 distinct colors)
-colors = [
-    '#1f77b4',  # tab:blue
-    '#ff7f0e',  # tab:orange
-    '#2ca02c',  # tab:green
-    '#d62728',  # tab:red
-    '#9467bd',  # tab:purple
-    '#8c564b',  # tab:brown
-    '#e377c2',  # tab:pink
-    '#7f7f7f',  # tab:gray
-    '#bcbd22',  # tab:olive
-    '#17becf'   # tab:cyan
-]
-
-def find_max(x, y, start, end):
-    """
-    Find the maximum y-value and corresponding x in a given x-range.
-
-    Parameters:
-        x (list or np.array): x-values
-        y (list or np.array): y-values
-        start (float): lower x-bound
-        end (float): upper x-bound
-
-    Returns:
-        tuple: (x, y) position of maximum y-value within range
-    """
-    ma = -1
-    xx = -1
-    yy = -1
-    for i in range(len(x)):
-        if start <= x[i] <= end:
-            if y[i] > ma:
-                ma = y[i]
-                xx = x[i]
-                yy = y[i]
[…]
