### 2025-08-08 1f0bac3
Delete src/HDV_commented.py
