+++ src/HDV_commented.py
@@ -1,765 +1,0 @@
-"""
-HDV.py - Hydrodynamic Voltammetry (HDV) Analysis Module
---------------------------------------------------------
-
-This module is part of the Envismetrics software suite and performs comprehensive 
-analysis of Hydrodynamic Voltammetry (HDV) data, including both Levich and 
-Koutecky-Levich plots for determining diffusion coefficients.
-
-Main Features:
---------------
-1. **step1(sigma)**:
-   - Loads all valid HDV data files (Excel or TXT), sorted by RPM.
-   - Visualizes raw and Gaussian-filtered current vs potential curves.
-   - Stores PNG figures for raw and smoothed voltammograms.
-
-2. **step2_1(all_params)**:
-   - Levich plot analysis for each potential point or interval.
-   - Calculates slope (B) and diffusion coefficient (D).
-   - Produces:
-     - Individual regression plots across rotation rates.
-     - Summary slope-D plots vs potential.
-     - Exportable CSV file with current vs √ω data.
-
-3. **step2_2(all_params)**:
-   - Koutecky-Levich plot analysis for kinetic and diffusion current separation.
-   - Estimates diffusion coefficients using inverse current vs ω⁻¹ᐟ².
-   - Produces:
-     - Regression line plots per potential.
-     - Combined B-D dual-axis plots.
-     - Exportable CSV of calculated parameters.
-
-Core Utilities:
----------------
-- `read_data()`: Handles file parsing, RPM extraction, and CSV caching.
-- `rpm_to_rads()`: Converts RPM to angular velocity (rad/s).
-- Utility functions: `find_y`, `extract_rpm`, `check_files`, etc.
-
-Inputs:
--------
-- User-defined electrochemical parameters: number of electrons (n), area (A), viscosity (ν), and bulk concentration (C).
-- Potential range, number of points, and sampling interval.
-
-Output:
--------
-- PNG plots, CSV results, and progress tracking in structured JSON.
-
-Dependencies:
--------------
-- numpy, pandas, matplotlib, scipy, sklearn
-- BaseModule.py
-- JSON configuration files
-
-Date: 2025  
-"""
-import re
-import numpy as np
-import pandas as pd
-import os as os
-import matplotlib
-import matplotlib.pyplot as plt
-import math
-import json
-from BaseModule import BaseModule
-from sklearn.linear_model import LinearRegression
-from scipy.ndimage import gaussian_filter
-matplotlib.use('Agg')
-
-# Function to find y value (I) corresponding to given x value (potential)
-# It finds the index of the element in the array x that is closest to the target_x value using
-def find_y(x, y, target_x):
-    index = np.argmin(np.abs(x - target_x))
-    return y[index]
-
-def find_y_exact(x,y,target_x):
-    index = np.where(x == target_x)[0]
-    return y[index]
-
-def rpm_to_rads(rpm):
-    rps = rpm / 60  # Convert RPM to RPS (revolutions per second)
-    rad_per_sec = 2 * math.pi * rps  # Convert RPS to rad/s
-    return rad_per_sec
-
-def reorder(filename):
-    match = re.search(r'(\d+)rpm', filename)
-    if match:
-        return int(match.group(1))
-    else:
-        # Handle files without RPM values
-        return -1  # You can use any default value or treatment
-
-def extract_rpm(filename):
-    pattern = r'(?:^|_)(\d+rpm)\.'
-    match = re.search(pattern, filename)
-    if match:
-        return match.group(1)
-    else:
-        return None
-
-def check_files(files):
-    for f in files:
-        ext = f.split('.')[-1].lower()
-        if ext not in ['xlsx', 'txt']:
-            return False
-    return True
-
-class HDV(BaseModule):
-    def __init__(self, version):
-        super().__init__(version)
-
-
-    def read_data(self):
[…]
