### 2025-08-08 Woffee/Envismetrics#14 @Grant-Xue
Thank you very much for your detailed feedback. I'd like to address the recent concerns as follows:

### 1. Regarding inline comments and "why" explanations

We've noted your point that many comments currently focus on *what* is being done rather than *why*. We expanded the inline comments to include the rationale behind key steps and design decisions, especially in areas where scientific assumptions or architectural choices matter (e.g., how certain equations are applied or why specific preprocessing is required). We'll push an updated commit reflecting these improvements soon.

### 2. On the use of `_commented.py` variants and multiple files (e.g., `CA_1.py`)

The current state includes commented versions that were uploaded **for internal testing only**. Because some comments required refactoring (e.g., changing variable names for clarity), we wanted to avoid unintended disruptions to the main codebase before fully verifying consistency across modules.

Our plan is as follows:
- Finalize and verify all `*_commented.py` versions through testing;
- Merge each into the primary module file (`CA.py`, `CV.py`, `HDV.py`) **with only comments and docstring edits**, ensuring no functional diffs remain;
- Remove legacy or temporary files (`CA_1.py`, etc.) in the same cleanup commit, to ensure a single, authoritative script per module.

### 3. On `BaseModule.py` comments

`BaseModule` underpins all modules and deserves clear documentation. We will add professional docstrings and inline comments (including reasons for design decisions like `self.version` handling, path setup, etc.) and commit them alongside the final module files.

### 4. On the Flask `app.py` file and route structure

We acknowledge that the structure may not be intuitive to all readers. Each `@app.route` indeed corresponds to a step/page in the web interface. We're currently annotating this file with block-level and route-level comments that:

- Describe which module (CV, CA, HDV) each route serves;
- Explain how the `step` or `func` query parameters control dynamic behavior;
- Clarify how uploaded data maps to the appropriate processing pipeline.

We aim to make the `app.py` logic comprehensible even to those unfamiliar with Flask.
