### 2025-08-08 8b22ff3
Delete src/BaseModule.py
