+++ src/BaseModule.py
@@ -1,0 +1,230 @@
+"""
+BaseModule.py
+
+This file defines the BaseModule class, which serves as the parent class for all electrochemical 
+analysis modules in Envismetrics (e.g., CV, HDV, CA).
+
+It provides core functionalities for:
+- Managing uploaded user data and file paths
+- Reading and converting electrochemical data files (.xlsx, .txt, .csv)
+- Maintaining analysis state across steps via JSON storage
+- Saving and loading intermediate data structures using pickle
+
+Utility functions are also provided for filename parsing and file format checks.
+
+Last updated: August 8, 2025
+License: MIT License
+
+Dependencies:
+    - os, json, re: file handling and text parsing
+    - pandas: data processing
+    - pickle: binary object serialization
+    - config: project-specific paths (UPLOAD_FOLDER, etc.)
+
+Example usage:
+    base = BaseModule(version="version_0808_1530")
+    df_list = base.read_data()
+    base.save_result_data(...)
+    data = base.pkl_load("some_object.pkl")
+"""
+
+
+import os
+import json
+import re
+import pandas as pd
+import pickle
+from config import *
+
+# === Utility Functions ===
+
+def reorder(filename):
+    """
+    Extract RPM value from the filename and return it as an integer.
+
+    Used to sort files based on their rotation speed.
+    Returns -1 if no RPM pattern is found, so these files will be placed at the beginning.
+    """
+    match = re.search(r'(\d+)rpm', filename)
+    if match:
+        return int(match.group(1))
+    else:
+        return -1  # Default sorting value if RPM not found
+
+
+def extract_rpm(filename):
+    """
+    Extract RPM string from filename using regex pattern.
+
+    Example:
+        Input: "sample_500rpm.xlsx"
+        Output: "500rpm"
+    """
+    pattern = r'(?:^|_)(\d+rpm)\.'
+    match = re.search(pattern, filename)
+    if match:
+        return match.group(1)
+    else:
+        return None
+
+
+def check_files(files):
+    """
+    Check whether all input files have supported extensions (.xlsx or .txt).
+
+    Returns:
+        True if all files are valid; otherwise, False.
+    """
+    for f in files:
+        ext = f.split('.')[-1].lower()
+        if ext not in ['xlsx', 'txt']:
+            return False
+    return True
+
+# === Base Class for All Modules ===
+
+class BaseModule(object):
+    """
+    Parent class for CV, HDV, and CA modules.
+
+    Handles:
+    - Reading uploaded file metadata
+    - Preprocessing raw files into DataFrames
+    - Saving/loading intermediate result files (JSON and pickle formats)
+    """
+
+    def __init__(self, version):
+        """
+        Constructor for BaseModule.
+
+        INPUT:
+            version (str): Unique identifier for this analysis run (used as folder name).
+
+        EFFECT:
+            Initializes paths for input and output data folders, reads result data if available.
+        """
+        self.version = version
+        self.files_info = os.path.join('uploads', version, 'fileinfo.json')
+        self.datapath = os.path.join('outputs', self.version)
+
+        if not os.path.exists(self.datapath):
+            os.mkdir(self.datapath)
+
+        # Load existing result data (e.g. form1, form2, etc.), or initialize a new dictionary
+        self.res_data = self.read_result_data()
+
+
+    def read_result_data(self):
+        """
+        Read existing result JSON (data.json) from disk, or initialize a new dictionary.
+
+        OUTPUT:
+            dict containing previously saved analysis results.
+        """
+        data_file = os.path.join(self.datapath, 'data.json')
+        if os.path.exists(data_file):
+            data = json.loads(open(data_file, 'r').read())
+        else:
+            data = {'version': self.version}
+        return data
[…]
