### 2025-08-08 9e71636
update figures' labels
