+++ src/app.py
@@ -1,15 +1,49 @@
 """
-运行逻辑：
+app.py - Main Flask Application for Envismetrics Electrochemical Analysis Suite
+-------------------------------------------------------------------------------
 
-1. 使用 JS 背后提交数据。
-2. 提交之后展跳转到结果页面，每秒检查一次是否执行完毕。
-    /<module>/<version>?step=xx
-    /check?module=xx&version=xx&step=xx
-3. 后台执行任务，任务完成后会生成结果文件。
-    /upload?module=xx&version=xx&step=xx
-    background_task()
-4. 前台检查执行完毕后，刷新页面，显示结果。
+This module defines the backend server logic for the Envismetrics software package.
+It provides a web-based user interface (UI) for performing electrochemical data
+analysis using CV (Cyclic Voltammetry), CA (Chronoamperometry), and HDV (Hydrodynamic Voltammetry) modules.
 
+Core Responsibilities:
+----------------------
+- Sets up and launches a Flask web application to serve Envismetrics UI.
+- Handles routing for various views including:
+    - Main landing page
+    - About page
+    - File upload and processing
+    - Results visualization and downloads
+- Coordinates frontend interactions with backend data analysis classes.
+
+Features:
+---------
+- Uploads and stores user data to a designated session directory.
+- Dynamically loads module-specific analysis pipelines (CV, CA, HDV).
+- Manages user sessions and filesystem separation by version (timestamp).
+- Uses Jinja2 templates for rendering HTML interfaces.
+
+Structure:
+----------
+- Uses the Flask web framework for routing and request handling.
+- Static files (JS, CSS) served from /static directory.
+- HTML templates located in /templates directory.
+
+Dependencies:
+-------------
+- Flask: lightweight web framework for Python
+- os, shutil, uuid: for session and file handling
+- Envismetrics modules (CV.py, CA.py, HDV.py)
+
+Typical Usage:
+--------------
+To launch the application, run this file directly:
+
+    $ python app.py
+
+Then navigate to http://localhost:5000/ in a web browser.
+
+Date: 2025
 """
 
 from werkzeug.utils import secure_filename
@@ -33,56 +67,137 @@ from CA import CA
 import sys
 import os
 
+# Add the current file's directory to the system path to enable local imports (e.g., config.py, utils.py)
 sys.path.append(os.path.dirname(__file__))
+
+# Import global configuration variables and constants (e.g., UPLOAD_FOLDER, ALLOWED_EXTENSIONS)
 from config import *
+
+# Import utility functions for logging and folder validation
 from utils import init_logging, check_folders
 
+# Initialize the Flask web application
 app = Flask(__name__)
+
+# Configure the upload directory for storing user-submitted files
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
+# Ensure required folders (e.g., upload, output) exist; create if missing
 check_folders()
+
+# Initialize logging system to record backend runtime events
 init_logging()
 
 def allowed_file(filename):
+    """
+    Check whether the uploaded filename has an allowed extension.
+
+    Args:
+        filename (str): The uploaded file's name.
+
+    Returns:
+        bool: True if file extension is allowed, False otherwise.
+
+    Purpose:
+        Prevents users from uploading unsupported file types, based on ALLOWED_EXTENSIONS in config.py.
+    """
     return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
 
 @app.route("/doc")
 def demo():
+    """
+    Route: /doc
+
+    Renders the documentation page.
+
+    Returns:
+        HTML template: Renders doc.html, optionally with notes or instructions.
+
+    Note:
[…]
