### 2025-08-11 Woffee/Envismetrics#12 @Grant-Xue
> *"My point is more that your file format is very limiting - would be better to have an entry box next to filename allowing you to enter the scan rate/rpm so that I don't need to rename all my files to comply with your format. It may not even be necessary, you could adapt `pattern = r'(?:^|_)(\d+rpm)\.' match = re.search(pattern, filename)` to extract the values you need as long as they are anywhere in the filename. The conditions have got even more restrictive, now CA need to start with a number - and everything has to have the module name at the end of the filename (not sure why this is necessary - it seems to work regardless and if I am uploading to CA module, I should know it needs to be CA data)."*

We appreciate this feedback and agree that the current filename restrictions are unnecessarily strict in some cases. The original purpose of requiring module suffixes (e.g., `_CA`, `_CV`, `_HDV`) was to prevent uploading data to the wrong module. However, since the module is already determined at upload time, this requirement adds friction without significant benefit.

**Planned changes:**
- Remove the `_CV`, `_HDV`, and `_CA` suffix requirement when uploading to the corresponding module.
- Remove the “number at the start” requirement for CA filenames.
- Relax the scan rate/RPM filename parsing to allow extraction from anywhere in the filename using a regex similar to the one suggested (`r'(?:^|_)(\d+rpm)\.'`).
- Retain filename-based scan rate/RPM extraction for now, as most users still store this information in filenames to avoid mixing datasets.

---

> *"Then you need to make sure all references to your software being useable for different workstations are removed from manuscript, app and MD files. It seems you can only use it on Autolab data."*

We will update the manuscript, app text, and MD files to clarify workstation compatibility. Currently:
- **Fully supported:** Autolab formats (CSV/TXT)
- **Partially supported:** EC-Lab `.txt` files (currently implemented in CV, soon to be extended to other modules)

We will also add EC-Lab sample data to the repository so users can test compatibility.

---

> *"Good, but you don't mention anything about .txt files in this MD file, and they should be equally valid in CA."*

We will update the `README_data_structure.md` to explicitly document `.txt` file formats, ensuring they are treated as equally valid alongside `.csv`.
