### 2025-08-11 Woffee/Envismetrics#18 @Grant-Xue
@MoonshinetheP
It seems that much of the confusion was caused by the way the test data was originally organized.
Previously, I placed data from 7 different experiments into a single folder, which unintentionally gave the impression that the software was combining results.

I have now restructured the test data directory so that each experiment is stored in its own folder. You can see the updated structure here:
https://github.com/Woffee/Envismetrics/tree/main/data/test_data/05132024_CA_D

From now on, please place the files from only one experiment (in a single folder) when running the analysis.
The software does not combine results from multiple experiments into one plot, and it is not designed to support that workflow.

I will address the remaining issues very soon.
