+++ src/CA.py
@@ -35,9 +35,12 @@ Dependencies:
 - numpy, pandas, matplotlib, scipy, sklearn
 - config.py
 - BaseModule.py
+
+Date: 2025
 """
 
-import os as os
+
+import os
 import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
@@ -51,35 +54,50 @@ from BaseModule import BaseModule
 
 def get_num(filename):
     """
-    Extracts the first number found in the filename using regex.
+    Extract the first integer number found in a filename using regular expressions.
 
-    Args:
-        filename (str): File name string (e.g., '3PFOA400ppm_75075_CA.xlsx')
+    Parameters:
+        filename (str): Input filename string (e.g., '3PFOA400ppm_75075_CA.xlsx').
 
     Returns:
-        int or None: Extracted number, or None if no match found.
+        int or None: The first numeric match converted to int, or None if no number is found.
 
+    Notes:
+        This is commonly used to extract concentration, ID, or numeric prefixes from filenames.
     """
     match = re.search(r'(\d+)', filename)
-
-    if match:
-        result = match.group(1)
-        return int(result)
-    else:
-        return None
-
+    return int(match.group(1)) if match else None
 
 class CA(BaseModule):
     def __init__(self, version):
         """
-        Initialize the CA module with a version identifier.
+        Initialize the Chronoamperometry (CA) module.
+
+        This class inherits from BaseModule and sets up the environment
+        for processing CA data associated with a given analysis version.
 
         Args:
-            version (str): Version name or ID for the run.
+            version (str): Unique identifier for this session/run.
+                           Used to organize output directories and results.
         """
         super().__init__(version)
 
     def check_columns(self, data):
+        """
+        Validate that all required columns are present in the input DataFrames.
+
+        Required columns:
+            - 'Time (s)': Time axis in seconds.
+            - 'WE(1).Current (A)': Working electrode current in amperes.
+            - 'WE(1).Potential (V)': Working electrode potential in volts.
+
+        Args:
+            data (list): List of dictionaries, each containing a 'df' key with a pandas DataFrame.
+
+        Returns:
+            str: Empty string if all columns are present; otherwise, returns an error message
+                 listing all missing columns.
+        """
         cols = ['Time (s)', 'WE(1).Current (A)', 'WE(1).Potential (V)']
         missing_cols = []
         for d in data:
@@ -93,76 +111,80 @@ class CA(BaseModule):
 
     def step1(self):
         """
-        Step 1: Load CA data files, plot potential and current over time, and save outputs.
+        Step 1: Visualization of raw Chronoamperometry data.
 
-        Outputs:
-            - CA_form1_p1.png: Potential vs Time for all files
-            - CA_form1_p2.png: Current vs Time for all files
-            - data.json: Metadata and output file paths
+        This method:
+        1. Loads pre-processed CA data using `read_data()`.
+        2. Verifies required columns using `check_columns()`.
+        3. Generates and saves two plots:
+            - Plot A: Applied potential (V) vs. time (s)
+            - Plot B: Current (A) vs. time (s)
+        4. Saves output metadata to a JSON file (`data.json`) for reproducibility.
+
+        Output Files:
+            - CA_form1_p1.png: Potential vs Time
+            - CA_form1_p2.png: Current vs Time
 
         Returns:
-            dict: Status and result metadata
+            dict: A structured response containing success status, version, result paths,
+                  or error message if any step fails.
         """
         data_file = os.path.join(self.datapath, 'data.json')
+
+        # Load previous metadata or initialize a new record
[…]
