+++ src/HDV.py
@@ -50,8 +50,8 @@ Dependencies:
 - BaseModule.py
 - JSON configuration files
 
+Date: 2025
 """
-
 import re
 import numpy as np
 import pandas as pd
@@ -67,61 +67,215 @@ matplotlib.use('Agg')
 
 # Function to find y value (I) corresponding to given x value (potential)
 # It finds the index of the element in the array x that is closest to the target_x value using
+
 def find_y(x, y, target_x):
+    """
+    WHAT:
+        Given two 1D arrays `x` and `y` (of equal length), return the value in `y`
+        that corresponds to the value in `x` closest to `target_x`.
+
+    INPUT:
+        x : array-like
+            Independent variable array (e.g., time, voltage).
+        y : array-like
+            Dependent variable array (e.g., current, intensity).
+        target_x : float
+            The x-value we want to find the closest match for.
+
+    OUTPUT:
+        y_value : float
+            The corresponding y-value at the nearest x to `target_x`.
+
+    WHY:
+        Experimental or simulation data often doesn't have exact x values;
+        this allows nearest-neighbor interpolation to approximate the y-value.
+    """
+    # Find the index of the x-value closest to target_x
     index = np.argmin(np.abs(x - target_x))
+    # Return the corresponding y-value
     return y[index]
 
-def find_y_exact(x,y,target_x):
+
+def find_y_exact(x, y, target_x):
+    """
+    WHAT:
+        Returns the y-values where x matches `target_x` exactly.
+
+    INPUT:
+        x : array-like
+            Independent variable array.
+        y : array-like
+            Dependent variable array.
+        target_x : float or int
+            The x-value to match exactly.
+
+    OUTPUT:
+        y_values : array
+            All y-values at positions where x == target_x.
+            (May be empty if no exact match is found.)
+
+    WHY:
+        Use this when exact x-values are expected and meaningful.
+        Ensures strict matching, useful for discrete systems.
+    """
+    # Find indices where x equals target_x
     index = np.where(x == target_x)[0]
+    # Return the matching y-values
     return y[index]
+
 
 def rpm_to_rads(rpm):
-    rps = rpm / 60  # Convert RPM to RPS (revolutions per second)
-    rad_per_sec = 2 * math.pi * rps  # Convert RPS to rad/s
+    """
+    WHAT:
+        Convert rotational speed in RPM (revolutions per minute) to angular
+        velocity in radians per second.
+
+    INPUT:
+        rpm : float or int
+            Rotational speed in revolutions per minute.
+
+    OUTPUT:
+        rad_per_sec : float
+            Angular velocity in radians per second.
+
+    WHY:
+        Most physics/engineering formulas involving angular motion
+        require radians/sec as standard SI unit.
+    """
+    rps = rpm / 60              # Convert RPM to RPS (revolutions per second)
+    rad_per_sec = 2 * math.pi * rps  # Convert to radians/second
     return rad_per_sec
 
+
 def reorder(filename):
+    """
+    WHAT:
+        Extract an integer RPM value from a filename string (e.g., 'test_500rpm.xlsx').
+
+    INPUT:
+        filename : str
+            The name of the file containing an RPM tag like '1000rpm'.
+
+    OUTPUT:
+        rpm : int
+            Extracted RPM as integer, or -1 if not found.
+
+    WHY:
+        Used to numerically sort or organize files based on RPM value.
+        Returns -1 for filenames without valid RPM to handle edge cases safely.
+    """
     match = re.search(r'(\d+)rpm', filename)
     if match:
-        return int(match.group(1))
+        return int(match.group(1))  # Extract numeric portion of 'xxxrpm'
     else:
[…]

[…]
