### 2025-08-12 5a121ba
Remove the requirement for the filename to have a module suffix at the end.
