### 2025-08-12 Woffee/Envismetrics#18 @Grant-Xue
> *"That is why they also give you the 'Corrected time / s' column, which you could use instead to make sure each transient starts at t = 0… For Cottrell, you need time starting from 0…"*

In our experiments the **first 120 s step is an intentional, user-defined pretreatment** to equilibrate the electrode/solution interface (remove residual films, stabilize diffusion layer, etc.). This is an experimental design choice and **independent of any workstation-provided “Corrected time” column**.

- For the overview plots (*Applied potential vs. time* and *Current vs. time*), we **retain the raw time axis** on purpose, so users can see the full sequence, including the pretreatment step and subsequent platforms.
- For **Cottrell analysis**, we already **zero each transient** (subtract the first timestamp of that transient) before fitting at the beginning, i.e., the fit is performed with \(t=0\) at the onset of the analyzed step. There’s no disagreement on using \(t=0\) for the fit—this is already what the code does.

---

> *"I don't really understand what is going on with the data in your test folders… ‘1_…’ all seem to be 120 s… ‘2_…’ start at 121.76 s; ‘3_…’ at 123.77 s + previous durations."*

This pattern reflects our **experimental protocol**, not software stitching. Every test experiment starts with a **120 s pretreatment platform** (step 1). Consequently, step 2 begins around ~120 s, step 3 around ~120 s + duration of step 2, and so on—hence the offsets you observed in raw time.

To avoid confusion, we have **restructured one of the test data** so that **each experiment resides in its own folder** (and filenames indicate step order/duration). The software does **not** concatenate independent experiments; it plots files **in the order provided** with their original timestamps.

Updated example:
https://github.com/Woffee/Envismetrics/tree/main/data/test_data/05132024_CA_DD

---
