### 2025-08-12 Woffee/Envismetrics#18 @Grant-Xue
> *"…you are considering only how your data is structured… you need to account for people using it in the way they want (they may not care about 'duplicate prefixes', and they may indeed want to see two consecutive transients). This is a pretty fundamental element to developing software for people to use in the laboratory - make it flexible."*

Thanks for raising this. To clarify our intended workflow:

- **Multi-step chronoamperometry (consecutive transients within one experiment) is supported.**
  Our paper (Fig. 4a) shows a **7-step CA trace (current vs time)** from a single experiment.

- The earlier “stacking” you observed came from our **test folder mixing multiple experiments together** in one session (my mistake in data organization), not from the software joining traces.
  We have since **restructured the test dataset** so that **each experiment resides in its own folder**, which avoids this confusion.

- **What we don’t support** (by design for now) is uploading **different experiments simultaneously** into one CA session, because that can produce misleading overlays.

**What we’ll do next (short-term):**
   - Restructured CA test files **one experiment per analysis session**; multi-step CA within that experiment is fine.
