### 2025-08-12 Woffee/Envismetrics#18 @Grant-Xue
> *"A long way to say we will address this in the future. Make sure it is adequately explained in the development roadmap, then"*

**Response:**
We agree that per-transient parameter input is essential for datasets where conditions such as concentration (c) and number of electrons (n) vary between potential platforms. As suggested,  ⚠️ we are almost **already implementing** an "Advanced Mode" in which users can enter these parameters as a JSON-style list (e.g., `c = [1e-3, 5e-4, 2e-3, ...]`), with the list length automatically matching the number of uploaded transients. Each entry is mapped to the corresponding step, ensuring accurate per-step analysis.

This is in line with recommended practices for multi-step electrochemical measurements, where parameters may differ between steps

A more user-friendly **parameter input table** (spreadsheet-style or CSV upload) is planned as a **future enhancement** and will be noted in the development roadmap.
