### 2025-08-12 Woffee/Envismetrics#18 @Grant-Xue
> *"No, this is something you are doing in this specific example. Again, if you want to publish the code you use as a software for others to use, it needs to be adaptable to other people's needs. Furthermore, this approach also assumes that people are saving their pretreatment steps as separate files, then uploading them so they can not be shown (and also shows that your approach cannot deal with double-step chronoamperommetry, which is more typical for D estimation than making several transients with different durations one after the other). In general, uploading something means I want to see it in full, you don't save data just to not look at it."*

We agree with the reviewer’s point that users should have access to all recorded potential steps, including the first one, as experimental setups and requirements can vary.
We have therefore modified the CA module so that **all potential steps are plotted and available for analysis by default**, rather than skipping the first step.
This ensures users can review their complete dataset in the software.
If the first step corresponds to a preconditioning period (as in our own workflow), users can simply disregard it during interpretation.
