### 2025-08-12 Woffee/Envismetrics#18 @Grant-Xue
> *"The user doesn't need to do anything, it's all in the calculations within your CA.py files. You have all the relevant constants already when you calculate Bt = ((n * F * A * C0) / (math.pi ** 0.5)) * t_inverse_05, you just need to use them to calculate D from the gradient of i vs 1/SQRT(t). A more 'immediately actionable' approach would be to show D, not gradient that equals SQRT(D)."*

The CA module already outputs the **diffusion coefficient \(D\)** in the downloadable results file.
On-screen we plot the regression for \(I\) vs. \(B(t)\) (slope = \(\sqrt{D}\)), and the export includes both the slope and the computed \(D\) (plus \(R^2\)) for each step.

**Example (from the results file):**

| Interval   | Slope       | D         | R²        |
|------------|-------------|-----------|-----------|
| interval2  | -0.00219997 | 4.84E-06  | 0.8133    |
| interval3  | 0.00354664  | 1.26E-05  | 0.9840    |
| interval4  | 0.00364917  | 1.33E-05  | 0.9845    |
| interval5  | 0.00968129  | 9.37E-05  | 0.9326    |
| interval6  | -0.00165190 | 2.73E-06  | 0.9320    |
| interval7  | -0.00313199 | 9.81E-06  | 0.8742    |
| interval8  | -0.00161681 | 2.61E-06  | 0.8980    |

To make this more visible, we now also **display \(D\) directly in the UI** alongside the slope, so users don’t need to download the file to see it.
