+++ src/utils.py
@@ -10,10 +10,11 @@ def check_folders():
         os.makedirs(UPLOAD_FOLDER, exist_ok=True)
     if not os.path.exists(OUTPUT_FOLDER):
         os.makedirs(OUTPUT_FOLDER, exist_ok=True)
+    os.makedirs(TMP_FOLDER, exist_ok=True)
 
 def init_logging():
     today = time.strftime("%Y-%m-%d", time.localtime())
     logging.basicConfig(level=logging.INFO,
                         format='%(asctime)s %(levelname)s %(filename)s line: %(lineno)s - %(message)s',
                         datefmt='%Y-%m-%d %H:%M:%S',
-                        filename=LOG_DIR + '/' + today + '.log')
+                        filename=os.path.join(LOG_DIR , today + '.log'))
