### 2025-08-13 3c9c359
show D in the CA results
