+++ requirements.txt
@@ -6,3 +6,12 @@ scikit-learn==1.0.2
 scipy==1.9.1
 openpyxl==3.0.10
 matplotlib==3.5.2
+
+## If windows users got errors, try these:
+
+# flask==2.0.3
+# jinja2==3.0.3
+# werkzeug==2.0.3
+# markupsafe==2.0.1
+# click==8.0.4
+# itsdangerous==2.0.1
