### 2025-08-13 b1ec4f2
Update paper.md
