### 2025-08-13 673d35f
Update paper.md
