+++ joss-paper/paper.md
@@ -57,27 +57,25 @@ Each technique supports specific analyses:
 
 While these methods are widely accepted, manual analysis can be labor-intensive and prone to inconsistency. To address this, **Envismetrics** is introduced as an open-source, browser-based Python application that automates data processing and analysis workflows for CV, LSV (RDE), and CA. It provides modules for filtering, peak detection, Levich regression, Randles–Ševčík analysis, and chronoamperometric fitting—offering visual outputs and tabulated results. By focusing on automation and reproducibility, Envismetrics lowers the barrier for electrochemical researchers—especially those dealing with large datasets or requiring rapid feedback—while preserving methodological rigor and transparency.
 
-# Statement of Need
+## Statement of Need
 
-Electrochemical researchers typically process and visualize experimental data using manual spreadsheet workflows (e.g., Excel) or general-purpose plotting software (e.g., Origin, SigmaPlot). While these tools are flexible, they require manual preprocessing, repeated formatting, and significant domain expertise for kinetic modeling. Proprietary instrument software (e.g., NOVA for Autolab) is primarily designed for device control and data acquisition; although it offers basic plotting, it is platform- and instrument-specific, and is rarely used as the main tool for in-depth data analysis [@Garg2021].
+Electrochemical researchers often rely on a mix of tools for data analysis and visualization. Manual spreadsheet workflows (e.g., Excel) and general-purpose plotting software (e.g., Origin, SigmaPlot) are flexible, but they require significant manual preprocessing, repeated formatting, and substantial domain expertise for kinetic modeling. Proprietary instrument software (e.g., NOVA for Autolab) is primarily intended for device control and data acquisition; while it offers basic plotting, it is instrument-specific and is rarely used for in-depth data analysis [@Garg2021]. Researchers with strong programming skills sometimes develop custom analysis scripts in MATLAB or Python, but this demands considerable coding effort, debugging, and domain-specific implementation, which can be a barrier for many users.
 
-Envismetrics addresses this gap by providing a modular, web-based platform dedicated to the automated analysis of electrochemical data, including cyclic voltammetry (CV), rotating disk electrode linear sweep voltammetry (LSV at RDE), and chronoamperometry (CA). It supports widely used plaintext formats such as .xlsx, .csv, and .txt—enabling users to export their data from proprietary systems and perform consistent, reproducible analysis without being tied to a specific instrument vendor. Its modular architecture makes it straightforward to incorporate new analytical techniques. Initially optimized for files exported from Autolab’s NOVA and BioLogic’s EC-Lab software, Envismetrics is continuously expanding compatibility with both new instruments and analytical workflows.
+Envismetrics addresses this gap by providing a modular, web-based platform dedicated to automated analysis of electrochemical data, including cyclic voltammetry (CV), rotating disk electrode linear sweep voltammetry (LSV at RDE), and chronoamperometry (CA). It supports widely used plaintext formats such as `.xlsx`, `.csv`, and `.txt`, allowing researchers to export data from most proprietary workstation software and perform consistent, reproducible analysis without being tied to a specific vendor.  
 
-While researchers with strong programming skills may develop custom analysis scripts in environments such as MATLAB or Python, this approach demands substantial coding effort, debugging, and domain-specific implementation. Envismetrics lowers this barrier by providing ready-to-use, validated analysis pipelines through an interactive browser interface—eliminating the need to write or maintain code while still enabling reproducible, script-like workflows.
+Its modular architecture supports both platform expansion (potential of compatibility with additional electrochemical workstations) and method expansion (new analytical techniques). For example, while Envismetrics currently supports CV, LSV at RDE, and CA, it can be extended to include electrochemical impedance spectroscopy (EIS) and other workflows with minimal changes to the core system.
 
[…]
