### 2025-08-13 Woffee/Envismetrics#9 @Grant-Xue
@MoonshinetheP

Thank you for the detailed feedback. We have made the following targeted revisions in line with your comments:

1. **Refocused comparison scope** – The Statement of Need and Table 1 now compare Envismetrics primarily to widely used data analysis tools such as Excel, Origin, and SigmaPlot, rather than to instrument control software like NOVA. Mentions of workstation software are retained only to explain data acquisition context, without implying direct competition.

2. **Clarified data import requirements** – We replaced “plaintext from any system” with:
   > “Supports `.xlsx`, `.csv`, and `.txt` formats exported from proprietary workstation software, enabling consistent analysis once data is available in a supported text-based format.”
   This makes explicit that users must first export their data from instrument software before using Envismetrics.

3. **Streamlined Table 1** – Removed columns/rows unrelated to Envismetrics’ capabilities (e.g., Hardware Dependency, Data Collection). Merged “Installation” and “Platform Support” into a single row. The table now focuses on usability, extensibility, and analysis capabilities—the aspects most relevant to software comparison.

We believe these changes make the scope, capabilities, and comparison points clearer and more consistent with your suggestions.
