### 2025-08-13 923839b
Update paper.md
