+++ joss-paper/paper.md
@@ -112,7 +112,11 @@ To aid in interpreting the equations below, Table 2 summarizes commonly used ele
 
 ## Data Processing
 
-Envismetrics supports commonly used plain-text electrochemical data formats, including .xlsx, .csv, and .txt, exported from potentiostat software such as EC-Lab (BioLogic), NOVA (Autolab), and Metrohm. Users can directly upload these files to the web-based interface without additional preprocessing, provided the files follow standard export structures. The software automatically parses time, current, and potential data for downstream analysis. File validation is included to help ensure compatibility and inform users of any formatting issues. Support for additional instrument vendors and data formats is actively expanding and documented in the repository.
+Envismetrics supports commonly used plain-text electrochemical data formats, including .xlsx, .csv, and .txt, exported from potentiostat software such as NOVA (Autolab Metrohm) and CV EC-Lab (BioLogic). Users can directly upload these files to the web-based interface without additional preprocessing, provided the files follow standard export structures. The software automatically parses time, current, and potential data for downstream analysis. File validation is included to help ensure compatibility and inform users of any formatting issues. Support for additional instrument vendors and data formats is actively expanding and documented in the repository.
+
+## Data Processing
+
+Envismetrics supports plain-text electrochemical data formats, including `.xlsx`, `.csv`, and `.txt`, exported from supported potentiostat software. The current release has been validated with files from Autolab’s NOVA and BioLogic’s EC-Lab (for CV analysis), with additional formats planned for future updates as documented in the repository. Users can upload these files directly to the web-based interface without additional preprocessing, provided they follow the standard export structures of the supported software. The software automatically parses time, current, and potential data for downstream analysis, with built-in file name and format validation to ensure compatibility and alert users to formatting issues.
 
 ![Data Import Window: Users can easily drag and drop or select their experimental data for quick and straightforward import.](Image_Set/1.png){ width=80% }
 
