+++ joss-paper/bibliography.bib
@@ -1,3 +1,33 @@
+@article{Klingler1981,
+author = {Klingler, R. J. and Kochi, J. K.},
+title = {Electron-transfer kinetics from cyclic voltammetry. Quantitative description of electrochemical reversibility},
+journal = {The Journal of Physical Chemistry},
+volume = {85},
+number = {12},
+pages = {1731-1741},
+year = {1981},
+doi = {10.1021/j150612a028},
+
+URL = { 
+    
+        https://doi.org/10.1021/j150612a028
+    
+    
+
+},
+eprint = { 
+    
+        https://doi.org/10.1021/j150612a028
+    
+    
+
+}
+
+}
+
+
+
+
 @article{TORABFAM2025107950,
 title = {Elucidating the electrochemical behavior and reaction pathway of 1,4-Dioxane: An integrated experimental and simulation approach},
 journal = {Electrochemistry Communications},
