### 2025-08-13 20c1d28
Update paper.md
