### 2025-08-13 61b7ac8
Update paper.md
