### 2025-08-13 974db52
Update paper.md
