### 2025-08-13 2bc653a
Update paper.md
