+++ joss-paper/paper.md
@@ -159,7 +159,7 @@ In *Envismetrics*, users can select potential values to automatically generate t
 
 <figure style="width: 50%; margin: auto; text-align: center;">
   <img src="Image_Set/KL_D23.png" alt="Koutecky–Levich plot module" style="width: 100%;" />
-  <figcaption><strong>Figure 2.</strong> Koutecky–Levich plot module.</figcaption>
+  <figcaption><strong>Figure 2.</strong> Koutecky–Levich plot module (logarithmic scale on the y-axis).</figcaption>
 </figure>
 
 ## Cyclic Voltammetry (CV) Module
