### 2025-08-13 Woffee/Envismetrics#10 @Grant-Xue
@MoonshinetheP

We thank the reviewer for the detailed and constructive feedback. We have carefully addressed each point as follows:

- **Potentiostat / Software Support** – All relevant sections (including Line 83) have been revised to explicitly state that the current release of *Envismetrics* is optimized for Autolab data.

- **CV Module RS Support** – We have clarified in both the manuscript and the user documentation how the CV module supports both reversible and irreversible Randles–Ševčík analysis, and where these functions are available in the web app and source code.

- **Function 4 (Nicholson-based \(k_0\))** – We agree that this method is non-standard for novice users and is sensitive to input parameters. Our goal is to continue exploring practical methods for obtaining kinetic parameters, and we therefore aim to introduce newly published approaches in future releases of the software. In the revised version, this function is now clearly marked as an “Experimental” feature in both the interface and documentation, with an explicit caution regarding its applicability and accuracy. ⚠️

- **Klinger–Kochi Method** – We have added a literature reference to support the description.

- **Levich & KL Analysis** – We acknowledge the reviewer’s point that the Levich equation applies strictly under steady-state, plateau conditions, and that KL at steady state yields the same transport-limited slope as Levich. In the revised version:
  - The \(D\) vs. \(E\) plot has been repositioned as an optional, diagnostic visualization and removed from the manuscript.
  - The manuscript and documentation now explicitly state that this plot should only be interpreted within the limiting-current plateau.
  - The main analysis workflow is plotting limiting current vs. rotation rate (Levich plot) as the primary method.
  - Figures 2a/2b have been replaced with corrected steady-state KL plots.

- **Peak Detection & Tafel Analysis** – We acknowledge that the description of peak detection was accidentally copied into the Tafel Analysis section. This has been corrected; no peak detection is performed in the Tafel Analysis module.
- We also updated the [Reproducing Manuscript Figures] at (https://github.com/Woffee/Envismetrics/blob/main/data/reproducibility_guide.md)
