### 2025-08-14 51a5d5b
Update paper.md
