### 2025-08-14 ff9fc38
Update paper.md
