+++ joss-paper/paper.md
@@ -184,23 +184,26 @@ In the implementation, setting \(\sqrt{n^{\prime} + \beta} = 1\) corresponds to 
 
 ### Function 4: Standard Rate Constant Calculation (Advanced)
 
-Function 4 implements an advanced, optional method for estimating the standard heterogeneous rate constant, $k_0$, using a dimensionless kinetic parameter $\Psi$ that relates $k_0$ to the system’s electrochemical and physical properties. This approach is based on the classical Nicholson model [@nicholson1965theory] and extended by Lavagnini *et al.* [@lavagnini2004extended] to cover a broader range of peak separations $\Delta E_p$, with additional support for the Klingler–Kochi formulation [@Klingler1981] in highly irreversible systems.
+Function 4 implements an advanced, optional method for estimating the standard heterogeneous rate constant, $$k_0$$, using a dimensionless kinetic parameter $$\Psi$$ that relates $$k_0$$ to the system’s electrochemical and physical properties. This approach is based on the classical Nicholson model [@nicholson1965theory] and extended by Lavagnini *et al.* [@lavagnini2004extended] to cover a broader range of peak separations ($$\Delta E_p$$), with additional support for the Klingler–Kochi formulation [@Klingler1981] in highly irreversible systems.
 
-$\Psi$ is estimated from the peak-to-peak separation using the empirical Lavagnini relationship:
-\[
+$$
 \Psi = \frac{0.6288 + 0.0021 \cdot X}{1 - 0.017 \cdot X}, \quad X = \Delta E_p \cdot n \ \ (\text{in mV})
-\]
-where $X$ is normalized to millivolts to maintain the dimensionless nature of \(\Psi\).
+$$
+where $$X$$ is normalized to millivolts to maintain the dimensionless nature of $$\Psi$$.
 
-For systems with large $\Delta E_p$ or highly irreversible behavior, the Klingler–Kochi expression is applied:
-\[
+For systems with large $$\Delta E_p$$ or highly irreversible behavior, the Klingler–Kochi expression is applied:
+$$
 \Psi = \frac{2.18}{\alpha \pi} \exp\left(-\frac{\alpha n \Delta E_p F}{2RT}\right)
-\]
+$$
 
 The standard rate constant is then obtained from:
-\[
+$$
 k^0 = \Psi \cdot \left(\frac{D \cdot n \cdot F}{R \cdot T}\right)^{1/2}
-\]
+$$
+
+**Assumptions and Scope**  
+This method assumes diffusion-controlled electron transfer without coupled chemical reactions or adsorption phenomena, and is most applicable to well-defined, peak-shaped CVs under quasi-reversible or irreversible conditions. The diffusion coefficient $$D$$ must be known or reliably estimated. Since the Lavagnini approach is empirical, optimal performance is expected when $$k_0$$ lies within an intermediate kinetic range. Users are advised to interpret $$k_0$$ results in accordance with electrochemical theory, and to restrict use of this feature to exploratory or comparative analysis rather than routine novice workflows.
+
 
 **Assumptions and Scope**  
 This method assumes diffusion-controlled electron transfer without coupled chemical reactions or adsorption phenomena, and is most applicable to well-defined, peak-shaped CVs under quasi-reversible or irreversible conditions. The diffusion coefficient $D$ must be known or reliably estimated. Since the Lavagnini approach is empirical, optimal performance is expected when $k_0$ lies within an intermediate kinetic range. Users are advised to interpret $k_0$ results in accordance with electrochemical theory, and to restrict use of this feature to exploratory or comparative analysis rather than routine novice workflows.
