### 2025-08-14 72a82cd
Update paper.md
