### 2025-08-14 592066f
Update paper.md
