+++ joss-paper/bibliography.bib
@@ -25,7 +25,19 @@ eprint = {
 
 }
 
-
+@article{WILOCH2024144089,
+title = {Redox behaviour of Cu-Aβ(4-16) complexes related to Alzheimer's Disease},
+journal = {Electrochimica Acta},
+volume = {485},
+pages = {144089},
+year = {2024},
+issn = {0013-4686},
+doi = {https://doi.org/10.1016/j.electacta.2024.144089},
+url = {https://www.sciencedirect.com/science/article/pii/S0013468624003311},
+author = {Magdalena Z. Wiloch and Steven Linfield and Natalia Baran and Wojciech Nogala and Martin Jönsson-Niedziółka},
+keywords = {Voltammetry, β-amyloid, Tyrosine, Copper complex, Alzheimer's Disease},
+abstract = {The development of Alzheimer's Disease (AD) has been linked to abnormal quantities of β-amyloid peptides in the brain. The majority of studies have focussed on Aβ(1-40/42) amyloids and their Cu(II)-Aβ(1-40/42) complexes which are responsible for production of reactive oxygen and nitrogen species (ROS and RNS), which are highly toxic to neurons. According to recent studies on amyloid plaques, Aβ(4–42), which is an N-truncated version of Aβ(1–42), is as prevalent as Aβ(1–42) in the brain. Although Cu(II) ions, bounded by Aβ(4–42), can be oxidized to highly reactive Cu(III) ions, its Cu(II) complexes do not appear to contribute to ROS/RNS formation. In this paper, the pH-dependent voltammetric response of Aβ(4–16)-Cu(II) complexes is investigated to understand the influence of the deprotonation of tyrosine within the complex towards the Cu(II)/Cu(III) reaction. The results will help to better understand the scavenging role of tyrosine in quenching highly reactive Cu(III) ions not only in Aβ(4-x)-Cu(II) complexes but also provide clues to the reactive properties of other tyrosine-containing amyloid-metal complexes.}
+}
 
 
 @article{TORABFAM2025107950,
