### 2025-08-14 b07b5e5
Update paper.md
