### 2025-08-14 c2b9c9d
Update paper.md
