+++ joss-paper/paper.md
@@ -58,23 +58,25 @@ While these methods are widely accepted, manual analysis can be labor-intensive 
 
 ## Statement of Need
 
-Electrochemical researchers often rely on a mix of tools for data analysis and visualization. Manual spreadsheet workflows (e.g., Excel) and general-purpose plotting software (e.g., Origin, SigmaPlot) are flexible, but they require significant manual preprocessing, repeated formatting, and substantial domain expertise for kinetic modeling. Proprietary instrument software (e.g., NOVA for Autolab) is primarily intended for device control and data acquisition; while it offers basic plotting, it is instrument-specific and is rarely used for in-depth data analysis [@Garg2021]. Researchers with strong programming skills sometimes develop custom analysis scripts in MATLAB or Python, but this demands considerable coding effort, debugging, and domain-specific implementation, which can be a barrier for many users.
+Electrochemical researchers often rely on a patchwork of tools for data analysis and visualization, each with significant limitations. Manual spreadsheet workflows (e.g., Excel) and general-purpose plotting software (e.g., Origin, SigmaPlot) are flexible but require labor-intensive preprocessing, repeated formatting, and substantial domain expertise for kinetic modeling. Proprietary instrument software (e.g., EC-lab & NOVA) is primarily designed for device control and data acquisition; its built-in plotting is basic, vendor-specific, and rarely used for advanced kinetic analysis [@Garg2021].
 
-Envismetrics addresses this gap by providing a modular, web-based platform dedicated to automated analysis of electrochemical data, including cyclic voltammetry (CV), rotating disk electrode linear sweep voltammetry (LSV at RDE), and chronoamperometry (CA). It supports widely used plaintext formats such as `.xlsx`, `.csv`, and `.txt`, allowing researchers to export data from most proprietary workstation software and perform consistent, reproducible analysis without being tied to a specific vendor.  
+Envismetrics directly addresses these limitations by offering automated, reproducible workflows for key electrochemical analyses—including peak detection, Levich and Randles–Ševčík analysis, diffusion coefficient and rate constant estimation—without requiring coding. Researchers with raw data from cyclic voltammetry (CV), rotating disk electrode linear sweep voltammetry (LSV at RDE), or chronoamperometry (CA) can obtain visual plots and tabulated results in minutes, avoiding the repetitive manual steps typical of Excel/Origin pipelines or custom MATLAB/Python scripts.
 
-Its modular architecture supports both platform expansion (potential of compatibility with additional electrochemical workstations) and method expansion (new analytical techniques). For example, while Envismetrics currently supports CV, LSV at RDE, and CA, it can be extended to include electrochemical impedance spectroscopy (EIS) and other workflows with minimal changes to the core system.
+The platform currently supports cyclic voltammetry (CV) data exported in widely used plaintext formats (.xlsx, .csv, .txt) from NOVA and EC-Lab software—reflecting the developer’s available instrumentation. Its modular architecture, however, is designed for both platform expansion (adding compatibility with additional electrochemical workstations) and method expansion (future integration of techniques such as electrochemical impedance spectroscopy).
 
-Envismetrics prioritizes data processing, reproducibility, and accessibility. It offers automated peak detection, Levich and Randles–Ševčík analysis, rate constant estimation, and stepwise analysis modules. Its browser-based design allows it to run on Windows, macOS, and Linux without installation, making it suitable for both research and teaching environments.
+Envismetrics runs entirely in the browser—requiring no installation or updates—and works seamlessly on Windows, macOS, and Linux. Its guided, user-friendly interface makes it equally well-suited for research laboratories and instructional settings. A direct feature comparison with other commonly used tools is provided in Table \ref{table:1}.
 
 
[…]
