### 2025-08-14 22e0f28
Update paper.md
