+++ joss-paper/paper.md
@@ -164,29 +164,35 @@ In *Envismetrics*, users can select potential values to automatically generate t
 
 ### Function 1: Plotting and Gaussian Filtering
 
-This function plots cyclic voltammetry data sorted based on the rate constant value and allows users to apply a Gaussian filter for smoothing. Users can input the sigma value to adjust the degree of smoothing. Both the original figure and the smoothed data will be displayed, allowing users to compare the raw and processed results.
+Plots cyclic voltammetry data and optionally applies Gaussian smoothing. Users can specify the `sigma` value to control smoothing intensity — higher values reduce high-frequency noise but may also attenuate sharp features. Both the raw and smoothed curves are displayed for direct comparison.  
+*(Identical in functionality to the HDV Gaussian filtering feature.)*
 
 ### Function 2: Peak Searching
 
-Peak searching is essential for calculating formal potential, peak separation, and performing Randles-Ševčík analysis. The software provides multiple searching methods, such as max/min and knee/elbow detection within specific ranges, allowing the analysis of multiple peaks and complex reactions. The software will record all the peak points for use in future analyses, and the results will be displayed in a plot.
+Peak identification is essential for calculating formal potentials, peak separations, and performing Randles–Ševčík analysis. *Envismetrics* primarily uses **Max/Min Search** to detect absolute peak positions within user-defined potential ranges. An additional **inflection-point detection** option is available for more complex waveforms. Users can define custom search ranges to target specific redox processes. Multiple peaks can be resolved in multi-step or complex reactions. All detected peak coordinates are stored for downstream analysis and are overlaid on the CV plot for visual verification.
+
 
 ### Function 3: Randles–Ševčík Analysis
 
-The Randles–Ševčík analysis utilizes equations that incorporate the transfer coefficient and calculate the diffusion coefficient from the peak current and scan rate. This function supports both reversible and irreversible versions of the Randles–Ševčík equation [@zanello2019inorganic]. The peak information data used in this analysis is obtained from Function 2 (Peak Searching):
+The Randles–Ševčík analysis estimates the diffusion coefficient $D$ from the relationship between peak current and scan rate, and can be applied to both reversible and irreversible electrochemical systems [@zanello2019inorganic]. Peak parameters for this calculation are obtained from **Function 2 (Peak Searching)**.
+
+For a **reversible** redox process, the Randles–Ševčík equation is:  
 
 $$
 I_{\text{peak}} = 0.4463 \ n \ F \ C \ A \sqrt{\frac{n F \nu D}{R T}}
 $$
 
+For an **irreversible** process, the equation becomes:  
+
 $$
 I_{\text{peak}} = 0.4463 \sqrt{n^{\prime} + \beta} \ n \ F \ C \ A \sqrt{\frac{n F \nu D}{R T}}
 $$
 
-In the implementation, setting \(\sqrt{n^{\prime} + \beta} = 1\) corresponds to the reversible case. For irreversible cases, \(\sqrt{n^{\prime} + \beta}\) is computed from user-specified \(\alpha\) and \(n'\) values. This design allows the same computational pipeline to handle both cases while preserving explicit parameter control for advanced users.
+In *Envismetrics*, setting $\sqrt{n' + \beta} = 1$ yields the reversible case. For irreversible systems, $\sqrt{n' + \beta}$ is calculated from the user-specified $\alpha$ and $n'$, allowing both scenarios to be handled within a single computational workflow while keeping full control over input parameters.
 
-### Function 4: Standard Rate Constant Calculation (Advanced)
[…]
