### 2025-08-14 82f82d2
Update paper.md
