### 2025-08-14 2306da3
Update paper.md
