+++ readme.md
@@ -203,44 +203,68 @@ By following these simple steps, you can efficiently utilize Envismetrics for yo
 <!-- ROADMAP -->
 ## Roadmap
 
-
 ### Hydrodynamic Voltammetry (HDV) Module
+1. **Data Import, Plotting, and Gaussian Filtering**
+   - Import data from supported potentiostats and file formats.
+   - Visualize data sorted by RPM (rotations per minute).
+   - Apply Gaussian filtering with user-defined sigma for noise reduction.
+2. **Levich and Koutecký–Levich Analysis**
+   - **Levich Analysis**: Generate plots and calculate diffusion coefficients from slope.
+   - **Koutecký–Levich Analysis**: Produce plots, perform linear regression, and analyze diffusion coefficients at selected potentials.
 
-1. Data Import, Plotting, and Gaussian Filtering
-    - Data Import: Supports importing data from various potentiostats and file formats.
-    - Plotting: Visualizes the imported data sorted by RPM (rotations per minute).
-    - Gaussian Filtering: Applies a Gaussian filter to smooth the plotted data.
-2. Levich and Koutecky-Levich Analysis
-	- Levich Plot and Analysis:
-		- Generates Levich plots directly from the data.
-		- Calculates the diffusion coefficient from the slope of the plot.
-	- Koutecky-Levich Plot and Analysis:
-		- Produces Koutecky-Levich plots.
-		- Performs linear regression to analyze the diffusion coefficient at various potentials.
+**Planned Features**
+- Automatic detection of limiting current plateaus for improved Levich/KL regression accuracy.
+- Enhanced fitting diagnostics with linearity feedback and outlier detection.
+- Sliding-range selection for potential window fitting.
+
+---
 
 ### Cyclic Voltammetry (CV) Module
+1. **Plotting and Gaussian Filtering**
+   - Plot CV data sorted by rate constant value.
+   - Apply Gaussian filtering with adjustable sigma values.
+2. **Peak Searching**
+   - Identify peaks within defined potential ranges (max/min detection).
+   - Record peak coordinates for downstream analysis.
+3. **Randles–Ševčík Analysis**
+   - Calculate diffusion coefficients from peak current vs. scan rate.
+4. **Standard Rate Constant Calculation**
+   - Estimate $k^0$ using peak separation and Nicholson/Lavagnini models.
+5. **Tafel Analysis**
+   - Determine anodic and cathodic transfer coefficients.
+   - Implement mass-transport correction for improved accuracy.
 
-1. Plotting and Gaussian Filtering
-	- Plotting: Visualizes cyclic voltammetry data sorted by rate constant value.
-	- Gaussian Filtering: Applies a Gaussian filter to smooth the data, with user-defined sigma values.
-2. Peak Searching
-	- Identifies peak points within specific ranges using various methods (max/min, knee/elbow detection).
-	- Records peak information for use in subsequent analyses.
-3. Randles–Ševčík Analysis
-	- Calculates the diffusion coefficient from peak current and scan rate.
-4. Rate Constant Calculation
-	- Calculates the rate constant using peak separation
-5. Tafel Analysis Module
-    - Determines anodic and cathodic transfer coefficients.
-    - Implements mass-transport corrected methods for enhanced accuracy.
+**Planned Features**
+- Fully automated peak detection across voltammetric cycles.
+- Regime diagnostics based on scan-rate normalization (e.g., $i_p$ vs. $\sqrt{v}$) to identify non-planar diffusion behavior.
+- Automatic exclusion of non-conforming voltammograms.
+- Real-time validation of input parameters (e.g., electrode radius, $D$, $C_0$) with warnings for unphysical values.
 
-### Step Techniques Module
+---
 
[…]
