### 2025-08-14 f22c0ca
Update paper.md
