### 2025-08-14 7049916
Update bibliography.bib
