+++ joss-paper/bibliography.bib
@@ -266,18 +266,15 @@ keywords = {Electrochemical techniques, Kinetic parameters, Exchange current, Di
 abstract = {The derivation and proposal of major electrochemical techniques used to determine and calculate the electrochemical kinetic parameters is basically based on the electrochemical reaction taking place at liquid/solid or liquid/liquid interface in which all the reactants and products are soluble in liquid aqueous solution or liquid mercury electrode, or are volatile gas. Such electrochemical reaction system is classical and traditional (ERS1). Recently, the electrochemical behavior of some materials used as the active electrode materials in chemical power sources has attracted much attention. In chemical power source systems, either reactant or product, or both are insoluble. This kind of electrochemical reaction system (ERS2) is slightly different from ERS1. The application of these electrochemical techniques/equations to chemical power sources’ system requires carefulness. The misuse of these electrochemical techniques can be easily found in the literatures and some of them even lead to a wrong conclusion. In this review, almost all the electrochemical techniques to measure the exchange current and diffusion coefficient were compiled for reference to the readers, including pulse step, electrochemical impedance, alternating cyclic voltammetry, etc. The necessary requirements/conditions to apply these techniques have been briefly discussed and some simple examples were also discussed for a better understanding.}
 }
 
-@article{SANECKI2003109,
-title = {The problem of the accuracy of electrochemical kinetic parameter determination for the ECE reaction mechanism},
-journal = {Journal of Electroanalytical Chemistry},
-volume = {546},
-pages = {109-121},
-year = {2003},
-issn = {1572-6657},
-doi = {10.1016/S0022-0728(03)00138-4},
-url = {https://www.sciencedirect.com/science/article/pii/S0022072803001384},
-author = {Przemysław Sanecki and Christian Amatore and Piotr Skitał},
-keywords = {ECE mechanism, Cyclic voltammetry, Digital simulation},
-abstract = {The theoretical and practical possibilities of the determination of elementary kinetic parameters k1, α1, k2, α2 for sequence A→k1,α1B→kfC→k2,α2D were examined for various k2/k1 ratios on the basis of electrochemical simulations. The errors of the kinetic parameters evaluated from theoretical CV curves or from those with artificial noise added were calculated and discussed. The subject is important for the modelling of the ECE mechanism and other consecutive reaction steps. The work demonstrates quantitatively the relations that are qualitatively familiar to chemists: a precise determination of second stage kinetic parameters in a consecutive process is impossible if their magnitude is much higher than that in the first stage.}
+@article{SANECKI2003,
+  title   = {The problem of the accuracy of electrochemical kinetic parameter determination for the {ECE} reaction mechanism},
+  author  = {Sanecki, Przemys{\l}aw and Amatore, Christian and Skita{\l}, Piotr},
+  journal = {Journal of Electroanalytical Chemistry},
+  year    = {2003},
+  volume  = {546},
+  pages   = {109--121},
+  doi     = {10.1016/S0022-0728(03)00138-4},
+  issn    = {1572-6657}
 }
 
 @article{C9CP05527D,
