+++ joss-paper/bibliography.bib
@@ -268,7 +268,7 @@ abstract = {The derivation and proposal of major electrochemical techniques used
 
 @article{SANECKI2003,
   title   = {The problem of the accuracy of electrochemical kinetic parameter determination for the {ECE} reaction mechanism},
-  author  = {Sanecki, Przemys{\l}aw and Amatore, Christian and Skita{\l}, Piotr},
+  author  = {Sanecki, Przemyslaw and Amatore, Christian and Skital, Piotr},
   journal = {Journal of Electroanalytical Chemistry},
   year    = {2003},
   volume  = {546},
