### 2025-08-14 3f3252e
Update bibliography_2.bib
