+++ joss-paper/bibliography_2.bib
@@ -1,1 +1,308 @@
-1
+@article{SANECKI2003,
+  title   = {The problem of the accuracy of electrochemical kinetic parameter determination for the {ECE} reaction mechanism},
+  author  = {Sanecki, Przemyslaw and Amatore, Christian and Skital, Piotr},
+  journal = {Journal of Electroanalytical Chemistry},
+  year    = {2003},
+  volume  = {546},
+  pages   = {109--121},
+  doi     = {10.1016/S0022-0728(03)00138-4},
+  issn    = {1572-6657}
+}
+
+@article{wang2020,
+  author  = {Wang, Hao and Sayed, Sayed Youssef and Luber, Erik J. and Olsen, Brian C. and Shirurkar, Shubham M. and Venkatakrishnan, Sankaranarayanan and Tefashe, Ushula M. and Farquhar, Anna K. and Smotkin, Eugene S. and McCreery, Richard L. and Buriak, Jillian M.},
+  title   = {Redox Flow Batteries: How to Determine Electrochemical Kinetic Parameters},
+  journal = {ACS Nano},
+  year    = {2020},
+  volume  = {14},
+  number  = {3},
+  pages   = {2575--2584},
+  doi     = {10.1021/acsnano.0c01281}
+}
+
+@article{WILOCH2024,
+  author  = {Wiloch, Magdalena Z. and Linfield, Steven and Baran, Natalia and Nogala, Wojciech and Jönsson-Niedziółka, Martin},
+  title   = {Redox Behaviour of Cu-Aβ(4-16) Complexes Related to Alzheimer's Disease},
+  journal = {Electrochimica Acta},
+  year    = {2024},
+  volume  = {485},
+  pages   = {144089},
+  doi     = {10.1016/j.electacta.2024.144089},
+  issn    = {0013-4686}
+}
+
+@article{XU2010,
+  author  = {Xu, Yanhui and Chen, Ying and Wu, Jun and Li, Decheng and Ju, Hua and Zheng, Junwei},
+  title   = {The Determination of the Kinetic Parameters of Electrochemical Reaction in Chemical Power Sources: A Critical Review},
+  journal = {International Journal of Hydrogen Energy},
+  year    = {2010},
+  volume  = {35},
+  number  = {12},
+  pages   = {6366--6380},
+  doi     = {10.1016/j.ijhydene.2010.03.115},
+  issn    = {0360-3199}
+}
+
+@article{Bacil2020,
+  author  = {Bacil, Raphael P. and Chen, Lifu and Serrano, Silvia H. P. and Compton, Richard G.},
+  title   = {Dopamine Oxidation at Gold Electrodes: Mechanism and Kinetics Near Neutral pH},
+  journal = {Physical Chemistry Chemical Physics},
+  year    = {2020},
+  volume  = {22},
+  number  = {2},
+  pages   = {607--614},
+  doi     = {10.1039/C9CP05527D}
+}
+
+@book{Bard2022,
+  title     = {Electrochemical Methods: Fundamentals and Applications},
+  author    = {Bard, A. J. and Faulkner, L. R. and White, H. S.},
+  isbn      = {9781119334064},
+  lccn      = {2022017533},
+  url       = {https://books.google.com/books?id=Sct6EAAAQBAJ},
+  year      = {2022},
+  publisher = {Wiley}
+}
+
+@article{Bruckenstein1977,
+  author  = {Bruckenstein, Stanley and Miller, Barry},
+  title   = {Unraveling reactions with rotating electrodes},
+  journal = {Accounts of Chemical Research},
+  volume  = {10},
+  number  = {2},
+  pages   = {54--61},
+  year    = {1977},
+  doi     = {10.1021/ar50110a004},
[…]
