### 2025-08-14 c3a8ee0
Update paper.md
