### 2025-08-14 Woffee/Envismetrics#11 @Grant-Xue
@MoonshinetheP,

I really want to thank you for your patience and detailed explanation. I finally understand the difference between the Markdown and PDF versions—previously, I didn’t know how to check the JOSS PDF version, but now I fully get it.

I have completely revised the `bibliography.bib` file so that everything follows a universal and consistent format. The citation keys have been updated to use the surname of the first author followed by the publication year, which I agree is a very good practice.

I have already checked the generated PDF, and I believe there will be no issues now. I also removed the conference abstract entry.

For the newly added or removed references:
- Some were added because another reviewer specifically requested additional references in certain sections.
- Some were removed because they lacked a DOI or ISSN, or were not peer-reviewed publications.

Thanks again for your guidance!
