### 2025-08-14 87514a1
Update paper.md
