+++ src/CA_commented.py
@@ -1,423 +1,0 @@
-"""
-CA.py - Chronoamperometry (CA) Analysis Module
-----------------------------------------------
-
-This module is part of the Envismetrics software suite and provides both visualization
-and diffusion coefficient analysis for Chronoamperometry (CA) data. It reads time-dependent
-potential and current measurements from electrochemical experiments and offers a standardized
-approach to visualize raw signals and compute key kinetic parameters using the Cottrell equation.
-
-Core Functions:
----------------
-1. `step1()`: Raw Data Visualization
-   - Generates two plots per dataset:
-     - A: Applied Potential vs Time
-     - B: Current vs Time
-   - Useful for preliminary data inspection and quality control.
-
-2. `step2(inter, n, a, c, x_range)`: Diffusion Coefficient Calculation
-   - Applies the Cottrell equation to estimate diffusion coefficients (D) from current-time curves.
-   - Performs linear regression on transformed data (Bt vs I), where Bt ~ t^(-1/2).
-   - Produces:
-     - Plot of raw current vs time.
-     - Plot of regression used to extract D.
-     - Summary table (CSV) with slope, D, and R² for each dataset.
-
-Features:
----------
-- Supports batch processing of multiple CA files.
-- Configurable analysis range (`x_range`) for linear regression.
-- Automatically handles file saving, figure generation, and output management.
-- Saves intermediate results and metadata to `data.json` and CSV for reproducibility.
-
-Dependencies:
--------------
-- numpy, pandas, matplotlib, scipy, sklearn
-- config.py
-- BaseModule.py
-
-Date: 2025  
-"""
-
-
-import os
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.ndimage import gaussian_filter
-from sklearn.linear_model import LinearRegression
-import math
-import re
-import json
-from config import *
-from BaseModule import BaseModule
-
-def get_num(filename):
-    """
-    Extract the first integer number found in a filename using regular expressions.
-
-    Parameters:
-        filename (str): Input filename string (e.g., '3PFOA400ppm_75075_CA.xlsx').
-
-    Returns:
-        int or None: The first numeric match converted to int, or None if no number is found.
-
-    Notes:
-        This is commonly used to extract concentration, ID, or numeric prefixes from filenames.
-    """
-    match = re.search(r'(\d+)', filename)
-    return int(match.group(1)) if match else None
-
-class CA(BaseModule):
-    def __init__(self, version):
-        """
-        Initialize the Chronoamperometry (CA) module.
-
-        This class inherits from BaseModule and sets up the environment 
-        for processing CA data associated with a given analysis version.
-
-        Args:
-            version (str): Unique identifier for this session/run. 
-                           Used to organize output directories and results.
-        """
-        super().__init__(version)
-
-    def check_columns(self, data):
-        """
-        Validate that all required columns are present in the input DataFrames.
-
-        Required columns:
-            - 'Time (s)': Time axis in seconds.
-            - 'WE(1).Current (A)': Working electrode current in amperes.
-            - 'WE(1).Potential (V)': Working electrode potential in volts.
-
-        Args:
-            data (list): List of dictionaries, each containing a 'df' key with a pandas DataFrame.
-
-        Returns:
-            str: Empty string if all columns are present; otherwise, returns an error message
-                 listing all missing columns.
-        """
-        cols = ['Time (s)', 'WE(1).Current (A)', 'WE(1).Potential (V)']
-        missing_cols = []
-        for d in data:
-            df = d['df']
-            for col in cols:
-                if col not in df.columns:
-                    missing_cols.append(col)
-        if len(missing_cols) > 0:
-            return "error: Missing columns: " + ", ".join(missing_cols)
-        return ''
-
-    def step1(self):
-        """
[…]
