### 2025-08-15 cf31d27
Delete src/CV_commented.py
