### 2025-08-15 31d3a24
Delete src/HDV_commented.py
