### 2025-08-18 37c1dc0
remove unused variables, functions
