+++ src/CV.py
@@ -83,136 +83,6 @@ colors = [
     '#17becf'   # tab:cyan
 ]
 
-def find_max(x, y, start, end):
-    """
-    Find the maximum y-value and corresponding x in a given x-range.
-
-    Parameters:
-        x (list or np.array): x-values
-        y (list or np.array): y-values
-        start (float): lower x-bound
-        end (float): upper x-bound
-
-    Returns:
-        tuple: (x, y) position of maximum y-value within range
-    """
-    ma = -1       # Initialize max y with a sentinel; assumes y-values are non-negative
-    xx = -1       # Placeholder for x corresponding to max y
-    yy = -1       # Placeholder for max y itself
-
-    # Iterate through all x-values to identify those within the specified range
-    for i in range(len(x)):
-        if start <= x[i] <= end:
-            # Update max only if current y-value exceeds previous max
-            if y[i] > ma:
-                ma = y[i]
-                xx = x[i]
-                yy = y[i]
-
-    # Return the (x, y) coordinates of the max y-value found in the given range
-    return xx, yy
-
-
-def find_min(x, y, start, end):
-    """
-    Find the minimum y-value and corresponding x in a given x-range.
-
-    Parameters:
-        x (list or np.array): x-values
-        y (list or np.array): y-values
-        start (float): lower x-bound
-        end (float): upper x-bound
-
-    Returns:
-        tuple: (x, y) position of minimum y-value within range
-    """
-    mi = 10000    # Initialize min y with a large number as sentinel
-                  # Assumes y-values are significantly smaller than 10^4
-    xx = -1       # Placeholder for x corresponding to min y
-    yy = -1       # Placeholder for min y itself
-
-    # Iterate through all x-values to check which ones fall in the target range
-    for i in range(len(x)):
-        if start <= x[i] <= end:
-            # Update min only if current y-value is smaller
-            if y[i] < mi:
-                mi = y[i]
-                xx = x[i]
-                yy = y[i]
-
-    # Return the (x, y) coordinates of the min y-value found in the specified range
-    return xx, yy
-
-
-def find_y(x, y, xi):
-    """
-    Find the y-value corresponding to a given x-value (xi).
-
-    Parameters:
-        x (list): list of x-values
-        y (list): list of y-values
-        xi (float): x target
-
-    Returns:
-        float: y-value at xi or -1 if not found
-    """
-    # Linear scan to find the x that exactly matches xi
-    for i in range(len(x)):
-        if x[i] == xi:
-            return y[i]
-
-    # If xi not found in x (e.g., due to floating-point mismatch), return fallback
-    return -1
-
-
-def separater(x, y, left, right):
-    """
-    Separate a cyclic voltammogram into upper and lower sweep segments.
-
-    Parameters:
-        x (pd.Series): potential values
-        y (pd.Series): current values
-        left (float): left potential bound
-        right (float): right potential bound
-
-    Returns:
-        tuple: (upperx, lowerx, uppery, lowery)
-    """
-    upperx = []  # Stores x-values (potential) for forward (upper) scan
-    lowerx = []  # Stores x-values for reverse (lower) scan
-    uppery = []  # Stores corresponding y-values (current) for upper scan
-    lowery = []  # Stores y-values for lower scan
-
-    x = x.tolist()  # Convert input from pd.Series to native Python list
-    y = y.tolist()
-
-    # Identify the index of key potential boundaries
-    boundary_l = x.index(left)
[…]
