### 2025-09-16 8b52790
Update LICENSE.txt
