### 2025-09-25 #8436 @Kevin-Mattheus-Moerman
@MoonshinetheP @oliverrdz @rena-96 thanks for your review efforts here! I have gone over the issues/concerns raised. Given the general concerns on software/paper quality, and the concerns relating to the use of LLMs/AI, I will now proceed to reject this submission.

@Woffee given the concerns shared by the reviewers, and in line with [our journal's policies on generative AI](https://joss.readthedocs.io/en/latest/submitting.html#interim-position-on-ai-enabled-software) I have decided to reject this work as unsuitable for publication in JOSS.
